Auto-Correlation functions

GaPSE.ξ_Doppler — Function

 ξ_Doppler(P1::Point, P2::Point, y, cosmo::Cosmology) :: Float64

Return the Doppler auto-correlation function, defined as follows:

\[\xi^{v_{\parallel}v_{\parallel}} (s_1, s_2, \cos{\theta}) = D_1 D_2 f_1 f_2 \mathcal{H}_1 \mathcal{H}_2 \mathcal{R}_1 \mathcal{R}_2 (J_{00}\, I^0_0(s) + J_{02}\,I^0_2(s) + J_{04}\,I^0_4(s) + J_{20}\,I^2_0(s))\]

where $D_1 = D(s_1)$, $D_2 = D(s_2)$ and so on, $\mathcal{H} = a H$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and the J coefficients are given by:

\[\begin{align*} J_{00} (s_1, s_2, y) & = \frac{1}{45} (y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2) \\ J_{02} (s_1, s_2, y) & = \frac{2}{63} (y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2) \\ J_{04} (s_1, s_2, y) & = \frac{1}{105} (y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2) \\ J_{20} (s_1, s_2, y) & = \frac{1}{3} y s^2 \end{align*}\]

Inputs

P1::Point and P2::Point: Point where the CF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on)

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

See also: Point, Cosmology

GaPSE.ξ_Lensing — Function

 ξ_Lensing(s1, s2, y, cosmo::Cosmology;
      en::Float64 = 1e6,
      Δχ_min::Float64 = 1e-3,
      N_χs::Integer = 100) :: Float64

Return the Lensing auto-correlation function $\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta}) = \int_0^{s_1} \mathrm{d} \chi_1 \int_0^{s_2} \mathrm{d} \chi_2 \frac{1}{2} \frac{ \mathcal{H}_0^4 \Omega_{ \mathrm{M0}}^2 D_1 D_2 (\chi_1 - s_1)(\chi_2 - s_2) }{ s_1 s_2 a(\chi_1) a(\chi_2) } (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi))\]

where $D_1 = D(\chi_1)$, $D_2 = D(\chi_2)$ and so on, $\mathcal{H} = a H$, $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2\chi_1\chi_2\cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$) and the $J$ coefficients are given by

\[\begin{align*} J_{00} & = - \frac{3 \chi_1^2 \chi_2^2}{4 \chi^4} (y^2 - 1) (8 y (\chi_1^2 + \chi_2^2) - 9 \chi_1 \chi_2 y^2 - 7 \chi_1 \chi_2) \\ J_{02} & = - \frac{3 \chi_1^2 \chi_2^2}{2 \chi^4} (y^2 - 1) (4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2) \\ J_{31} & = 9 y \chi^2 \\ J_{22} & = \frac{9 \chi_1 \chi_2}{4 \chi^4} [ 2 (\chi_1^4 + \chi_2^4) (7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2) (y^2+1) + \chi_1^2 \chi_2^2 (11 y^4 + 14 y^2 + 23)] \end{align*}\]

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_Lensing.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

Δχ_min::Float64 = 1e-6 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\chi\to0^{+}} (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi)) = \frac{4}{15} \, (5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2)\]

So, when it happens that $\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_Lensing, integrand_on_mu_Lensing integral_on_mu, ξ_multipole

GaPSE.ξ_LocalGP — Function

 ξ_LocalGP(P1::Point, P2::Point, y, cosmo::Cosmology) :: Float64

Return the local gravitational potential auto-correlation function, defined as follows:

\[\xi^{\phi\phi} (s_1, s_2, \cos{\theta}) = \frac{9 \mathcal{H}_0^4 \Omega_{M0}^2 D(s_1) D(s_2)s^4}{4 a(s_1) a(s_2)} (1 + \mathcal{R}_1 + \mathcal{R}_2 + \mathcal{R}_1\mathcal{R}_2) \tilde{I}^4_0(s)\]

where $D_1 = D(s_1)$, $D_2 = D(s_2)$ and so on, $\mathcal{H} = a H$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and:

\[\tilde{I}^4_0 (s) &= \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_0(q s) - 1}{(q s)^4}\]

Inputs

P1::Point and P2::Point: Point where the CF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on)

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

See also: Point, Cosmology

GaPSE.ξ_IntegratedGP — Function

 ξ_IntegratedGP(s1, s2, y, cosmo::Cosmology; 
      en::Float64 = 1e10,
      N_χs::Integer = 100) :: Float64

Return the integrated gravitational potential auto-correlation function $\xi^{\int\phi\int\phi}(s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{\int\phi\int\phi} (s_1, s_2, \cos{\theta}) = \int_0^{s_1} \mathrm{d} \chi_1 \int_0^{s_2}\mathrm{d} \chi_2 \; J_{40}(s_1, s_2, y, \chi_1, \chi_2) \, \tilde{I}^4_0(\chi)\]

where $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \, \chi_2 \, y}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and:

\[\begin{align*} J_{40}(s_1, s_2, y, \chi_1, \chi_2) &= \frac{ 9 \mathcal{H}_0^4 \Omega_{M0}^2 D(\chi_1) D(\chi_2) \chi^4 }{ a(\chi_1) a(\chi_2) s_1 s_2} (s_2 \mathcal{H}(\chi_2) \mathcal{R}(s_2) (f(\chi_2)-1) - 1) (s_1 \mathcal{H}(\chi_1) \mathcal{R}(s_1) (f(\chi_1)-1) - 1) \\[5pt] \tilde{I}^4_0 (s) &= \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_0(q s) - 1}{(q s)^4} \end{aling*}\]

and $P(q)$ is the input power spectrum.

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_Lensing.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e10: just a float number used in order to deal better with small numbers.

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_IntegratedGP, integrand_on_mu_IntegratedGP integral_on_mu, ξ_multipole

GaPSE.integrand_ξ_Lensing — Function

 integrand_ξ_Lensing(
      IP1::Point, IP2::Point,
      P1::Point, P2::Point,
      y, cosmo::Cosmology;
      Δχ_min::Float64 = 1e-4) :: Float64

Return the integrand of the Lensing auto-correlation function $\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta})$, i.e. the function $f(s_1, s_2, y, \chi_1, \chi_2)$ defined as follows:

\[f(s_1, s_2, y, \chi_1, \chi_2) = \frac{1}{2} \frac{ \mathcal{H}_0^4 \Omega_{ \mathrm{M0}}^2 D_1 D_2 (\chi_1 - s_1)(\chi_2 - s_2) }{ s_1 s_2 a(\chi_1) a(\chi_2) } (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi))\]

where $D_1 = D(\chi_1)$, $D_2 = D(\chi_2)$ and so on, $\mathcal{H} = a H$, $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2\chi_1\chi_2\cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$) and the $J$ coefficients are given by

\[\begin{align*} J_{00} & = - \frac{3 \chi_1^2 \chi_2^2}{4 \chi^4} (y^2 - 1) (8 y (\chi_1^2 + \chi_2^2) - 9 \chi_1 \chi_2 y^2 - 7 \chi_1 \chi_2) \\ J_{02} & = - \frac{3 \chi_1^2 \chi_2^2}{2 \chi^4} (y^2 - 1) (4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2) \\ J_{31} & = 9 y \chi^2 \\ J_{22} & = \frac{9 \chi_1 \chi_2}{4 \chi^4} [ 2 (\chi_1^4 + \chi_2^4) (7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2) (y^2+1) + \chi_1^2 \chi_2^2 (11 y^4 + 14 y^2 + 23)] \end{align*}\]

Inputs

IP1::Point and IP2::Point: Point inside the integration limits, placed at comoving distance χ1 and χ2 respectively.

P1::Point and P2::Point: extreme Point of the integration, placed at comoving distance s1 and s2 respectively.

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

Δχ_min::Float64 = 1e-6 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\chi\to0^{+}} (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi)) = \frac{4}{15} \, (5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2)\]

So, when it happens that $\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

See also: ξ_Lensing, integrand_on_mu_Lensing integral_on_mu, ξ_multipole

GaPSE.integrand_ξ_IntegratedGP — Function

 integrand_ξ_IntegratedGP(IP1::Point, IP2::Point,
      P1::Point, P2::Point,
      y, cosmo::Cosmology) :: Float64

Return the integrand of the integrated gravitational potential auto-correlation function $\xi^{\int\phi\int\phi} (s_1, s_2, \cos{\theta})$, i.e. the function $f(s_1, s_2, y, \chi_1, \chi_2)$ defined as follows:

\[f(s_1, s_2, y, \chi_1, \chi_2) = J_{40}(s_1, s_2, y, \chi_1, \chi_2) \tilde{I}^4_0(\chi)\]

where $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \, \chi_2 \, y}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and:

\[\begin{split} &J_{40}(s_1, s_2, y, \chi_1, \chi_2) = \frac{ 9 \mathcal{H}_0^4 \Omega_{M0}^2 D(\chi_1) D(\chi_2) \chi^4 }{ a(\chi_1) a(\chi_2) s_1 s_2} (s_2 \mathcal{H}(\chi_2) \mathcal{R}(s_2) (f(\chi_2)-1) - 1) (s_1 \mathcal{H}(\chi_1) \mathcal{R}(s_1) (f(\chi_1)-1) - 1)\\[5pt] &\tilde{I}^4_0 (s) = \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_0(q s) - 1}{(q s)^4} \end{split}\]

Inputs

IP1::Point and IP2::Point: Point inside the integration limits, placed at comoving distance χ1 and χ2 respectively.

P1::Point and P2::Point: extreme Point of the integration, placed at comoving distance s1 and s2 respectively.

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

See also: ξ_IntegratedGP, integrand_on_mu_IntegratedGP integral_on_mu, ξ_multipole