Doppler Auto-Correlation functions

GaPSE.ξ_doppler — Function

 ξ_doppler(P1::Point, P2::Point, y, cosmo::Cosmology; enhancer = 1.0) :: Float64

Return the Doppler auto-correlation function, defined as follows:

\[\xi^{v_{\parallel}v_{\parallel}} (s_1, s_2, \cos{\theta}) = D_1 D_2 f_1 f_2 \mathcal{H}_1 \mathcal{H}_2 \mathcal{R}_1 \mathcal{R}_2 (J_{00}\, I^0_0(s) + J_{02}\,I^0_2(s) + J_{04}\,I^0_4(s) + J_{20}\,I^2_0(s))\]

where $D_1 = D(s_1)$, $D_2 = D(s_2)$ and so on, $\mathcal{H} = a H$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and the J coefficients are given by:

\[\begin{align*} J_{00} (s_1, s_2, y) & = \frac{1}{45} (y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2) \\ J_{02} (s_1, s_2, y) & = \frac{2}{63} (y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2) \\ J_{04} (s_1, s_2, y) & = \frac{1}{105} (y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2) \\ J_{20} (s_1, s_2, y) & = \frac{1}{3} y s^2 \end{align*}\]

Inputs

P1::Point and P2::Point: Point where the CF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on)

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

enhancer::Float64 = 1.0: just a float number used in order to deal better with small numbers; the returned value is actually enhancer * f, where f is the true value calculated as shown before.

See also: Point, Cosmology

Missing docstring.

Missing docstring for GaSPE.integrand_on_mu_doppler. Check Documenter's build log for details.