Integrated Gravitational Potential Auto-Correlation functions

GaPSE.integrand_ξ_integratedGP — Function

 integrand_ξ_integratedGP(IP1::Point, IP2::Point,
      P1::Point, P2::Point,
      y, cosmo::Cosmology) :: Float64

Return the integrand of the integrated gravitational potential auto-correlation function $\xi^{\int\phi\int\phi} (s_1, s_2, \cos{\theta})$, i.e. the function $f(s_1, s_2, y, \chi_1, \chi_2)$ defined as follows:

\[f(s_1, s_2, y, \chi_1, \chi_2) = J_{40}(s_1, s_2, y, \chi_1, \chi_2) \tilde{I}^4_0(\chi)\]

where $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \, \chi_2 \, y}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and:

\[\begin{split} &J_{40}(s_1, s_2, y, \chi_1, \chi_2) = \frac{ 9 \mathcal{H}_0^4 \Omega_{M0}^2 D(\chi_1) D(\chi_2) \chi^4 }{ a(\chi_1) a(\chi_2) s_1 s_2} (s_2 \mathcal{H}(\chi_2) \mathcal{R}(s_2) (f(\chi_2)-1) - 1) (s_1 \mathcal{H}(\chi_1) \mathcal{R}(s_1) (f(\chi_1)-1) - 1)\\[5pt] &\tilde{I}^4_0 (s) = \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_0(q s) - 1}{(q s)^4} \end{split}\]

Inputs

IP1::Point and IP2::Point: Point inside the integration limits, placed at comoving distance χ1 and χ2 respectively.

P1::Point and P2::Point: extreme Point of the integration, placed at comoving distance s1 and s2 respectively.

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

See also: ξ_integratedGP, integrand_on_mu_integratedGP integral_on_mu, ξ_multipole

GaPSE.ξ_integratedGP — Function

 ξ_integratedGP(s1, s2, y, cosmo::Cosmology; 
      en::Float64 = 1e10,
      N_χs::Integer = 100) :: Float64

Return the integrated gravitational potential auto-correlation function $\xi^{\int\phi\int\phi}(s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{\int\phi\int\phi} (s_1, s_2, \cos{\theta}) = \int_0^{s_1} \mathrm{d} \chi_1 \int_0^{s_2}\mathrm{d} \chi_2 \; J_{40}(s_1, s_2, y, \chi_1, \chi_2) \, \tilde{I}^4_0(\chi)\]

where $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \, \chi_2 \, y}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and:

\[\begin{align*} J_{40}(s_1, s_2, y, \chi_1, \chi_2) &= \frac{ 9 \mathcal{H}_0^4 \Omega_{M0}^2 D(\chi_1) D(\chi_2) \chi^4 }{ a(\chi_1) a(\chi_2) s_1 s_2} (s_2 \mathcal{H}(\chi_2) \mathcal{R}(s_2) (f(\chi_2)-1) - 1) (s_1 \mathcal{H}(\chi_1) \mathcal{R}(s_1) (f(\chi_1)-1) - 1) \\[5pt] \tilde{I}^4_0 (s) &= \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_0(q s) - 1}{(q s)^4} \end{aling*}\]

and $P(q)$ is the input power spectrum.

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_lensing.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e10: just a float number used in order to deal better with small numbers.

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_integratedGP, integrand_on_mu_integratedGP integral_on_mu, ξ_multipole

Missing docstring.

Missing docstring for GaSPE.integrand_on_mu_integratedGP. Check Documenter's build log for details.