Lensing Auto-Correlation functions

GaPSE.integrand_ξ_lensing — Function

 integrand_ξ_lensing(
      IP1::Point, IP2::Point,
      P1::Point, P2::Point,
      y, cosmo::Cosmology;
      Δχ_min::Float64 = 1e-4) :: Float64

Return the integrand of the lensing auto-correlation function $\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta})$, i.e. the function $f(s_1, s_2, y, \chi_1, \chi_2)$ defined as follows:

\[f(s_1, s_2, y, \chi_1, \chi_2) = \frac{1}{2} \frac{ \mathcal{H}_0^4 \Omega_{ \mathrm{M0}}^2 D_1 D_2 (\chi_1 - s_1)(\chi_2 - s_2) }{ s_1 s_2 a(\chi_1) a(\chi_2) } (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi))\]

where $D_1 = D(\chi_1)$, $D_2 = D(\chi_2)$ and so on, $\mathcal{H} = a H$, $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2\chi_1\chi_2\cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$) and the $J$ coefficients are given by

\[\begin{align*} J_{00} & = - \frac{3 \chi_1^2 \chi_2^2}{4 \chi^4} (y^2 - 1) (8 y (\chi_1^2 + \chi_2^2) - 9 \chi_1 \chi_2 y^2 - 7 \chi_1 \chi_2) \\ J_{02} & = - \frac{3 \chi_1^2 \chi_2^2}{2 \chi^4} (y^2 - 1) (4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2) \\ J_{31} & = 9 y \chi^2 \\ J_{22} & = \frac{9 \chi_1 \chi_2}{4 \chi^4} [ 2 (\chi_1^4 + \chi_2^4) (7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2) (y^2+1) + \chi_1^2 \chi_2^2 (11 y^4 + 14 y^2 + 23)] \end{align*}\]

Inputs

IP1::Point and IP2::Point: Point inside the integration limits, placed at comoving distance χ1 and χ2 respectively.

P1::Point and P2::Point: extreme Point of the integration, placed at comoving distance s1 and s2 respectively.

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

Δχ_min::Float64 = 1e-6 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\chi\to0^{+}} (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi)) = \frac{4}{15} \, (5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2)\]

So, when it happens that $\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

See also: ξ_lensing, integrand_on_mu_lensing integral_on_mu, ξ_multipole

GaPSE.ξ_lensing — Function

 ξ_lensing(s1, s2, y, cosmo::Cosmology;
      en::Float64 = 1e6,
      Δχ_min::Float64 = 1e-3,
      N_χs::Integer = 100) :: Float64

Return the lensing auto-correlation function $\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta}) = \int_0^{s_1} \mathrm{d} \chi_1 \int_0^{s_2} \mathrm{d} \chi_2 \frac{1}{2} \frac{ \mathcal{H}_0^4 \Omega_{ \mathrm{M0}}^2 D_1 D_2 (\chi_1 - s_1)(\chi_2 - s_2) }{ s_1 s_2 a(\chi_1) a(\chi_2) } (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi))\]

where $D_1 = D(\chi_1)$, $D_2 = D(\chi_2)$ and so on, $\mathcal{H} = a H$, $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2\chi_1\chi_2\cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$) and the $J$ coefficients are given by

\[\begin{align*} J_{00} & = - \frac{3 \chi_1^2 \chi_2^2}{4 \chi^4} (y^2 - 1) (8 y (\chi_1^2 + \chi_2^2) - 9 \chi_1 \chi_2 y^2 - 7 \chi_1 \chi_2) \\ J_{02} & = - \frac{3 \chi_1^2 \chi_2^2}{2 \chi^4} (y^2 - 1) (4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2) \\ J_{31} & = 9 y \chi^2 \\ J_{22} & = \frac{9 \chi_1 \chi_2}{4 \chi^4} [ 2 (\chi_1^4 + \chi_2^4) (7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2) (y^2+1) + \chi_1^2 \chi_2^2 (11 y^4 + 14 y^2 + 23)] \end{align*}\]

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_lensing.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

Δχ_min::Float64 = 1e-6 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\chi\to0^{+}} (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi)) = \frac{4}{15} \, (5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2)\]

So, when it happens that $\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_lensing, integrand_on_mu_lensing integral_on_mu, ξ_multipole

Missing docstring.

Missing docstring for GaSPE.integrand_on_mu_lensing. Check Documenter's build log for details.