Lensing Auto-Correlation functions

GaPSE.integrand_ξ_lensing — Function

 integrand_ξ_lensing(χ1, χ2, s1, s2, y; tol = 0.5, 
      enhancer = 1, Δχ_min = 1e-6) :: Float64

Return the integarnd of the lensing auto-correlation function $\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta})$, i.e. the function $f(s_1, s_2, y, \chi_1, \chi_2)$ defined as follows:

\[f(s_1, s_2, y, \chi_1, \chi_2) = \frac{1}{2} \frac{ \mathcal{H}_0^4 \Omega_{ \mathrm{M0}}^2 D_1 D_2 (\chi_1 - s_1)(\chi_2 - s_2) }{ s_1 s_2 a(\chi_1) a(\chi_2) } (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi))\]

where $D_1 = D(\chi_1)$, $D_2 = D(\chi_2)$ and so on, $\mathcal{H} = a H$, $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2\chi_1\chi_2\cos{\theta}}$ and the $J$ coefficients are given by (with $y = \cos{\theta}$)

\[\begin{align*} J_{00} & = - \frac{3 \chi_1^2 \chi_2^2}{4 \chi^4} (y^2 - 1) (8 y (\chi_1^2 + \chi_2^2) - 9 \chi_1 \chi_2 y^2 - 7 \chi_1 \chi_2) \\ J_{02} & = - \frac{3 \chi_1^2 \chi_2^2}{2 \chi^4} (y^2 - 1) (4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2) \\ J_{31} & = 9 y \chi^2 \\ J_{22} & = \frac{9 \chi_1 \chi_2}{4 \chi^4} [ 2 (\chi_1^4 + \chi_2^4) (7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2) (y^2+1) + \chi_1^2 \chi_2^2 (11 y^4 + 14 y^2 + 23)] \end{align*}\]

Optional arguments

tol = 0.5 : if $s = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 s_2 y} \leq \mathrm{tol}$, then the returned value is 0.0; it prevents computational problems conserning too close points which are insignificants (cosnidering that tol is a distance, so it is measured in $h_0^{-1}\,\mathrm{Mpc}$).

enhancer = 1 : multiply the resulting $f(s_1, s_2, y, \chi_1, \chi_2)$ value; it is very useful for interal computations in other functions (for instance map_integral_on_mu_lensing), in order to deal better with small float numbers.

Δχ_min = 1e-6 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\chi\to0^{+}} (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi)) = \frac{4}{15} \, (5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2)\]

So, when it happens that $\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

See also: ξ_lensing, integrand_on_mu_lensing integral_on_mu_lensing, map_integral_on_mu_lensing

GaPSE.ξ_lensing — Function

 ξ_lensing(s1, s2, y; tol = 0.5, 
      enhancer = 1, Δχ_min = 1e-6, kwargs...) :: Tuple{Float64, Float64}

Return the Lensing Auto-correlation function $\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta})$ , defined as follows:

\[\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta}) = \int_0^{s_1} \mathrm{d} \chi_1 \int_0^{s_2} \mathrm{d} \chi_2 \frac{1}{2} \frac{ \mathcal{H}_0^4 \Omega_{ \mathrm{M0}}^2 D_1 D_2 (\chi_1 - s_1)(\chi_2 - s_2) }{ s_1 s_2 a(\chi_1) a(\chi_2) } (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi))\]

where $D_1 = D(\chi_1)$, $D_2 = D(\chi_2)$ and so on, $\mathcal{H} = a H$, $\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2\chi_1\chi_2\cos{\theta}}$ and the $J$ coefficients are given by (with $y = \cos{\theta}$)

\[\begin{align*} J_{00} & = - \frac{3 \chi_1^2 \chi_2^2}{4 \chi^4} (y^2 - 1) (8 y (\chi_1^2 + \chi_2^2) - 9 \chi_1 \chi_2 y^2 - 7 \chi_1 \chi_2) \\ J_{02} & = - \frac{3 \chi_1^2 \chi_2^2}{2 \chi^4} (y^2 - 1) (4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2) \\ J_{31} & = 9 y \chi^2 \\ J_{22} & = \frac{9 \chi_1 \chi_2}{4 \chi^4} [ 2 (\chi_1^4 + \chi_2^4) (7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2) (y^2+1) + \chi_1^2 \chi_2^2 (11 y^4 + 14 y^2 + 23)] \end{align*}\]

The computation is made applying hcubature (see the Hcubature Julia package) to the integrand function integrand_ξ_lensing.

Optional arguments

tol = 0.5 : if $s = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 s_2 y} \leq \mathrm{tol}$, then the returned value is 0.0; it prevents computational problems conserning too close points which are insignificants (cosnidering that tol is a distance, so it is measured in $h_0^{-1}\,\mathrm{Mpc}$).

enhancer = 1 : multiply the resulting value; it is very useful for interal computations in other functions (for instance map_integral_on_mu_lensing), in order to deal better with small float numbers.

Δχ_min = 1e-6 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\chi\to0^{+}} (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi)) = \frac{4}{15} \, (5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2)\]

So, when it happens that $\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

kwargs... : keyword arguments which should be passed to hcubature when performing the 2-dims integral; we shall recomend to set atol≥1e-4 and rtol≥-3, as a consequence of the fast increase for the evaluation time needed for this integral.

Return

A Tuple{Float64, Float64} : the former is the integral result, the latter its error.

See also: integrand_ξ_lensing, integrand_on_mu_lensing integral_on_mu_lensing, map_integral_on_mu_lensing

Missing docstring.

Missing docstring for GaSPE.integrand_on_mu_lensing. Check Documenter's build log for details.

GaPSE.integral_on_mu_lensing — Function

 integral_on_mu_lensing(s1, s;  L::Integer = 0, enhancer = 1e6, 
      Δχ_min = 1e-6, tol = 0.5, χ_atol = 1e-3, χ_rtol = 1e-3, 
      kwargs...) :: Tuple{Float64, Float64}

Evaluate the integral on $\mu$ of the lensing auto-correlation function $\xi^{\kappa\kappa} (s_1, s_2, \cos{\theta})$, weighted with the window functions $F(x, \mu)$ and $\phi(s_2)$ and the Legendre polynomial $\mathcal{L}(\mu)$, i.e. the following function $f_\mathrm{in}(s_1, s)$:

\[f^{\kappa\kappa}(s_1, s) = \int_{-1}^{+1} \mathrm{d} \mu \; \xi^{\kappa\kappa}(s_1, s, \mu) \, \phi(s_2) \, \mathcal{L}_L(\mu) \, F\left(\frac{s}{s_1}, \mu \right)\]

The computation is made applying quadgk (see the QuadGK Julia package) to the integrand function integrand_on_mu_lensing.

Optional arguments

L::Integer = 0 : Lagrange polynomial degree to be used in the computation.

tol = 0.5 : if $s \leq \mathrm{tol}$, then the integrand value is 0.0; it prevents computational problems conserning too close points which are insignificants (cosnidering that tol is a distance, so it is measured in $h_0^{-1}\,\mathrm{Mpc}$).

enhancer = 1e6 : inside the integration, multiply the results in order to deal with values not too close to zero; at the end, the result is also divided for this number, restoring the correct value.

Δχ_min = 1e-6 : a Float64 parameter used inside integrand_ξ_lensing in order to avoid computatinal divergences; it should be 0<Δχ_min<<1, see the integrand_ξ_lensing docstring for more informations.

χ_atol = 1e-3 : absolute tolerance to be used in the computation of the 2-dims integral on $\chi_1$ and $\chi_2$, made inside ξ_lensing; for computational time reasons, it's better to use χ_atol ≥ 1e-5

χ_rtol = 1e-3 : relative tolerance to be used in the computation of the 2-dims integral on $\chi_1$ and $\chi_2$, made inside ξ_lensing; for computational time reasons, it's better to use χ_rtol ≥ 1e-4.

kwargs... : keyword arguments which should be passed to quadgk when performing the 1-dim integral; we shall recomend to set atol≥1e-4 and rtol≥-3, as a consequence of the fast increase for the evaluation time needed for this integral.

Return

A Tuple{Float64, Float64} : the former is the integral result, the latter its error.

See also: integrand_ξ_lensing, ξ_lensing integrand_on_mu_lensing, map_integral_on_mu_lensing, spline_F, ϕ

GaPSE.map_integral_on_mu_lensing — Function

 map_integral_on_mu_lensing(s1 = s_eff; L::Integer = 0, pr::Bool = true, 
      Δχ_min = 1e-6, χ_atol = 1e-3, χ_rtol = 1e-3, enhancer = 1e6, 
      tol = 0.5, kwargs...
      ) :: Tuple{Vector{Float64}, Vector{Float64}, Vector{Float64}}

Evaluate integral_on_mu_lensing in a range of distance values, returning value and error for each integral.

Arguments

s1 = s_eff : as already mentioned, in this program we use the effective redshift approximation, so the integral on $s_1$ is replaced with an evaluation of the integrand in $s1 = s_eff$.

Optional arguments

L::Integer = 0 : Lagrange polynomial degree to be used in the computation.

pr::Bool = true : tells if the println messages should be printed or not.

tol = 0.5 : if during the evaluation of the integral inside integral_on_mu_lensing happens that $s = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 s_2 y} \leq \mathrm{tol}$, then the integrand value is 0.0; it prevents computational problems conserning too close points which are insignificants (cosnidering that tol is a distance, so it is measured in $h_0^{-1}\,\mathrm{Mpc}$).

enhancer = 1e6 : inside the integration, multiply the results in order to deal with values not too close to zero; at the end, the result is also divided for this number, restoring the correct value.

Δχ_min = 1e-6 : a Float64 parameter used inside integrand_ξ_lensing in order to avoid computatinal divergences; it should be 0<Δχ_min<<1, see the integrand_ξ_lensing docstring for more informations.

χ_atol = 1e-3 : absolute tolerance to be used in the computation of the 2-dims integral on $\chi_1$ and $\chi_2$, made inside ξ_lensing; for computational time reasons, it's better to use χ_atol ≥ 1e-5

χ_rtol = 1e-3 : relative tolerance to be used in the computation of the 2-dims integral on $\chi_1$ and $\chi_2$, made inside ξ_lensing; for computational time reasons, it's better to use χ_rtol ≥ 1e-4.

kwargs... : keyword arguments which should be passed to quadgk when performing the 1-dim integral in integral_on_mu_lensing; we shall recomend to set atol≥1e-4 and rtol≥-3, as a consequence of the fast increase for the evaluation time needed for this integral.

Return

A Tuple{Vector{Float64}, Vector{Float64}, Vector{Float64}}:

the first Vector{Float64} contains the distance s where the integarl is evaluated;

the second one the integral results at the corresponding distance;

the third and last one contains the error in the integral estimation, as returned by integral_on_mu_lensing

See also: z_eff, z_eff, integrand_ξ_lensing, ξ_lensing, integrand_on_mu_lensing, integral_on_mu_lensing, print_map_integral_on_mu_lensing

GaPSE.print_map_int_on_mu_lensing — Function

 print_map_int_on_mu_lensing(out::String; L::Integer = 0,
      s1 = s_eff, pr::Bool = true, kwargs...)

Print the output of map_integral_on_mu_lensing to the given input filename out, with some meta-data at the beginning of it. If that file already exists, it will be destroyed an re-created.

Arguments

s1 = s_eff : as already mentioned, in this program we use the effective redshift approximation, so the integral on $s_1$ is replaced with an evaluation of the integrand in $s1 = s_eff$.

Optional arguments

kwargs... : all the keyword arguments supported by map_integral_on_mu_lensing; see its docstring for more details.

See also: z_eff, z_eff, integrand_ξ_lensing, ξ_lensing, integrand_on_mu_lensing, integral_on_mu_lensing, map_integral_on_mu_lensing