Background functions

GaPSE.f0 — Constant

 const f0 :: Float64

Linear growth rate at present time. Its value is equal to:

\[ f_0 \simeq 0.5126998572951\]

GaPSE.D0 — Constant

 const D0 :: Float64

Linear growth factor at present time. Its value is equal to:

\[ D_0 = 1.0\]

GaPSE.ℋ0 — Constant

 const ℋ0 :: Float64

Comoving Hubble constant at present time. Its value is, in natural system (where the speed of light c=1): $\mathcal{H}_0 \simeq 3.335641\times10^{-4} \; h_0^{-1}\mathrm{Mpc}$

GaPSE.BackgroundData — Type

 BackgroundData(
      z::Vector{Float64}
      conftime::Vector{Float64}
      comdist::Vector{Float64}
      angdist::Vector{Float64}
      lumdist::Vector{Float64}
      D::Vector{Float64}
      f::Vector{Float64}
      ℋ::Vector{Float64}
      ℋ_p::Vector{Float64})

Struct that contains all the relevant cosmological information for future computations. The data are stored with increasing distance values (so the first ones are associated to z=0). It is internally used in Cosmology.

Arguments

z::Vector{Float64} : redshifts (adimensionals).

conftime::Vector{Float64} : conformal times, measured in [Mpc/h].

comdist::Vector{Float64} : comoving distances, measured in [Mpc/h].

angdist::Vector{Float64} : angular diameter distances, measured in [Mpc/h].

lumdist::Vector{Float64} : luminosity distances, measured in [Mpc/h].

D::Vector{Float64} : linear growth factors, normalized to 1.0 at the present day (adimensional).

f::Vector{Float64} : linear growth rates (adimensional).

ℋ::Vector{Float64} : comoving Hubble parameters, measured in [h/Mpc].

ℋ_p::Vector{Float64} : derivatives of the comoving Hubble parameter wrt the conformal time. It is here manually computed with the Dierckx function derivative.

Constructors

BackgroundData(file::String, z_max; names = NAMES_BACKGROUND, h = 0.7)

file::string : input file where the data are stored; it is expected that such file is a background output of the CLASS program (link: https://github.com/lesgourg/class_public)

z_max : the maximum redhsift we are interested in our analysis. The constructor will store the data necessary for a study only in 0 < z < z_max, for optimisation purposes (More precisely, the maximum distance stored will be 3*z_max).

names = NAMES_BACKGROUND : the column names of the file. If the colum order change from the default one NAMES_BACKGROUND, you must set as input the vector of string with the correct one, with the SAME names. They are, with the default order:

z , proper time [Gyr] , conf. time [Mpc] , H [1/Mpc] , comov. dist. , ang.diam.dist. , lum. dist. , comov.snd.hrz. , (.)rho_g , (.)rho_b , (.)rho_cdm , (.)rho_lambda , (.)rho_ur , (.)rho_crit , gr.fac. D , gr.fac. f ,

h = 0.7 : the adimensional hubble constant. By default, CLASS background data are measured with it numerically expressed (so distances are measured in Mpc, for example), while this code works with h in the unit of measure (so distances are measured in Mpc/h, for example). Change this value to 1.0 if the input data do not have this issue, or to your value of interest (0.67, 0.5, ...).

See also: CosmoParams, Cosmology