Cross-Correlation functions

GaPSE.ξ_Lensing_Doppler — Function

 ξ_Lensing_Doppler(s1, s2, y, cosmo::Cosmology;
      en::Float64 = 1e6, N_χs::Integer = 100):: Float64

Return the Lensing-Doppler cross-correlation function $\xi^{v_{\parallel}\kappa} (s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{v_{\parallel}\kappa} (s_1, s_2, \cos{\theta}) = \mathcal{H}_0^2 \Omega_{M0} D(s_2) f(s_2) \mathcal{H}(s_2) \mathcal{R}(s_2) \int_0^{s_1} \mathrm{d} \chi_1 \frac{ D(\chi_1) (\chi_1 - s_1) }{a(\chi_1) s_1} \left( J_{00} I^0_0(\Delta\chi_1) + J_{02} I^0_2(\Delta\chi_1) + J_{04} I^0_4(\Delta\chi_1) + J_{20} I^2_0(\Delta\chi_1) \right)\]

where $\mathcal{H} = a H$, $\Delta\chi_1= \sqrt{\chi_1^2 + s_2^2 - 2 \chi_1 s_2 \cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$) and the $J$ coefficients are given by:

\[\begin{align*} J_{00} & = \frac{1}{15}(\chi_1^2 y + \chi_1(4 y^2 - 3) s_2 - 2 y s_2^2) \\ J_{02} & = \frac{1}{42 \Delta\chi_1^2} (4 \chi_1^4 y + 4 \chi_1^3 (2 y^2 - 3) s_2 + \chi_1^2 y (11 - 23 y^2) s_2^2 + \chi_1 (23 y^2 - 3) s_2^3 - 8 y s_2^4) \\ J_{04} & = \frac{1}{70 \Delta\chi_1^2} (2 \chi_1^4 y + 2 \chi_1^3 (2y^2 - 3) s_2 - \chi_1^2 y (y^2 + 5) s_2^2 + \chi_1 (y^2 + 9) s_2^3 - 4 y s_2^4) \\ J_{20} & = y \Delta\chi_1^2 \end{align*}\]

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_Lensing_Doppler.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

Δχ_min::Float64 = 1e-6 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\chi\to0^{+}} (J_{00} \, I^0_0(\chi) + J_{02} \, I^0_2(\chi) + J_{31} \, I^3_1(\chi) + J_{22} \, I^2_2(\chi)) = \frac{4}{15} \, (5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2)\]

So, when it happens that $\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_Lensing_Doppler, int_on_mu_Lensing_Doppler integral_on_mu, ξ_multipole

Missing docstring.

Missing docstring for GaPSE.ξ_Doppler_Lensing. Check Documenter's build log for details.

GaPSE.ξ_Doppler_LocalGP — Function

 ξ_Doppler_LocalGP(P1::Point, P2::Point, y, cosmo::Cosmology) :: Float64

Return the Doppler-LocalGP cross-correlation function, defined as follows:

\[\xi^{v_{\parallel}\phi} (s_1, s_2, \cos{\theta}) = \frac{3}{2 a(s_2)} \mathcal{H}(s_1) f(s_1) D(s_1) \mathcal{R}(s_1) \mathcal{H}_0^2 \Omega_{M0} D(s_2) (1 + \mathcal{R}(s_2)) (s_1 - s_2\cos{\theta}) s^2 I^3_1(s)\]

where $\mathcal{H} = a H$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$ and :

\[I^n_l(s) = \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_l(qs)}{(q s)^n}\]

Inputs

P1::Point and P2::Point: Point where the CF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on)

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

See also: Point, Cosmology

Missing docstring.

Missing docstring for GaPSE.ξ_LocalGP_Doppler. Check Documenter's build log for details.

GaPSE.ξ_Doppler_IntegratedGP — Function

 ξ_Doppler_IntegratedGP(s1, s2, y, cosmo::Cosmology;
      en::Float64 = 1e6, N_χs::Integer = 100):: Float64

Return the Doppler-LocalGP cross-correlation function $\xi^{v_{\parallel}\int\phi} (s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{v_{\parallel}\int\phi} (s_1, s_2, \cos{\theta}) = 3 \mathcal{H}(s_1) f(s_1) D(s_1) \mathcal{H_0}^2 \Omega_{M0} \mathcal{R}(s_1) \int_0^{s_2} \mathrm{d}\chi_2 \, J_{31} \, I^3_1(\chi)\]

where $\mathcal{H} = a H$, $\chi = \sqrt{s_1^2 + \chi_2^2 - 2 s_1 \chi_2 \cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$) and:

\[J_{31} = \frac{D(\chi_2) (s_1 - \chi_2 \cos{\theta})}{a(\chi_2)} \chi^2 \left( - \frac{1}{s_2} + \mathcal{R}(s_2) \mathcal{H}(\chi_2) (f(\chi_2) - 1) \right)\]

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_Doppler_IntegratedGP.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_Doppler_IntegratedGP, int_on_mu_Doppler_IntegratedGP integral_on_mu, ξ_multipole

Missing docstring.

Missing docstring for GaPSE.ξ_IntegratedGP_Doppler. Check Documenter's build log for details.

GaPSE.ξ_Lensing_LocalGP — Function

 ξ_Lensing_LocalGP(s1, s2, y, cosmo::Cosmology;
      en::Float64 = 1e6, N_χs::Integer = 100):: Float64

Return the Lensing-LocalGP cross-correlation function $\xi^{\kappa \phi} (s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{\kappa \phi} (s_1, s_2, \cos{\theta}) = \frac{ 9 \mathcal{H}_0^4 \Omega_{M0}^2 D(s_2) (1 + \mathcal{R}(s_2)) s_2 }{4 a(s_2) s_1} \int_0^{s_1} \mathrm{d}\chi_1 \frac{D(\chi_1)(s_1 - \chi_1) }{a(\chi_1)} \left( J_{31} I^3_1(\Delta\chi_1) + J_{22} I^2_2(\Delta\chi_1) \right)\]

where $\mathcal{H} = a H$, $\Delta\chi_1 = \sqrt{\chi_1^2 + s_2^2 - 2 \chi_1 s_2\cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$) and the $J$ coefficients are given by

\[\begin{align*} J_{31} & = -2 y \Delta\chi_1^2 \\ J_{22} & = \chi_1 s_2 (1 - y^2) \end{align*}\]

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_Lensing_LocalGP.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_Lensing_LocalGP, int_on_mu_Lensing_LocalGP integral_on_mu, ξ_multipole

Missing docstring.

Missing docstring for GaPSE.ξ_LocalGP_Lensing, GaPSE.ξ_Lensing_IntegratedGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_IntegratedGP_Lensing. Check Documenter's build log for details.

GaPSE.ξ_LocalGP_IntegratedGP — Function

 ξ_LocalGP_IntegratedGP(s1, s2, y, cosmo::Cosmology;
      en::Float64 = 1e6, N_χs::Integer = 100):: Float64

Return the LocalGP-IntegratedGP cross-correlation function $\xi^{v_{\parallel}\int \phi} (s_1, s_2, \cos{\theta})$, defined as follows:

\[\xi^{v_{\parallel}\int \phi} (s_1, s_2, \cos{\theta}) = \frac{9 \mathcal{H}_0^4 \Omega_{M0}^2 D(s_1) (\mathcal{R}(s_1) +1)}{2 a(s_1)} \int_0^{s_2} \mathrm{d}\chi_2 \frac{D(\chi_2) \Delta\chi_2^4}{ a(\chi_2)} \left( \mathcal{H}(\chi_2)( f(\chi_2) - 1) \mathcal{R}(s_2) - \frac{1}{s_2} \right) \tilde{I}^4_0(\Delta\chi_2)\]

where $\mathcal{H} = a H$, $\Delta\chi_2 = \sqrt{s_1^2 + \chi_2^2 - 2 s_1 \chi_2 \cos{\theta}}$, $y = \cos{\theta} = \hat{\mathbf{s}}_1 \dot \hat{\mathbf{s}}_2$).

The computation is made applying trapz (see the Trapz Julia package) to the integrand function integrand_ξ_LocalGP_IntegratedGP.

Inputs

s1 and s2: comovign distances where the function must be evaluated

y: the cosine of the angle between the two points P1 and P2

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs::Integer = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s1) (for χ2); it has been checked that with N_χs ≥ 50 the result is stable.

See also: integrand_ξ_LocalGP_IntegratedGP, int_on_mu_LocalGP_IntegratedGP integral_on_mu, ξ_multipole

Missing docstring.

Missing docstring for GaPSE.ξ_IntegratedGP_LocalGP. Check Documenter's build log for details.