GNC TPCFs

Two-Point Auto-Correlation Functions

GaPSE.ξ_GNC_Newtonian — Function

ξ_GNC_Newtonian(P1::Point, P2::Point, y, cosmo::Cosmology; 
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel) ::Float64

ξ_GNC_Newtonian(s1, s2, y, cosmo::Cosmology; 
    kwargs...) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Newtonian auto-correlation effect arising from the Galaxy Number Counts (GNC).

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{\delta\delta} ( s_1 , s_2 , y ) = D_1 D_2 \left[ J_{00}^{\delta\delta} I_0^0 (s) + J_{02}^{\delta\delta} I_2^0 (s) + J_{04}^{\delta\delta} I_4^0 (s) \right] \, , \end{split}\]

with

\[\begin{split} J_{00}^{\delta\delta} &= \frac{1}{15} \left[ 15 b_1 b_2 + 5 b_1 f_2 + 5 b_2 f_1 + (2 y^2 + 1) f_1 f_2 \right] \, , \\ %%%%%%%%%%%%%%%%%% J_{02}^{\delta\delta} &= - \frac{1}{21 s^2} \left\{ s_1^2 \left[ 14 b_2 f_1 + 7 b_1 f_2 (3 y^2 - 1) + (11 y^2 + 1) f_1 f_2 \right] \right. \\ &\left.\qquad \qquad + s_2^2 \left[ 14 b_1 f_2 + 7 b_2 f_1 (3 y^2 - 1) + (11 y^2 + 1) f_1 f_2 \right] \right. \\ &\left. \qquad \qquad - 4 y s_1 s_2 \left[ 7 b_2 f_1 + 7 b_1 f_2 + (y^2 + 5) f_1 f_2 \right] \right\} \, , \\ %%%%%%%%%%%%%%%%%% J_{04}^{\delta\delta} &= \frac{f_1 f_2}{35 s^4} \left[ 4 (3 y^2 - 1) (s_1^4 + s_2^4) + 3 (3 + y^2)^2 s_1^2 s_2^2 - 8 y s_1 s_2 (s_1^2 + s_2^2) (3 + y^2) \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathcal{R}_1 = \mathcal{R}(s_1)$, ... is computed by func_ℛ_GNC in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathcal{R}(s) = 5 s_{\mathrm{b}}(s) + \frac{2 - 5 s_{\mathrm{b}}(s)}{\mathcal{H}(s) \, s} + \frac{\dot{\mathcal{H}}(s)}{\mathcal{H}(s)^2} - \mathit{f}_{\mathrm{evo}} \quad ;\]

$b_1$, $s_{\mathrm{b}, 1}$, $\mathit{f}_{\mathrm{evo}, 1}$ (and $b_2$, $s_{\mathrm{b}, 2}$, $\mathit{f}_{\mathrm{evo}, 2}$) : galaxy bias, magnification bias (i.e. the slope of the luminosity function at the luminosity threshold), and evolution bias for the first (second) effect;

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

In this TPCF there are no observer terms. The obs keyword is inserted only for compatibility with the other GNC TPCFs.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

b1=nothing, s_b1=nothing, 𝑓_evo1=nothing and b2=nothing, s_b2=nothing, 𝑓_evo2=nothing: galaxy, magnification and evolutionary biases respectively for the first and the second effect computed in this TPCF:

if not set (i.e. if you leave the default value nothing) the values stored in the input cosmo will be considered;

if you set one or more values, they will override the cosmo ones in this computation;

the two sets of values should be different only if you are interested in studing two galaxy species;

only the required parameters for the chosen TPCF will be used, depending on its analytical expression; all the others will have no effect, we still inserted them for pragmatical code compatibilities.

s_lim=nothing : parameter used in order to avoid the divergence of the $\mathcal{R}$ and $\mathfrak{R}$ denominators: when $0 \leq s \leq s_\mathrm{lim}$ the returned values are

\[\forall \, s \in [ 0, s_\mathrm{lim} ] \; : \quad \mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}_0 \, s_\mathrm{lim}} \; , \quad \mathcal{R}(s) = 5 s_{\mathrm{b}} + \frac{2 - 5 s_{\mathrm{b}}}{\mathcal{H}_0 \, s_\mathrm{lim}} + \frac{\dot{\mathcal{H}}}{\mathcal{H}_0^2} - \mathit{f}_{\mathrm{evo}} \; .\]

If nothing, the default value stored in cosmo will be considered.

obs::Union{Bool,Symbol} = :noobsvel : do you want to consider the observer terms in the computation of the chosen GNC TPCF effect?

:yes or true -> all the observer effects will be considered

:no or false -> no observer term will be taken into account

:noobsvel -> the observer terms related to the observer velocity (that you can find in the CF concerning Doppler) will be neglected, the other ones will be taken into account

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_GNC_multipole, print_map_ξ_GNC_multipole

GaPSE.ξ_GNC_Doppler — Function

ξ_GNC_Doppler(P1::Point, P2::Point, y, cosmo::Cosmology;
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel) ::Float64

ξ_GNC_Doppler(s1, s2, y, cosmo::Cosmology; 
    kwargs...) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Doppler auto-correlation effect arising from the Galaxy Number Counts (GNC).

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{v_{\parallel} v_{\parallel}}( s_1, s_2 , y ) &= D_1 D_2 J_{\alpha}^{v_{\parallel} v_{\parallel}} \left[ J_{\beta}^{v_{\parallel} v_{\parallel}} \left( \frac{1}{45} I_0^0(s) + \frac{2}{63} I_2^0 (s) + \frac{1}{105} I_4^0(s) \right) + J^{v_{\parallel} v_{\parallel}}_{20} I_0^2 (s) \right] \\ & + D_1 \left[ J^{v_{\parallel} v_{\parallel}}_{31} ( s_1 , s_2) I_1^3 (s_1) + J^{v_{\parallel} v_{\parallel}}_{11} ( s_1 , s_2) I_1^1 (s_1) + J^{v_{\parallel} v_{\parallel}}_{13} ( s_1 ,s_2) I_3^1 (s_1) \right] \\ & + D_2 \left[ J^{v_{\parallel} v_{\parallel}}_{31} ( s_2 , s_1) I_1^3 (s_2) + J^{v_{\parallel} v_{\parallel}}_{11} ( s_2 , s_1) I_1^1 (s_2) + J^{v_{\parallel} v_{\parallel}}_{13} ( s_2 , s_1) I_3^1 (s_2) \right] \\ & + J^{v_{\parallel} v_{\parallel}}_{\sigma2} \sigma_2 \, , \end{split}\]

with

\[\begin{split} J_{\alpha}^{v_{\parallel} v_{\parallel}} &= f_1 f_2\mathcal{H}_1 \mathcal{H}_2 \mathcal{R}_1 \mathcal{R}_2 \, , \\ %%%%%%%%%%%%%% J_{\beta}^{v_{\parallel} v_{\parallel}} &= y^2 s_1 s_2 - 2 y (s_1^2 + s_2^2) + 3 s_1 s_2 \, , \\ %%%%%%%%%%%%%% J_{20}^{v_{\parallel} v_{\parallel}} &= \frac{1}{3} y s^2 \, , \\ %%%%%%%%%%%%%% J_{31}^{v_{\parallel} v_{\parallel}} (s_1, s_2) &= -y f_0 \mathcal{H}_0 s_1^2 f_1 \mathcal{R}_1 (\mathcal{R}_2 - 5 s_{\mathrm{b}, 2} + 2) \, , \\ %%%%%%%%%%%%%% J_{11}^{v_{\parallel} v_{\parallel}} (s_1, s_2) &= \frac{1}{5} y f_0 \mathcal{H}_0 s_1^2 f_1 \mathcal{H}_1 \mathcal{R}_1 (\mathcal{R}_2 - 5 s_{\mathrm{b}, 2} + 2) \, , \\ %%%%%%%%%%%%%% J_{13}^{v_{\parallel} v_{\parallel}} (s_1, s_2) &= \frac{1}{5} y f_0 \mathcal{H}_0 s_1^2 f_1 \mathcal{H}_1 \mathcal{R}_1 (\mathcal{R}_2 - 5 s_{\mathrm{b}, 2} + 2) \, , \\ %%%%%%%%%%%%%% J_{\sigma2}^{v_{\parallel} v_{\parallel}} &= \frac{1}{3} y f_0^2 \mathcal{H}_0^2 (\mathcal{R}_1 - 5 s_{\mathrm{b}, 1} + 2) (\mathcal{R}_2 - 5 s_{\mathrm{b}, 2} + 2) \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathcal{R}_1 = \mathcal{R}(s_1)$, ... is computed by func_ℛ_GNC in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathcal{R}(s) = 5 s_{\mathrm{b}}(s) + \frac{2 - 5 s_{\mathrm{b}}(s)}{\mathcal{H}(s) \, s} + \frac{\dot{\mathcal{H}}(s)}{\mathcal{H}(s)^2} - \mathit{f}_{\mathrm{evo}} \quad ;\]

$b_1$, $s_{\mathrm{b}, 1}$, $\mathit{f}_{\mathrm{evo}, 1}$ (and $b_2$, $s_{\mathrm{b}, 2}$, $\mathit{f}_{\mathrm{evo}, 2}$) : galaxy bias, magnification bias (i.e. the slope of the luminosity function at the luminosity threshold), and evolution bias for the first (second) effect;

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

In this TPCF, the only non-observer term is the first one (i.e. the bracket proportional to $D(s_1) \, D(s_2)$). The other three ones are all observer terms related to the observer velocity, so if you set obs = :no, obs = false or even obs = :noobsvel they will not be computed.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

b1=nothing, s_b1=nothing, 𝑓_evo1=nothing and b2=nothing, s_b2=nothing, 𝑓_evo2=nothing: galaxy, magnification and evolutionary biases respectively for the first and the second effect computed in this TPCF:

if not set (i.e. if you leave the default value nothing) the values stored in the input cosmo will be considered;

if you set one or more values, they will override the cosmo ones in this computation;

the two sets of values should be different only if you are interested in studing two galaxy species;

only the required parameters for the chosen TPCF will be used, depending on its analytical expression; all the others will have no effect, we still inserted them for pragmatical code compatibilities.

s_lim=nothing : parameter used in order to avoid the divergence of the $\mathcal{R}$ and $\mathfrak{R}$ denominators: when $0 \leq s \leq s_\mathrm{lim}$ the returned values are

\[\forall \, s \in [ 0, s_\mathrm{lim} ] \; : \quad \mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}_0 \, s_\mathrm{lim}} \; , \quad \mathcal{R}(s) = 5 s_{\mathrm{b}} + \frac{2 - 5 s_{\mathrm{b}}}{\mathcal{H}_0 \, s_\mathrm{lim}} + \frac{\dot{\mathcal{H}}}{\mathcal{H}_0^2} - \mathit{f}_{\mathrm{evo}} \; .\]

If nothing, the default value stored in cosmo will be considered.

obs::Union{Bool,Symbol} = :noobsvel : do you want to consider the observer terms in the computation of the chosen GNC TPCF effect?

:yes or true -> all the observer effects will be considered

:no or false -> no observer term will be taken into account

:noobsvel -> the observer terms related to the observer velocity (that you can find in the CF concerning Doppler) will be neglected, the other ones will be taken into account

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_GNC_multipole, print_map_ξ_GNC_multipole

GaPSE.ξ_GNC_Lensing — Function

ξ_GNC_Lensing(P1::Point, P2::Point, y, cosmo::Cosmology;
    en::Float64 = 1e6, Δχ_min::Float64 = 1e-1,
    N_χs_2::Int = 100,
    s_b1=nothing, s_b2=nothing, 𝑓_evo1=nothing, 𝑓_evo2=nothing,
    s_lim=nothing, obs::Union{Bool,Symbol}=:noobsvel,
    suit_sampling::Bool=true ) ::Float64

ξ_GNC_Lensing(s1, s2, y, cosmo::Cosmology; 
    kwargs...) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Lensing auto-correlation effect arising from the Galaxy Number Counts (GNC).

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{\kappa\kappa} (s_1, s_2, y) = \int_0^{s_1} \mathrm{d}\chi_1 \int_0^{s_2} \mathrm{d}\chi_2\; J^{\kappa\kappa}_{\alpha} &\left[ J^{\kappa\kappa}_{00} I_0^0(\Delta\chi) + J^{\kappa\kappa}_{02} I_2^0(\Delta\chi) + \right.\\ &\left. J^{\kappa\kappa}_{31} I_1^3(\Delta\chi) + J^{\kappa\kappa}_{22} I_2^2(\Delta\chi) \right] \, , \end{split}\]

with

\[\begin{split} J^{\kappa\kappa}_{\alpha} & = \frac{ \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D(\chi_1) D(\chi_2) }{ s_1 s_2 a(\chi_1) a(\chi_2)} (\chi_1 - s_1)(\chi_2 - s_2) (5 s_{\mathrm{b}, 1} - 2)(5 s_{\mathrm{b}, 2} - 2) \, , \\ %%%%&%%%%%%%%%%%%% J^{\kappa\kappa}_{00} & = -\frac{ 3 \chi_1^2 \chi_2^2}{4 \Delta\chi^4} (y^2 - 1) \left[ 8 y (\chi_1^2 + \chi_2^2) - 9\chi_1\chi_2y^2 - 7\chi_1\chi_2 \right] \, , \\ %%%%&%%%%%%%%%%%%% J^{\kappa\kappa}_{02} & = -\frac{ 3 \chi_1^2 \chi_2^2}{2 \Delta\chi^4}(y^2 - 1) \left[ 4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2 \right] \, , \\ %%%%%%%%%%%%%%%%%% J^{\kappa\kappa}_{31} & = 9 y \Delta\chi^2 \, , \\ %%%%%%%%%%%%%%%%%% J^{\kappa\kappa}_{22} & = \frac{9 \chi_1 \chi_2}{4 \Delta\chi^4} \left[ 2(\chi_1^4 + \chi_2^4)(7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2)(y^2 + 1) + \right.\\ &\left.\qquad\qquad\qquad \chi_1^2 \chi_2^2 (11y^4 + 14y^2 + 23) \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathcal{R}_1 = \mathcal{R}(s_1)$, ... is computed by func_ℛ_GNC in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathcal{R}(s) = 5 s_{\mathrm{b}}(s) + \frac{2 - 5 s_{\mathrm{b}}(s)}{\mathcal{H}(s) \, s} + \frac{\dot{\mathcal{H}}(s)}{\mathcal{H}(s)^2} - \mathit{f}_{\mathrm{evo}} \quad ;\]

$b_1$, $s_{\mathrm{b}, 1}$, $\mathit{f}_{\mathrm{evo}, 1}$ (and $b_2$, $s_{\mathrm{b}, 2}$, $\mathit{f}_{\mathrm{evo}, 2}$) : galaxy bias, magnification bias (i.e. the slope of the luminosity function at the luminosity threshold), and evolution bias for the first (second) effect;

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

In this TPCF there are no observer terms. The obs keyword is inserted only for compatibility with the other GNC TPCFs.

This function is computed integrating integrand_ξ_GNC_Lensing with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

b1=nothing, s_b1=nothing, 𝑓_evo1=nothing and b2=nothing, s_b2=nothing, 𝑓_evo2=nothing: galaxy, magnification and evolutionary biases respectively for the first and the second effect computed in this TPCF:

if not set (i.e. if you leave the default value nothing) the values stored in the input cosmo will be considered;

if you set one or more values, they will override the cosmo ones in this computation;

the two sets of values should be different only if you are interested in studing two galaxy species;

only the required parameters for the chosen TPCF will be used, depending on its analytical expression; all the others will have no effect, we still inserted them for pragmatical code compatibilities.

s_lim=nothing : parameter used in order to avoid the divergence of the $\mathcal{R}$ and $\mathfrak{R}$ denominators: when $0 \leq s \leq s_\mathrm{lim}$ the returned values are

\[\forall \, s \in [ 0, s_\mathrm{lim} ] \; : \quad \mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}_0 \, s_\mathrm{lim}} \; , \quad \mathcal{R}(s) = 5 s_{\mathrm{b}} + \frac{2 - 5 s_{\mathrm{b}}}{\mathcal{H}_0 \, s_\mathrm{lim}} + \frac{\dot{\mathcal{H}}}{\mathcal{H}_0^2} - \mathit{f}_{\mathrm{evo}} \; .\]

If nothing, the default value stored in cosmo will be considered.

obs::Union{Bool,Symbol} = :noobsvel : do you want to consider the observer terms in the computation of the chosen GNC TPCF effect?

:yes or true -> all the observer effects will be considered

:no or false -> no observer term will be taken into account

:noobsvel -> the observer terms related to the observer velocity (that you can find in the CF concerning Doppler) will be neglected, the other ones will be taken into account

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

Δχ_min::Float64 = 1e-4 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\Delta\chi \to 0^{+}} \left( J_{00}^{\kappa\kappa} \, I^0_0(\Delta\chi) + J_{02}^{\kappa\kappa} \, I^0_2(\Delta\chi) + J_{31}^{\kappa\kappa} \, I^3_1(\Delta\chi) + J_{22}^{\kappa\kappa} \, I^2_2(\Delta\chi) \right) = 3 \, \sigma_2 + \frac{6}{5} \, \chi_2^2 \, \sigma_0\]

So, when it happens that $\Delta\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

N_χs_2::Int = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s2) (for χ2); it has been checked that with N_χs_2 ≥ 50 the result is stable.

suit_sampling::Bool = true : this bool keyword can be found in all the TPCFs which have at least one χ integral; it is conceived to enable a sampling of the χ integral(s) suited for the given TPCF; however, it actually have an effect only in the TPCFs that have such a sampling implemented in the code. Currently, only ξ_GNC_Newtonian_Lensing (and its simmetryc TPCF) has it.

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_GNC_multipole, print_map_ξ_GNC_multipole, integrand_ξ_GNC_Lensing

GaPSE.ξ_GNC_LocalGP — Function

 ξ_GNC_LocalGP(P1::Point, P2::Point, y, cosmo::Cosmology; 
      obs::Union{Bool, Symbol} = :noobsvel) ::Float64

 ξ_GNC_LocalGP(s1, s2, y, cosmo::Cosmology; 
      kwargs...) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Local Gravitational Potential (GP) auto-correlation effect arising from the Galaxy Number Counts (GNC).

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{\phi\phi}( s_1 , s_2, y ) &= D_1 D_2 J_{40}^{\phi\phi}( s_1, s_2 ) \tilde{I}_0^4 (s) + D_2 J_{40}^{\phi_0 \phi}(s_1,s_2) \tilde{I}_0^4 (s_2) \\ &+ D_1 J_{40}^{\phi_0 \phi}(s_2,s_1) \tilde{I}_0^4 (s_1) + J_{\sigma4}^{\phi\phi} (s_1, s_2) \sigma_4 \, , \end{split}\]

with

\[\begin{split} J_{40}^{\phi\phi} (s_1, s_2) &= \frac{s^4}{4 a_1 a_2} \left[ 2 a_2 f_2 (\mathit{f}_{\mathrm{evo}, 2} - 3) \mathcal{H}_2^2 + 3 \mathcal{H}_0^2 \Omega_{\mathrm{M}0} (f_2 + \mathcal{R}_2 + 5 s_{\mathrm{b}, 2} - 2) \right] \times \\ &\qquad\qquad \left[ 2 a_1 f_1(\mathit{f}_{\mathrm{evo}, 1}-3) \mathcal{H}_1^2 + 3 \mathcal{H}_0^2 \Omega_{\mathrm{M}0} (f_1 + \mathcal{R}_1 + 5 s_{\mathrm{b}, 1} - 2) \right] \, ,\\ %%%%%%%%%%%%%%%%% J_{40}^{\phi_0 \phi}(s_1, s_2) &= \frac{\mathcal{H}_0 s_2^4}{4 s_1 a_2} \left[ \mathcal{H}_0 s_1 (2 f_0-3 \Omega_{\mathrm{M}0}) \mathcal{R}_1 + 2 f_0 (5 s_{\mathrm{b}, 1}-2) \right] \times \\ &\qquad\qquad\left[ 2 a_2 f_2 (\mathit{f}_{\mathrm{evo}, 2} - 3) \mathcal{H}_2^2 + 3 \mathcal{H}_0^2 \Omega_{\mathrm{M}0} (f_2 + \mathcal{R}_2 + 5 s_{\mathrm{b}, 2} - 2) \right] \, , \\ %%%%%%%%%%%%%%%%%%%% J_{\sigma 4}(s_1, s_2)&= \frac{\mathcal{H}_0^2 }{4 s_1 s_2} \left[ \mathcal{H}_0 s_1 (2 f_0-3 \Omega_{\mathrm{M}0}) \mathcal{R}_1 + 2 f_0(5 s_{\mathrm{b}, 1}-2) \right] \times \\ &\qquad\qquad\left[ 3 \Omega_{\mathrm{M}0} \mathcal{H}_0 s_2 \mathcal{R}_2 + 2 f_0 (\mathcal{H}_0 s_2 \mathcal{R}_2 + 5 s_{\mathrm{b}, 2} - 2) \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathcal{R}_1 = \mathcal{R}(s_1)$, ... is computed by func_ℛ_GNC in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathcal{R}(s) = 5 s_{\mathrm{b}}(s) + \frac{2 - 5 s_{\mathrm{b}}(s)}{\mathcal{H}(s) \, s} + \frac{\dot{\mathcal{H}}(s)}{\mathcal{H}(s)^2} - \mathit{f}_{\mathrm{evo}} \quad ;\]

$b_1$, $s_{\mathrm{b}, 1}$, $\mathit{f}_{\mathrm{evo}, 1}$ (and $b_2$, $s_{\mathrm{b}, 2}$, $\mathit{f}_{\mathrm{evo}, 2}$) : galaxy bias, magnification bias (i.e. the slope of the luminosity function at the luminosity threshold), and evolution bias for the first (second) effect;

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

In this TPCF, the only non-observer term is the first one (i.e. the bracket proportional to $D(s_1) \, D(s_2)$), so if you set obs = :no or obs = false that is the only one computed. The other three terms are all observer terms, but not related to the observer velocity, so if you set obs = :noobsvel they will still be computed.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

b1=nothing, s_b1=nothing, 𝑓_evo1=nothing and b2=nothing, s_b2=nothing, 𝑓_evo2=nothing: galaxy, magnification and evolutionary biases respectively for the first and the second effect computed in this TPCF:

if not set (i.e. if you leave the default value nothing) the values stored in the input cosmo will be considered;

if you set one or more values, they will override the cosmo ones in this computation;

the two sets of values should be different only if you are interested in studing two galaxy species;

only the required parameters for the chosen TPCF will be used, depending on its analytical expression; all the others will have no effect, we still inserted them for pragmatical code compatibilities.

s_lim=nothing : parameter used in order to avoid the divergence of the $\mathcal{R}$ and $\mathfrak{R}$ denominators: when $0 \leq s \leq s_\mathrm{lim}$ the returned values are

\[\forall \, s \in [ 0, s_\mathrm{lim} ] \; : \quad \mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}_0 \, s_\mathrm{lim}} \; , \quad \mathcal{R}(s) = 5 s_{\mathrm{b}} + \frac{2 - 5 s_{\mathrm{b}}}{\mathcal{H}_0 \, s_\mathrm{lim}} + \frac{\dot{\mathcal{H}}}{\mathcal{H}_0^2} - \mathit{f}_{\mathrm{evo}} \; .\]

If nothing, the default value stored in cosmo will be considered.

obs::Union{Bool,Symbol} = :noobsvel : do you want to consider the observer terms in the computation of the chosen GNC TPCF effect?

:yes or true -> all the observer effects will be considered

:no or false -> no observer term will be taken into account

:noobsvel -> the observer terms related to the observer velocity (that you can find in the CF concerning Doppler) will be neglected, the other ones will be taken into account

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_GNC_multipole, print_map_ξ_GNC_multipole

GaPSE.ξ_GNC_IntegratedGP — Function

ξ_GNC_IntegratedGP(
    P1::Point, P2::Point, y, cosmo::Cosmology;
    en::Float64=1e10, N_χs_2::Int=100, 
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool,Symbol}=:noobsvel,
    suit_sampling::Bool=true
    ) ::Float64

ξ_GNC_IntegratedGP(s1, s2, y, cosmo::Cosmology; 
    kwargs...) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Integrated Gravitational Potential (GP) auto-correlation effect arising from the Galaxy Number Counts (GNC).

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{\int\!\phi \int \!\phi }( s_1 , s_2, y ) = \int_0^{s_1}\mathrm{d} \chi_1 \int_0^{s_2}\mathrm{d} \chi_2 \; J^{\int \!\phi \int \!\phi}_{40} \tilde{I}_0^4 ( \Delta\chi) \, , \end{split}\]

with

\[\begin{split} J^{\int \!\phi\int \!\phi}_{40} = \frac{ 9 \Delta\chi ^4 \mathcal{H}_0^4 \Omega_{\mathrm{M}, 0}^2 D(\chi_1) D(\chi_2) }{ a(\chi_1) a(\chi_2) s_1 s_2 } &\left[ s_1 (f(\chi_1) - 1) \mathcal{H}(\chi_1) \mathcal{R}_1 - 5 s_{\mathrm{b}, 1} + 2 \right] \times \nonumber \\ &\left[ s_2 (f(\chi_2) - 1) \mathcal{H}(\chi_2) \mathcal{R}_2 - 5 s_{\mathrm{b}, 2} + 2 \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathcal{R}_1 = \mathcal{R}(s_1)$, ... is computed by func_ℛ_GNC in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathcal{R}(s) = 5 s_{\mathrm{b}}(s) + \frac{2 - 5 s_{\mathrm{b}}(s)}{\mathcal{H}(s) \, s} + \frac{\dot{\mathcal{H}}(s)}{\mathcal{H}(s)^2} - \mathit{f}_{\mathrm{evo}} \quad ;\]

$b_1$, $s_{\mathrm{b}, 1}$, $\mathit{f}_{\mathrm{evo}, 1}$ (and $b_2$, $s_{\mathrm{b}, 2}$, $\mathit{f}_{\mathrm{evo}, 2}$) : galaxy bias, magnification bias (i.e. the slope of the luminosity function at the luminosity threshold), and evolution bias for the first (second) effect;

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

In this TPCF there are no observer terms. The obs keyword is inserted only for compatibility with the other GNC TPCFs.

This function is computed integrating integrand_ξ_GNC_IntegratedGP with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

b1=nothing, s_b1=nothing, 𝑓_evo1=nothing and b2=nothing, s_b2=nothing, 𝑓_evo2=nothing: galaxy, magnification and evolutionary biases respectively for the first and the second effect computed in this TPCF:

if not set (i.e. if you leave the default value nothing) the values stored in the input cosmo will be considered;

if you set one or more values, they will override the cosmo ones in this computation;

the two sets of values should be different only if you are interested in studing two galaxy species;

only the required parameters for the chosen TPCF will be used, depending on its analytical expression; all the others will have no effect, we still inserted them for pragmatical code compatibilities.

s_lim=nothing : parameter used in order to avoid the divergence of the $\mathcal{R}$ and $\mathfrak{R}$ denominators: when $0 \leq s \leq s_\mathrm{lim}$ the returned values are

\[\forall \, s \in [ 0, s_\mathrm{lim} ] \; : \quad \mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}_0 \, s_\mathrm{lim}} \; , \quad \mathcal{R}(s) = 5 s_{\mathrm{b}} + \frac{2 - 5 s_{\mathrm{b}}}{\mathcal{H}_0 \, s_\mathrm{lim}} + \frac{\dot{\mathcal{H}}}{\mathcal{H}_0^2} - \mathit{f}_{\mathrm{evo}} \; .\]

If nothing, the default value stored in cosmo will be considered.

obs::Union{Bool,Symbol} = :noobsvel : do you want to consider the observer terms in the computation of the chosen GNC TPCF effect?

:yes or true -> all the observer effects will be considered

:no or false -> no observer term will be taken into account

:noobsvel -> the observer terms related to the observer velocity (that you can find in the CF concerning Doppler) will be neglected, the other ones will be taken into account

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs_2::Int = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s2) (for χ2); it has been checked that with N_χs_2 ≥ 50 the result is stable.

suit_sampling::Bool = true : this bool keyword can be found in all the TPCFs which have at least one χ integral; it is conceived to enable a sampling of the χ integral(s) suited for the given TPCF; however, it actually have an effect only in the TPCFs that have such a sampling implemented in the code. Currently, only ξ_GNC_Newtonian_Lensing (and its simmetryc TPCF) has it.

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_GNC_multipole, print_map_ξ_GNC_multipole, integrand_ξ_GNC_IntegratedGP

Two-Point Auto-Correlation Functions integrands

GaPSE.integrand_ξ_GNC_Lensing — Function

integrand_ξ_GNC_Lensing(
    IP1::Point, IP2::Point,
    P1::Point, P2::Point,
    y, cosmo::Cosmology;
    Δχ_min::Float64=1e-1, b1=nothing, b2=nothing, 
    s_b1=nothing, s_b2=nothing, 𝑓_evo1=nothing, 𝑓_evo2=nothing,
    s_lim=nothing, obs::Union{Bool,Symbol}=:noobsvel
    ) ::Float64

integrand_ξ_GNC_Lensing(
    χ1::Float64, χ2::Float64,
    s1::Float64, s2::Float64,
    y, cosmo::Cosmology;
    kwargs... )::Float64

Return the integrand of the Two-Point Correlation Function (TPCF) of the Lensing auto-correlation effect arising from the Galaxy Number Counts (GNC).

In the first method, you should pass the two extreme Points (P1 and P2) and the two intermediate integrand Points (IP1 and IP2) where to evaluate the function. In the second method (that internally recalls the first), you must provide the four corresponding comoving distances s1, s2, χ1, χ2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this integrand is the following:

\[\begin{equation} f^{\kappa\kappa} (\chi_1, \chi_2, s_1, s_2, y) = J^{\kappa\kappa}_{\alpha} \left[ J^{\kappa\kappa}_{00} I_0^0(\Delta\chi) + J^{\kappa\kappa}_{02} I_2^0(\Delta\chi) + J^{\kappa\kappa}_{31} I_1^3(\Delta\chi) + J^{\kappa\kappa}_{22} I_2^2(\Delta\chi) \right] \, , \end{equation}\]

with

\[\begin{split} J^{\kappa\kappa}_{\alpha} & = \frac{ \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D(\chi_1) D(\chi_2) }{ s_1 s_2 a(\chi_1) a(\chi_2)} (\chi_1 - s_1)(\chi_2 - s_2) (5 s_{\mathrm{b}, 1} - 2)(5 s_{\mathrm{b}, 2} - 2) \, , \\ %%%%&%%%%%%%%%%%%% J^{\kappa\kappa}_{00} & = -\frac{ 3 \chi_1^2 \chi_2^2}{4 \Delta\chi^4} (y^2 - 1) \left[ 8 y (\chi_1^2 + \chi_2^2) - 9\chi_1\chi_2y^2 - 7\chi_1\chi_2 \right] \, , \\ %%%%&%%%%%%%%%%%%% J^{\kappa\kappa}_{02} & = -\frac{ 3 \chi_1^2 \chi_2^2}{2 \Delta\chi^4}(y^2 - 1) \left[ 4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2 \right] \, , \\ %%%%%%%%%%%%%%%%%% J^{\kappa\kappa}_{31} & = 9 y \Delta\chi^2 \, , \\ %%%%%%%%%%%%%%%%%% J^{\kappa\kappa}_{22} & = \frac{9 \chi_1 \chi_2}{4 \Delta\chi^4} \left[ 2(\chi_1^4 + \chi_2^4)(7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2)(y^2 + 1) + \right.\\ &\left.\qquad\qquad\qquad \chi_1^2 \chi_2^2 (11y^4 + 14y^2 + 23) \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathcal{R}_1 = \mathcal{R}(s_1)$, ... is computed by func_ℛ_GNC in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathcal{R}(s) = 5 s_{\mathrm{b}}(s) + \frac{2 - 5 s_{\mathrm{b}}(s)}{\mathcal{H}(s) \, s} + \frac{\dot{\mathcal{H}}(s)}{\mathcal{H}(s)^2} - \mathit{f}_{\mathrm{evo}} \quad ;\]

$b_1$, $s_{\mathrm{b}, 1}$, $\mathit{f}_{\mathrm{evo}, 1}$ (and $b_2$, $s_{\mathrm{b}, 2}$, $\mathit{f}_{\mathrm{evo}, 2}$) : galaxy bias, magnification bias (i.e. the slope of the luminosity function at the luminosity threshold), and evolution bias for the first (second) effect;

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

In this TPCF there are no observer terms. The obs keyword is inserted only for compatibility with the other GNC TPCFs.

This function is used inside ξ_GNC_Lensing with trapz() from the Trapz Julia package.

Inputs

IP1::Point, IP2::Point, P1::Point, P2::Point or χ1, χ2, s1, s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

b1=nothing, s_b1=nothing, 𝑓_evo1=nothing and b2=nothing, s_b2=nothing, 𝑓_evo2=nothing: galaxy, magnification and evolutionary biases respectively for the first and the second effect computed in this TPCF:

if not set (i.e. if you leave the default value nothing) the values stored in the input cosmo will be considered;

if you set one or more values, they will override the cosmo ones in this computation;

the two sets of values should be different only if you are interested in studing two galaxy species;

only the required parameters for the chosen TPCF will be used, depending on its analytical expression; all the others will have no effect, we still inserted them for pragmatical code compatibilities.

s_lim=nothing : parameter used in order to avoid the divergence of the $\mathcal{R}$ and $\mathfrak{R}$ denominators: when $0 \leq s \leq s_\mathrm{lim}$ the returned values are

\[\forall \, s \in [ 0, s_\mathrm{lim} ] \; : \quad \mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}_0 \, s_\mathrm{lim}} \; , \quad \mathcal{R}(s) = 5 s_{\mathrm{b}} + \frac{2 - 5 s_{\mathrm{b}}}{\mathcal{H}_0 \, s_\mathrm{lim}} + \frac{\dot{\mathcal{H}}}{\mathcal{H}_0^2} - \mathit{f}_{\mathrm{evo}} \; .\]

If nothing, the default value stored in cosmo will be considered.

obs::Union{Bool,Symbol} = :noobsvel : do you want to consider the observer terms in the computation of the chosen GNC TPCF effect?

:yes or true -> all the observer effects will be considered

:no or false -> no observer term will be taken into account

:noobsvel -> the observer terms related to the observer velocity (that you can find in the CF concerning Doppler) will be neglected, the other ones will be taken into account

Δχ_min::Float64 = 1e-4 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[ \lim_{\Delta\chi \to 0^{+}} \left( J_{00}^{\kappa\kappa} \, I^0_0(\Delta\chi) + J_{02}^{\kappa\kappa} \, I^0_2(\Delta\chi) + J_{31}^{\kappa\kappa} \, I^3_1(\Delta\chi) + J_{22}^{\kappa\kappa} \, I^2_2(\Delta\chi) \right) = 3 \, \sigma_2 + \frac{6}{5} \, \chi_2^2 \, \sigma_0\]

So, when it happens that $\Delta\chi < \Delta\chi_\mathrm{min}$, the function considers this limit as the result of the parenthesis instead of calculating it in the normal way; it prevents computational divergences.

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_GNC_multipole, print_map_ξ_GNC_multipole, ξ_GNC_Lensing

GaPSE.integrand_ξ_GNC_IntegratedGP — Function

integrand_ξ_GNC_IntegratedGP(
    IP1::Point, IP2::Point,
    P1::Point, P2::Point,
    y, cosmo::Cosmology; 
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool,Symbol}=:noobsvel
    ) ::Float64

integrand_ξ_GNC_IntegratedGP(
    χ1::Float64, χ2::Float64,
    s1::Float64, s2::Float64,
    y, cosmo::Cosmology;
    kwargs...) ::Float64

Return the integrand of the Two-Point Correlation Function (TPCF) of the Integrated Gravitational Potential (GP) auto-correlation effect arising from the Galaxy Number Counts (GNC).

In the first method, you should pass the two extreme Points (P1 and P2) and the two intermediate integrand Points (IP1 and IP2) where to evaluate the function. In the second method (that internally recalls the first), you must provide the four corresponding comoving distances s1, s2, χ1, χ2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this integrand is the following:

\[\begin{equation} f^{\int \!\phi\int \!\phi} (\chi_1, \chi_2, s_1, s_2, y) = J^{\int \!\phi \int \!\phi}_{40} \tilde{I}_0^4 ( \Delta\chi) \, , \end{equation}\]

with

\[\begin{split} J^{\int \!\phi\int \!\phi}_{40} = \frac{ 9 \Delta\chi ^4 \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D(\chi_1) D(\chi_2) }{ a(\chi_1) a(\chi_2) s_1 s_2 } &\left[ s_1 (f(\chi_1) - 1) \mathcal{H}(\chi_1) \mathcal{R}_1 - 5 s_{\mathrm{b}, 1} + 2 \right] \times \nonumber \\ &\left[ s_2 (f(\chi_2) - 1) \mathcal{H}(\chi_2) \mathcal{R}_2 - 5 s_{\mathrm{b}, 2} + 2 \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathcal{R}_1 = \mathcal{R}(s_1)$, ... is computed by func_ℛ_GNC in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathcal{R}(s) = 5 s_{\mathrm{b}}(s) + \frac{2 - 5 s_{\mathrm{b}}(s)}{\mathcal{H}(s) \, s} + \frac{\dot{\mathcal{H}}(s)}{\mathcal{H}(s)^2} - \mathit{f}_{\mathrm{evo}} \quad ;\]

$b_1$, $s_{\mathrm{b}, 1}$, $\mathit{f}_{\mathrm{evo}, 1}$ (and $b_2$, $s_{\mathrm{b}, 2}$, $\mathit{f}_{\mathrm{evo}, 2}$) : galaxy bias, magnification bias (i.e. the slope of the luminosity function at the luminosity threshold), and evolution bias for the first (second) effect;

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

In this TPCF there are no observer terms. The obs keyword is inserted only for compatibility with the other GNC TPCFs.

This function is used inside ξ_GNC_IntegratedGP with trapz() from the Trapz Julia package.

Inputs

IP1::Point, IP2::Point, P1::Point, P2::Point or χ1, χ2, s1, s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

b1=nothing, s_b1=nothing, 𝑓_evo1=nothing and b2=nothing, s_b2=nothing, 𝑓_evo2=nothing: galaxy, magnification and evolutionary biases respectively for the first and the second effect computed in this TPCF:

if not set (i.e. if you leave the default value nothing) the values stored in the input cosmo will be considered;

if you set one or more values, they will override the cosmo ones in this computation;

the two sets of values should be different only if you are interested in studing two galaxy species;

only the required parameters for the chosen TPCF will be used, depending on its analytical expression; all the others will have no effect, we still inserted them for pragmatical code compatibilities.

s_lim=nothing : parameter used in order to avoid the divergence of the $\mathcal{R}$ and $\mathfrak{R}$ denominators: when $0 \leq s \leq s_\mathrm{lim}$ the returned values are

\[\forall \, s \in [ 0, s_\mathrm{lim} ] \; : \quad \mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}_0 \, s_\mathrm{lim}} \; , \quad \mathcal{R}(s) = 5 s_{\mathrm{b}} + \frac{2 - 5 s_{\mathrm{b}}}{\mathcal{H}_0 \, s_\mathrm{lim}} + \frac{\dot{\mathcal{H}}}{\mathcal{H}_0^2} - \mathit{f}_{\mathrm{evo}} \; .\]

If nothing, the default value stored in cosmo will be considered.

obs::Union{Bool,Symbol} = :noobsvel : do you want to consider the observer terms in the computation of the chosen GNC TPCF effect?

:yes or true -> all the observer effects will be considered

:no or false -> no observer term will be taken into account

:noobsvel -> the observer terms related to the observer velocity (that you can find in the CF concerning Doppler) will be neglected, the other ones will be taken into account

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_GNC_multipole, print_map_ξ_GNC_multipole, ξ_GNC_IntegratedGP