Tool functions

GaPSE.InputPS — Type

 InputPS(
      l_si::Float64
      l_b::Float64
      l_a::Float64
      left::Float64

      spline::Dierckx.Spline1D
      
      r_si::Float64
      r_b::Float64
      r_a::Float64
      right::Float64)

Store the Input Power Spectrum.

Arguments

l_si, l_b, l_a :: Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the LEFT edge; when an input value x < left is given, the returned one is obtained from power_law with this coefficients ( where, of course, l_si is the exponent, l_b the coefficient and l_a the spurious adding constant).

left::Float64 : the break between the left power-law (for x <left) and the spline (for x ≥ left); its value is the fit_min of the used constructor.

spline::Dierckx.Spline1D : spline that interpolates between the real values of the power spectrum inside the range left ≤ x ≤ right

right::Float64 : the break between the right power-law (for x > left) and the spline (for x ≤ right); its value is the fit_max of the used constructor.

r_si, r_b, r_a :: Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the RIGHT edge; when an input value x > right is given, the returned one is obtained from power_law with this coefficients ( where, of course, r_si is the exponent, r_bthe coefficient andrathe spurious adding constant). NOTE: for numerical issues, only the "pure" power-lawy = f(x) = b + x^scan be used. In other words, it always setra = 0.0`.

Constructors

InputPS(file::String; fit_left_min = 1e-6, fit_left_max = 3e-6, fit_right_min = 1e1, fit_right_max = 2e1) : read the IPS from the given input file; it can contain comments (defined with a starting # on each line), but the file structure must be space-separated in two colums (former for k values, latter for P ones).

fit_left_min = 1e-6, fit_left_max = 3e-6 : the limits (min and max) where the PS must be fitted with a power law, for small wavenumbers.

fit_right_min = 1e1, fit_right_max = 2e1 : the limits (min and max) where the PS must be fitted with a power law, for high wavenumbers.

InputPS(ks::AbstractVector{T1}, pks::AbstractVector{T2}; fit_left_min = 1e-6, fit_left_max = 3e-6, fit_right_min = 1e1, fit_right_max = 2e1)

ks::AbstractVector{T1}, pks::AbstractVector{T2} : self-explanatory ks and pks array-like values.

fit_left_min = 1e-6, fit_left_max = 3e-6 : the limits (min and max) where the PS must be fitted with a power law, for small wavenumbers.

fit_right_min = 1e1, fit_right_max = 2e1 : the limits (min and max) where the PS must be fitted with a power law, for high wavenumbers.

All the power-law fitting (both "pure" and spurious) are made through the local function power_law_from_data.

See also: power_law_from_data

GaPSE.IntegralIPS — Type

 IntegralIPS(
      l_si::Float64
      l_b::Float64
      l_a::Float64
      left::Float64

      spline::Dierckx.Spline1D

      r_si::Float64
      r_b::Float64
      r_a::Float64
      right::Float64
 )

Contains all the information useful in order to return the value of an integral obtained from the Input Power Spectrum.

Arguments

l_si, l_b, l_a :: Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the LEFT edge; when an input value x < left is given, the returned one is obtained from power_law with this coefficients ( where, of course, l_si is the exponent, l_b the coefficient and l_a the spurious adding constant).

left::Float64 : the break between the left power-law (for x <left) and the spline (for x ≥ left); its value is the fit_min of the used constructor.

spline::Dierckx.Spline1D : spline that interpolates between the real values of the integral calculated inside the range left ≤ x ≤ right

right::Float64 : the break between the right power-law (for x > left) and the spline (for x ≤ right); its value is the fit_max of the used constructor.

r_si, r_b, r_a :: Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the RIGHT edge; when an input value x > right is given, the returned one is obtained from power_law with this coefficients ( where, of course, r_si is the exponent, r_bthe coefficient andrathe spurious adding constant). NOTE: for numerical issues, only the "pure" power-lawy = f(x) = b + x^scan be used. In other words, it always setra = 0.0`.

Constructors

There are two type of integrals we are interested in, and so two constructors are here provided:

IntegralIPS(ips, l, n; N::Integer = 1024, kmin = 1e-4, kmax = 1e3, s0 = 1e-3, fit_left_min = 2.0, fit_left_max = 10.0, p0_left = nothing, con = false, fit_right_min = nothing, fit_right_max = nothing, p0_right = nothing) This is the one used for the "classical" $I_\ell_n$ integrals:

\[I_\ell^n(s) = \int_0^\infty \frac{\mathrm{d} q}{2 \pi^2} q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]

where, for a generic Iab name, $\ell$ is the FIRST number (a) and $n$ the second (b). The integral obtained with this constructor is calculated through xicalc, and expanded with power-laws at the edges.

ips: the function/spline that gives the Input Power Spectrum

l and n: self-explanatory degree of the integral, with the convenction above mentioned

kmin = 1e-4, kmax = 1e3, s0 = 1e-3 : values to be passed to xicalc for the integration

fit_left_min = 2.0, fit_left_max = 10.0, : the limits (min and max) where the integral $I_\ell^n$ must be fitted with a power law, for small distances. This operation is necessary, because xicalc, in this context, gives wrong results for too small input distance s; nevertheless, all these $I_\ell^n$ integrals have fixed power-law trends for $s \rightarrow 0$, so this approach gives good results.

p0_left = nothing : vector with the initial values for the left power-law fitting; its length must be 2 (if you want to fit with a pure power-law $y = f(x) = b * x^s$, so only l_si and l_b are matter of concern) or 3 (if you want to fit with a spurious power-law $y = f(x) = a + b * x^s$, so you are also interested in l_a), depending on the value of con; if nothing, it will be automatically set p0 = [-1.0, 1.0, 0.0] for con==true and p0 = [-1.0, 1.0] for con==false.

con::Bool = false : do you want that the fit of all the $I_\ell^n$ for the LEFT edge is not a simple power-law $y = f(x) = b \, x^s$, but also consider a constant $a$, such that $y = f(x) = a + b \, x^s$? For the LEFT side, there is not a lot of difference empirically. For the RIGHT side, there is not such an option due to numerical problems (it's like is always set con==false).

fit_right_min = nothing, fit_right_max = nothing : the limits (min and max) where the integral $I_\ell^n$ must be fitted with a power law, for high distances. These $I_\ell^n$ integrals have fixed power-law trends for $s \rightarrow infty$, so this approach gives good results. If nothing, the last 15 points returned from xicalc are used for this fitting. NOTE: for numerical issues, only the "pure" power-law $y = f(x) = b + x^s$ can be used.

p0_right = nothing : vector with the initial values for the left power-law fitting; its length must be 2 (to fit with a pure power-law $y = f(x) = b * x^s$, so only r_si and r_b are matter of concern); if nothing, it will be automatically set p0 = [-4.0, 1.0, 0.0].

IntegralIPS(ips, func::Function; N::Integer = 1024, kmin = 1e-4, kmax = 1e3, fit_left_min = 0.1, fit_left_max = 1.0, p0_left = nothing, con = false, fit_right_min = nothing, fit_right_max = nothing, p0_right = nothing, kwargs...) This is the one used for the "strange" $\tilde{I}$ integrals, such as:

\[\tilde{I}^4_0 (s) = \int \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4} \;.\]

The integral obtained with this constructor is calculated through the input function func, and expanded with power-laws at the edges. For \tilde{I}^4_0, the function is func_I04_tilde.

ips: the function/spline that gives the Input Power Spectrum

func: function that return the value of this specific integral in a given value

kmin = 1e-4, kmax = 1e3, s0 = 1e-3 : values to be passed to func as extremes of integration

fit_left_min = 0.1, fit_left_max = 1.0, : the limits (min and max) where the integral $\tilde{I}$ must be fitted with a power law, for small distances. This operation is necessary, because xicalc, in this context, gives wrong results for too small input distance s; nevertheless, all this $\tilde{I}$ integral have fixed power-law trends for $s \rightarrow 0$, so this approach gives good results.

p0_left = nothing : vector with the initial values for the left power-law fitting; its length must be 2 (if you want to fit with a pure power-law $y = f(x) = b * x^s$, so only l_si and l_b are matter of concern) or 3 (if you want to fit with a spurious power-law $y = f(x) = a + b * x^s$, so you are also interested in l_a), depending on the value of con; if nothing, it will be automatically set p0 = [-2.0, -1.0, 0.0] for con==true and p0 = [-2.0, -1.0] for con==false.

con::Bool = false : do you want that the fit of all the $I_\ell^n$ for the LEFT edge is not a simple power-law $y = f(x) = b \, x^s$, but also consider a constant $a$, such that $y = f(x) = a + b \, x^s$? For the LEFT side, there is not a lot of difference empirically. For the RIGHT side, there is not such an option due to numerical problems (it's like is always set con==false).

fit_right_min = nothing, fit_right_max = nothing : the limits (min and max) where the integral $I_\ell^n$ must be fitted with a power law, for high distances. These $I_\ell^n$ integrals have fixed power-law trends for $s \rightarrow infty$, so this approach gives good results. If nothing, the last 15 points returned from xicalc are used for this fitting. NOTE: for numerical issues, only the "pure" power-law $y = f(x) = b + x^s$ can be used.

p0_right = nothing : vector with the initial values for the left power-law fitting; its length must be 2 (to fit with a pure power-law $y = f(x) = b * x^s$, so only r_si and r_b are matter of concern); if nothing, it will be automatically set p0 = [-4.0, -1.0].

All the power-law fitting (both "pure" and spurious) are made through the local function power_law_from_data.

See also: power_law_from_data, func_I04_tilde

GaPSE.IPSTools — Type

 IPSTools(
      I00::IntegralIPS
      I20::IntegralIPS
      I40::IntegralIPS
      I02::IntegralIPS
      I22::IntegralIPS
      I31::IntegralIPS
      I13::IntegralIPS
      I11::IntegralIPS

      I04_tilde::IntegralIPS

      σ_0::Float64
      σ_1::Float64
      σ_2::Float64
      σ_3::Float64

      fit_min::Union{Float64,Nothing}
      fit_max::Union{Float64,Nothing}
      k_min::Float64
      k_max::Float64
      s_0::Float64
      )

Struct that contains all the useful functions and values obtained from the Input Power Spectrum.

Arguments

I00, I20, I40, I02, I22, I31, I13, I11 ::IntegralIPS: they return the value of the corresponding integral:

\[I_\ell^n(s) = \int_0^\infty \frac{\mathrm{d} q}{2 \pi^2} q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]

where, for a generic Iab name, $\ell$ is the FIRST number (a) and $n$ the second (b). These integrals are performed through xicalc, with kmin, kmax, s0 = 1e-5, 1e3, 1e-3; at the edges they are fitted with power laws (for s < fit_min and s > max_s_returned_from_xi_calc).

I04_tilde::IntegralIPS: it returns the value of the integral:

\[\tilde{I}^4_0 (s) = \int \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4} \;.\]

This integral is calculated brute-force with quadgk, and fitted with power-laws at the edges (for s < 0.1 and s > 1e4).

σ_0, σ_1, σ_2, σ_3 :: Float64: these are the results of the following integral:

\[\sigma_i = \int_{k_\mathrm{min}}^{k_\mathrm{max}} \frac{\mathrm{d} q}{2 \pi^2} \, q^{2-i} \, P(q)\]

fit_min, fit_max :: Float64: the limits (min and max) where the integral $I_\ell^n$ must be fitted with a power law, for small distances. This operation is necessary, because xicalc, in this context, gives wrong results for too small input distance s; nevertheless, all these $I_\ell^n$ integrals have fixed power-law trends for $s \rightarrow 0$, so this approach gives good results.

k_min k_max::Float64 : because some of the $\sigma_i$ integrals from $q = 0$ to $q = +\infty$ diverge, it is common practice to cut the integrals at the edges, so they are calculated from $q = k_\mathrm{min}$ to $q = k_\mathrm{max}$

Constructors

IPSTools(ips::InputPS; N::Integer = 1024, fit_min::Float64 = 0.05, fit_max::Float64 = 0.5, k_min::Float64 = 1e-6, k_max::Float64 = 10.0 con::Bool = false )

ips::InputPS : the Input Power Spectrum to be used in all the calculations.

N::Integer = 1024 : number of points to be used in the xicalc function

k_min::Float64 = 1e-6, k_max::Float64 = 10.0 : integrations extremes of the $\sigma_i$s

con::Bool = false : do you want that the fit of all the $I_\ell^n$ for the LEFT edge is not a simple power-law $y = f(x) = b \, x^s$, but also consider a constant $a$, such that $y = f(x) = a + b \, x^s$? For the LEFT side, there is not a lot of difference empirically. For the RIGHT side, there is not such an option due to numerical problems (it's like is always set con==false).

See also: IntegralIPS, InputPS