Kernels

GaPSE.lr — Function

lr(a, b, n, i) ::Float64

Return the ith-number in the linear range [a,b] subdivided in n pieces:

\[\mathrm{lr}(a,b,n,i) = a + \frac{i-1}{n-1}\,(b-a)\]

It means that, with n>2, 1 ≤ i ≤ n and:

i=1 => lr(a, b, n, 1) = a;

1<i<n => a < lr(a, b, n, i) < b;

i=n => lr(a, b, n, 1) = b.

It's used inside kernel_1d_P1!, kernel_1d_P2! and kernel_2d!

GaPSE.kernel_1d_P1! — Function

@kernel function kernel_1d_P1!(results_vector, integrand, P1, P2, y, cosmo, N_χs, kwargs...) ::Nothing

One-dimensional (1d) kernel used for the parallelisation of the 1d integrals over χ1 (P1). The inputs are:

results_vector : 1d vector where the results will be stored;

integrand : function to be integrated (e.g. GaPSE.integrand_Lensing_Doppler);

P1 and P2, y and cosmo : Points, angle cosine and Cosmology to be given to the integrand;

N_χs : number of points to be used sampling the interval 1e-6 ≤ χ1 ≤ P1.comdist;

kwargs... : other kwargs to be passed to integrand.

Example

s1, s2, N_χs = cosmo.s_eff, 1.3*cosmo.s_eff, 50

χ1s = s1 .* range(1e-6, 1, length=N_χs)
P1, P2 = GaPSE.Point(s1, cosmo), GaPSE.Point(s2, cosmo)

int_ξs = KernelAbstractions.zeros(backend, Float64, N_χs)
kernel! = kernel_1d_P1!(backend)
kernel!(int_ξs, GaPSE.integrand_ξ_GNC_Lensing_Doppler, P1, P2, y, cosmo, N_χs, kwargs...; ndrange=size(int_ξs))
KernelAbstractions.synchronize(backend)

res = trapz(χ1s, int_ξs)

See also: kernel_1d_P2!, kernel_2d!

GaPSE.kernel_1d_P2! — Function

@kernel function kernel_1d_P2!(results_vector, integrand, P1, P2, y, cosmo, N_χs, kwargs...) ::Nothing

One-dimensional (1d) kernel used for the parallelisation of the 1d integrals over χ2 (P2). The inputs are:

results_vector : 1d vector where the results will be stored;

integrand : function to be integrated (e.g. GaPSE.integrand_ξ_GNC_Newtonian_IntegratedGP);

P1 and P2, y and cosmo : Points, angle cosine and Cosmology to be given to the integrand;

N_χs : number of points to be used sampling the interval 1e-6 ≤ χ2 ≤ P2.comdist;

kwargs... : other kwargs to be passed to integrand.

Example

s1, s2, N_χs = cosmo.s_eff, 1.3*cosmo.s_eff, 50

χ2s = s2 .* range(1e-6, 1, length=N_χs)
P1, P2 = GaPSE.Point(s1, cosmo), GaPSE.Point(s2, cosmo)

int_ξs = KernelAbstractions.zeros(backend, Float64, N_χs)
kernel! = kernel_1d_P2!(backend)
kernel!(int_ξs, GaPSE.integrand_ξ_GNC_Newtonian_IntegratedGP, P1, P2, y, cosmo, N_χs, kwargs...; ndrange=size(int_ξs))
KernelAbstractions.synchronize(backend)

res = trapz(χ2s, int_ξs)

See also: kernel_1d_P1!, kernel_2d!

GaPSE.kernel_2d! — Function

@kernel function kernel_2d!(results_vector, integrand, P1, P2, y, cosmo, N_χs_2, kwargs...) ::Nothing

Two-dimensional (2d) kernel used for the parallelisation of the 2d integrals over χ1 and χ2. The inputs are:

results_vector : 2d vector where the results will be stored;

integrand : function to be integrated (e.g. GaPSE.integrand_ξ_GNC_AutoLensing);

P1 and P2, y and cosmo : Points, angle cosine and Cosmology to be given to the integrand;

N_χs_2 : number of points to be used sampling each of the intervals 1e-6 ≤ χ1 ≤ P1.comdist and 1e-6 ≤ χ2 ≤ P2.comdist;

kwargs... : other kwargs to be passed to integrand.

Example

s1, s2, N_χs_2 = cosmo.s_eff, 1.3*cosmo.s_eff, 50

χ1s = s1 .* range(1e-6, 1, length=N_χs_2)
χ2s = s2 .* range(1e-6, 1, length=N_χs_2)
P1, P2 = GaPSE.Point(s1, cosmo), GaPSE.Point(s2, cosmo)

int_ξs = KernelAbstractions.zeros(backend, Float64, N_χs_2, N_χs_2)
kernel! = kernel_2d!(backend)
kernel!(int_ξs, GaPSE.integrand_ξ_GNC_Lensing, P1, P2, y, cosmo, N_χs_2, kwargs...; ndrange=size(int_ξs))
KernelAbstractions.synchronize(backend)

res = trapz((χ1s, χ2s), reshape(int_ξs, N_χs_2, N_χs_2))

See also: kernel_1d_P1!, kernel_1d_P2!