LD TPCFs

Two-Point Auto-Correlation Functions

GaPSE.ξ_LD_Doppler — Function

ξ_LD_Doppler(P1::Point, P2::Point, y, cosmo::Cosmology) ::Float64

ξ_LD_Doppler(s1, s2, y, cosmo::Cosmology) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Doppler auto-correlation effect arising from the Luminosity Distance (LD) perturbations.

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this term is the following:

\[\begin{split} \xi^{v_{\parallel}v_{\parallel}} (s_1, s_2, y) = \mathcal{J}^{v_{\parallel}v_{\parallel}}_{\alpha} \left[ \mathcal{J}^{v_{\parallel}v_{\parallel}}_{00} I_0^0(s) + \mathcal{J}^{v_{\parallel}v_{\parallel}}_{02} I_2^0(s) + \mathcal{J}^{v_{\parallel}v_{\parallel}}_{04} I_4^0(s) + \mathcal{J}^{v_{\parallel}v_{\parallel}}_{20} I_0^2(s) \right] \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{v_{\parallel}v_{\parallel}}_{\alpha} & = D_1 D_2 f_1 f_2 \mathcal{H}_1 \mathcal{H}_2 \mathfrak{R}_1 \mathfrak{R}_2 \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{v_{\parallel}v_{\parallel}}_{00} & = \frac{1}{45} \left[ y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2 \right] \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{v_{\parallel}v_{\parallel}}_{02} & = \frac{2}{63} \left[ y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2 \right] \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{v_{\parallel}v_{\parallel}}_{04} & = \frac{1}{105} \left[ y^2 s_1 s_2 - 2y(s_1^2 + s_2^2) + 3s_1 s_2 \right] \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{v_{\parallel}v_{\parallel}}_{20} & = \frac{1}{3} y s^2 \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_Lensing — Function

ξ_LD_Lensing(P1::Point, P2::Point, y, cosmo::Cosmology;
    en::Float64 = 1e6, Δχ_min::Float64 = 1e-3,
    N_χs_2::Int = 100) ::Float64

ξ_LD_Lensing(s1, s2, y, cosmo::Cosmology; kwargs...) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Lensing auto-correlation effect arising from the Luminosity Distance (LD) perturbations.

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this term is the following:

\[\begin{split} \xi^{\kappa\kappa} (s_1, s_2, y) = \int_0^{s_1} \mathrm{d}\chi_1 \int_0^{s_2} \mathrm{d}\chi_2 \; \mathcal{J}^{\kappa\kappa}_{\alpha} \left[ \mathcal{J}^{\kappa\kappa}_{00}I^0_0(\Delta\chi) + \mathcal{J}^{\kappa\kappa}_{02} I^0_2(\Delta\chi) + \mathcal{J}^{\kappa\kappa}_{31}I^3_1(\Delta\chi) + \mathcal{J}^{\kappa\kappa}_{22}I^2_2(\Delta\chi) \right] \nonumber \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{\kappa\kappa}_{\alpha} & = \frac{ \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D(\chi_1) D(\chi_2) }{ s_1 s_2 a(\chi_1) a(\chi_2) }(\chi_1 - s_1)(\chi_2 - s_2) \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa\kappa}_{00} & = -\frac{ 3 \chi_1^2 \chi_2^2}{4 \Delta\chi^4} (y^2 - 1) \left[ 8 y (\chi_1^2 + \chi_2^2) - 9\chi_1\chi_2y^2 - 7\chi_1\chi_2 \right] \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa\kappa}_{02} & = -\frac{ 3 \chi_1^2 \chi_2^2}{2 \Delta\chi^4}(y^2 - 1) \left[ 4 y (\chi_1^2 + \chi_2^2) - 3 \chi_1 \chi_2 y^2 - 5 \chi_1 \chi_2 \right] \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa\kappa}_{31} & = 9 y \Delta\chi^2 \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa\kappa}_{22} & = \frac{9 \chi_1 \chi_2}{4 \Delta\chi^4} \left[ 2(\chi_1^4 + \chi_2^4)(7 y^2 - 3) - 16 y \chi_1 \chi_2 (\chi_1^2 + \chi_2^2)(y^2 + 1) + \right.\\ &\left.\qquad\qquad\qquad \chi_1^2 \chi_2^2 (11y^4 + 14y^2 + 23) \right] \nonumber \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

This function is computed integrating integrand_ξ_LD_Lensing with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

Keyword arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs_2::Int = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s2) (for χ2); it has been checked that with N_χs_2 ≥ 50 the result is stable.

Δχ_min::Float64 = 1e-4 : when $\Delta\chi = \sqrt{\chi_1^2 + \chi_2^2 - 2 \, \chi_1 \chi_2 y} \to 0^{+}$, some $I_\ell^n$ term diverges, but the overall parenthesis has a known limit:

\[\lim_{\Delta\chi\to 0^{+}} \left(J^{\kappa\kappa}_{00} \, I^0_0(\Delta\chi) + J^{\kappa\kappa}_{02} \, I^0_2(\Delta\chi) + J^{\kappa\kappa}_{31} \, I^3_1(\Delta\chi) + J^{\kappa\kappa}_{22} \, I^2_2(\Delta\chi) \right) = \frac{4}{15} \, \left(5 \, \sigma_2 + \frac{2}{3} \, σ_0 \,s_1^2 \, \chi_2^2\right)\]

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_LocalGP — Function

ξ_LD_LocalGP(P1::Point, P2::Point, y, cosmo::Cosmology) ::Float64

ξ_LD_LocalGP(s1, s2, y, cosmo::Cosmology; kwargs...) ::Float64

Return the Two-Point Correlation Function (TPCF) of the Local Gravitational Potential (GP) auto-correlation effect arising from the Luminosity Distance (LD) perturbations.

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this term is the following:

\[\begin{split} \xi^{\phi\phi} (s_1, s_2, y) = \frac{9 \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D_1 D_2 s^4 }{4 a_1 a_2} \left(1 + \mathfrak{R}_1 + \mathfrak{R}_2 + \mathfrak{R}_1 \mathfrak{R}_2 \right) \tilde{I}^4_0(s) \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_IntegratedGP — Function

ξ_LD_IntegratedGP(P1::Point, P2::Point, y, cosmo::Cosmology; 
    en::Float64 = 1e10, N_χs_2::Int = 100) :: Float64

ξ_LD_IntegratedGP(s1, s2, y, cosmo::Cosmology; kwargs... )::Float64

Return the Two-Point Correlation Function (TPCF) of the Integrated Gravitational Potential (GP) auto-correlation effect arising from the Luminosity Distance (LD) perturbations.

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this term is the following:

\[\begin{split} \xi^{\int\!\phi\int\!\phi} (s_1, s_2, y) = \int_0^{s_1}\mathrm{d}\chi_1\int_0^{s_2}\mathrm{d}\chi_2 \; \mathcal{J}^{\int\!\phi\int\!\phi}_{40} \tilde{I}^4_0(\chi) \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{\int\!\phi\int\!\phi}_{40} = \frac{ 9 \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D(\chi_1) D(\chi_2) \Delta\chi^4 }{ a(\chi_1)\, a(\chi_2)\, s_1\, s_2 } &\left[ s_2 \mathcal{H}(\chi_2) \mathfrak{R}_2(f(\chi_2) - 1) - 1 \right] \times \\ &\left[ s_1 \mathcal{H}(\chi_1) \mathfrak{R}_1(f(\chi_1) - 1) - 1 \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

This function is computed integrating integrand_ξ_LD_IntegratedGP with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

Keyword Arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs_2::Int = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s2) (for χ2); it has been checked that with N_χs_2 ≥ 50 the result is stable.

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

Two-Point Cross-Correlation Functions

GaPSE.ξ_LD_Doppler_Lensing — Function

ξ_LD_Doppler_Lensing(
    s1, s2, y, cosmo::Cosmology; 
    kwargs... ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler and the Lensing effects arising from the Luminosity Distance (LD) perturbations.

It's computed through the symmetric function ξ_LD_Lensing_Doppler; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole, ξ_LD_Lensing_Doppler

GaPSE.ξ_LD_Lensing_Doppler — Function

ξ_LD_Lensing_Doppler(
    s1, s2, y, cosmo::Cosmology;
    en::Float64 = 1e6, N_χs::Int = 100 ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing and the Doppler effects arising from the Luminosity Distance (LD) perturbations.

You must provide the two comoving distances s1 and s2 where to evaluate the function. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this integrand is the following:

\[\begin{split} \xi^{\kappa v_{\parallel}} (s_1, s_2, y) = \int_0^{s_1} &\mathrm{d}\chi_1 \; \mathcal{J}^{\kappa v_{\parallel}}_{\alpha} \times \\ &\left[ \mathcal{J}^{\kappa v_{\parallel}}_{00} I_0^0(\Delta\chi_1) + \mathcal{J}^{\kappa v_{\parallel}}_{02} I_2^0(\Delta\chi_1) + \mathcal{J}^{\kappa v_{\parallel}}_{04} I_4^0(\Delta\chi_1) + \mathcal{J}^{\kappa v_{\parallel}}_{20} I_0^2(\Delta\chi_1) \right] \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{\kappa v_{\parallel}}_{\alpha} &= \mathcal{H}_0^2 \Omega_{\mathrm{M}0} D_2 f_2 \mathcal{H}_2 \mathfrak{R}_2 \frac{D(\chi_1) (\chi_1 - s_1)}{a(\chi_1) s_1} \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa v_{\parallel}}_{00} & = \frac{1}{15} \left[ \chi_1^2 y + \chi_1(4 y^2 - 3) s_2 - 2 y s_2^2 \right] \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa v_{\parallel}}_{02} & = \frac{1}{42 \Delta\chi_1^2} \left[ 4 \chi_1^4 y + 4 \chi_1^3 (2 y^2 - 3) s_2 + \chi_1^2 y (11 - 23 y^2) s_2^2 + \right.\\ &\left.\qquad\qquad\qquad \chi_1 (23 y^2 - 3) s_2^3 - 8 y s_2^4 \right] \nonumber \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa v_{\parallel}}_{04} & = \frac{1}{70 \Delta\chi_1^2} \left[ 2\chi_1^4 y + 2 \chi_1^3 (2 y^2 - 3) s_2 - \chi_1^2 y (y^2 + 5) s_2^2 + \right.\\ &\left.\qquad\qquad\qquad \chi_1(y^2 + 9) s_2^3 - 4 y s_2^4 \right] \nonumber \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa v_{\parallel}}_{20} & = y \Delta\chi_1^2 \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

This function is computed integrating integrand_ξ_LD_Lensing_Doppler with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

Keyword Arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs::Int = 100: number of points to be used for sampling the integral along the range (0, s1) (for χ1); it has been checked that with N_χs ≥ 50 the result is stable.

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_Doppler_LocalGP — Function

ξ_LD_Doppler_LocalGP(P1::Point, P2::Point, y, cosmo::Cosmology ) ::Float64
ξ_LD_Doppler_LocalGP(s1, s2, y, cosmo::Cosmology ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler and the Local Gravitational Potential (GP) effects arising from the Luminosity Distance (LD) perturbations.

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{v_{\parallel}\phi} (s_1, s_2, y) = \mathcal{J}^{v_{\parallel}\phi}_{31} I^3_1(s) \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{v_{\parallel}\phi}_{31} = \frac{3}{2 a_2} \mathcal{H}_1 f_1 D_1 \mathfrak{R}_1 \mathcal{H}_0^2 \Omega_{\mathrm{M}0} D_2 (1 + \mathfrak{R}_2)(s_1 - s_2 y) \, s^2 \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_LocalGP_Doppler — Function

ξ_LD_LocalGP_Doppler(
    s1, s2, y, cosmo::Cosmology; 
    kwargs... ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) and the Doppler effects arising from the Luminosity Distance (LD) perturbations.

It's computed through the symmetric function ξ_LD_Doppler_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole, ξ_LD_Doppler_LocalGP

GaPSE.ξ_LD_Doppler_IntegratedGP — Function

ξ_LD_Doppler_IntegratedGP(
    s1, s2, y, cosmo::Cosmology;
    en::Float64 = 1e6, N_χs::Int = 100 ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler and the Integrated Gravitational Potential (GP) effects arising from the Luminosity Distance (LD) perturbations.

You must provide the two comoving distances s1 and s2 where to evaluate the function. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{v_{\parallel} \int\!\phi} \left(\chi_2, s_1 , s_2, y \right) &= \int_0^{s_2} \mathrm{d}\chi_2 \; \mathcal{J}^{v_{\parallel} \int\!\phi}_{31} \,I_1^3(\Delta\chi_2) \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{v_{\parallel} \int\!\phi}_{31} = 3 \mathcal{H}_1 f_1 D_1 \mathcal{H}_0^2 \Omega_{\mathrm{M}0} \mathfrak{R}_1 \frac{ D(\chi_2)(s_1 - \chi_2 y) }{ a(\chi_2) s_2 } \Delta\chi_2^2 \left[ s_2 \mathfrak{R}_2 \mathcal{H}(\chi_2)(f_2 -1) - 1 \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

This function is computed integrating integrand_ξ_LD_Doppler_IntegratedGP with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

Keyword Arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs::Int = 100: number of points to be used for sampling the integral along the range (0, s2) (for χ2); it has been checked that with N_χs ≥ 100 the result is stable.

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_IntegratedGP_Doppler — Function

ξ_LD_IntegratedGP_Doppler(
    s1, s2, y, cosmo::Cosmology; 
    kwargs... ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) and the Doppler effects arising from the Luminosity Distance (LD) perturbations.

It's computed through the symmetric function ξ_LD_Doppler_IntegratedGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole, ξ_LD_Doppler_IntegratedGP

GaPSE.ξ_LD_Lensing_LocalGP — Function

ξ_LD_Lensing_LocalGP(
    s1, s2, y, cosmo::Cosmology;
    en::Float64 = 1e6, N_χs::Int = 100 ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing and the Local Gravitational Potential (GP) effects arising from the Luminosity Distance (LD) perturbations.

You must provide the two comoving distances s1 and s2 where to evaluate the function. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{\kappa \phi} (s_1, s_2, y) = \int_0^{s_1} \mathrm{d}\chi_1 \, \mathcal{J}^{\kappa \phi}_{\alpha} \left[ \mathcal{J}^{\kappa \phi}_{31} I_1^3(\Delta\chi_1) + \mathcal{J}^{\kappa \phi}_{22} I_2^2(\Delta\chi_1) \right] \nonumber \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{\kappa \phi}_{\alpha} &= \frac{ 9 \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D_2 s_2 }{4 a_2 s_1} (1 + \mathfrak{R}_2) \frac{D(\chi_1)(s_1 - \chi_1)}{a(\chi_1)} \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa \phi}_{31} & = -2 y \Delta\chi_1^2 \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa \phi}_{22} & = \chi_1 s_2 (1 - y^2) \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

This function is computed integrating integrand_ξ_LD_Lensing_LocalGP with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

Keyword Arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs::Int = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1); it has been checked that with N_χs ≥ 50 the result is stable.

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_LocalGP_Lensing — Function

ξ_LD_LocalGP_Lensing(
    s1, s2, y, cosmo::Cosmology; 
    kwargs... ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) and the Lensing effects arising from the Luminosity Distance (LD) perturbations.

It's computed through the symmetric function ξ_LD_Lensing_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole, ξ_LD_Lensing_LocalGP

GaPSE.ξ_LD_Lensing_IntegratedGP — Function

ξ_LD_Lensing_IntegratedGP(
    P1::Point, P2::Point, y, cosmo::Cosmology;
    en::Float64 = 1e6, N_χs_2::Int = 100 ) ::Float64

ξ_LD_Lensing_IntegratedGP(
    s1, s2, y, cosmo::Cosmology; kwargs... ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing and the Integrated Gravitational Potential (GP) effects arising from the Luminosity Distance (LD) perturbations.

In the first method, you should pass the two Point (P1 and P2) where to evaluate the function, while in the second method (that internally recalls the first) you must provide the two corresponding comoving distances s1 and s2. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{\kappa\int\!\phi} (s_1, s_2, y) = \int_0^{s_1} \mathrm{d}\chi_1 \int_0^{s_2} \mathrm{d}\chi_2 \; \mathcal{J}^{\kappa\int\!\phi}_{\alpha}\left[ \mathcal{J}^{\kappa\int\!\phi}_{31} I_1^3(\Delta\chi) + \mathcal{J}^{\kappa\int\!\phi}_{22} I_2^2(\Delta\chi) \right] \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{\kappa\int\!\phi}_{\alpha} & = \frac{9}{2} \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 \frac{ D(\chi_1) D(\chi_2) \chi_2 (s_1 - \chi_1) }{ s_1 a(\chi_1) a(\chi_2) } \left[ \mathcal{H}(\chi_2)(f(\chi_2) - 1)\mathfrak{R}_2 -\frac{1}{s_2} \right] \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa\int\!\phi}_{31} & = -2 y \Delta\chi^2 \, , \\ %%%%%%%%%%%%%%%%%%%%%%%% \mathcal{J}^{\kappa\int\!\phi}_{22} & = \chi_1 \chi_2(1 - y^2) \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

This function is computed integrating integrand_ξ_LD_Lensing_IntegratedGP with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1 and s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

Keyword Arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs_2::Int = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1) and (0, s2) (for χ2); it has been checked that with N_χs_2 ≥ 50 the result is stable.

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_IntegratedGP_Lensing — Function

ξ_LD_IntegratedGP_Lensing(
    s1, s2, y, cosmo::Cosmology; 
    kwargs... ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) and the Lensing effects arising from the Luminosity Distance (LD) perturbations.

It's computed through the symmetric function ξ_LD_Lensing_IntegratedGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole, ξ_LD_Lensing_IntegratedGP

GaPSE.ξ_LD_LocalGP_IntegratedGP — Function

ξ_LD_LocalGP_IntegratedGP(
    s1, s2, y, cosmo::Cosmology;
    en::Float64 = 1e6, N_χs::Int = 100 ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) and the Integrated GP effects arising from the Luminosity Distance (LD) perturbations.

You must provide the two comoving distances s1 and s2 where to evaluate the function. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of this TPCF is the following:

\[\begin{split} \xi^{\phi\int\!\phi} (s_1, s_2, y) = \int_0^{s_2} \mathrm{d}\chi_2 \, \mathcal{J}^{\phi \int\!\phi}_{40} \tilde{I}^4_0(\Delta\chi_2) \, , \end{split}\]

with

\[\begin{split} \mathcal{J}^{\phi \int\!\phi}_{40} = \frac{9 \mathcal{H}_0^4 \Omega_{\mathrm{M}0}^2 D_1}{2 a_1} (\mathfrak{R}_1 + 1) \frac{D(\chi_2) \Delta\chi_2^4}{a(\chi_2)} \left[ \mathcal{H}(\chi_2)(f(\chi_2) - 1)\mathfrak{R}_2 - \frac{1}{s_2} \right] \, , \end{split}\]

where:

$s_1$ and $s_2$ are comoving distances;

$D_1 = D(s_1)$, ... is the linear growth factor (evaluated in $s_1$);

$a_1 = a(s_1)$, ... is the scale factor (evaluated in $s_1$);

$f_1 = f(s_1)$, ... is the linear growth rate (evaluated in $s_1$);

$\mathcal{H}_1 = \mathcal{H}(s_1)$, ... is the comoving Hubble distances (evaluated in $s_1$);

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$;

$\mathfrak{R}_1 = \mathfrak{R}(s_1)$, ... is computed by func_ℛ_LD in cosmo::Cosmology (and evaluated in $s_1$ ); the definition of $\mathcal{R}(s)$ is the following:

\[\mathfrak{R}(s) = 1 - \frac{1}{\mathcal{H}(s) s} ;\]

$\Omega_{\mathrm{M}0} = \Omega_{\mathrm{cdm}} + \Omega_{\mathrm{b}}$ is the sum of cold-dark-matter and barionic density parameters (again, stored in cosmo);

$I_\ell^n$ and $\sigma_i$ are defined as

\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n} \quad , \quad \sigma_i = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^{2-i} \, P(q)\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\tilde{I}_0^4$ is defined as

\[\tilde{I}_0^4 = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_0(qs) - 1}{(qs)^4}\]

with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo) and $j_\ell$ as spherical Bessel function of order $\ell$;

$\mathcal{H}_0$, $f_0$ and so on are evaluated at the observer position (i.e. at present day);

$\Delta\chi_1 := \sqrt{\chi_1^2 + s_2^2-2\,\chi_1\,s_2\,y}$ and $\Delta\chi_2 := \sqrt{s_1^2 + \chi_2^2-2\,s_1\,\chi_2\,y}$;

$s=\sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\Delta\chi := \sqrt{\chi_1^2 + \chi_2^2-2\,\chi_1\,\chi_2\,y}$.

This function is computed integrating integrand_ξ_LD_LocalGP_IntegratedGP with trapz() from the Trapz Julia package.

Inputs

P1::Point and P2::Point, or s1,s2: Point/comoving distances where the TPCF has to be calculated; they contain all the data of interest needed for this calculus (comoving distance, growth factor and so on).

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $\Omega_{\mathrm{M}0}$, ...

Keyword arguments

en::Float64 = 1e6: just a float number used in order to deal better with small numbers;

N_χs::Int = 100: number of points to be used for sampling the integral along the ranges (0, s1) (for χ1); it has been checked that with N_χs ≥ 50 the result is stable.

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole

GaPSE.ξ_LD_IntegratedGP_LocalGP — Function

ξ_LD_IntegratedGP_LocalGP(
    s1, s2, y, cosmo::Cosmology; 
    kwargs... ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) and the Local GP effects arising from the Luminosity Distance (LD) perturbations.

It's computed through the symmetric function ξ_LD_IntegratedGP_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_LD_multipole, map_ξ_LD_multipole, print_map_ξ_LD_multipole, ξ_LD_IntegratedGP_LocalGP