Two-Point Auto-Correlation Functions

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Newtonian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Doppler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Lensing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_LocalGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_IntegratedGP. Check Documenter's build log for details.

Two-Point Cross-Correlation Functions

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Doppler_Newtonian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Lensing_Newtonian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_LocalGP_Newtonian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_IntegratedGP_Newtonian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Doppler_Doppler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Lensing_Doppler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_LocalGP_Doppler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_IntegratedGP_Doppler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Doppler_Lensing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Lensing_Lensing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_LocalGP_Lensing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_IntegratedGP_Lensing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Doppler_LocalGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Lensing_LocalGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_LocalGP_LocalGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_IntegratedGP_LocalGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Doppler_IntegrateGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_Lensing_IntegrateGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_LocalGP_IntegrateGP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.ξ_LDxGNC_IntegratedGP_IntegrateGP. Check Documenter's build log for details.

Two-Point Cross-Correlation Function multipoles

GaPSE.integrand_ξ_LDxGNC_multipole — Function

 integrand_ξ_LDxGNC_multipole(s1, s, μ, effect::Function, cosmo::Cosmology;
      L::Int = 0, use_windows::Bool = true, kwargs...)

 integrand_ξ_LDxGNC_multipole(s1, s, μ, effect::String, cosmo::Cosmology; kwargs...)

Return the integrand on $\mu = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}$ of the chosen two-point cross correlation between a perturbation of the Luminosity Distance (LD) effect and a Galaxy Number Counts (GNC) effect effect term, i.e. the following function $f(s_1, s, \mu)$:

\[ f_L(s_1, s, \mu) = \xi \left(s_1, s_2, y\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$s_2 = s_2(s_1, s, \mu) = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}$

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2 = \frac{\mu \, s + s_1}{s_2(s_1, s, \mu)}$

$\xi$ is the corresponding LDxGNC TPCF effect

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

In the former method you have to pass as an input the LDxGNC effect function you want to integrate, while in the (recommended) latter one it's necessary to specify the name of the LDxGNC TPCF term among the following:

doppler_newton , lensing_newton , localgp_newton , integratedgp_newton , doppler_doppler , lensing_doppler , localgp_doppler , integratedgp_doppler , doppler_lensing , lensing_lensing , localgp_lensing , integratedgp_lensing , doppler_localgp , lensing_localgp , localgp_localgp , integratedgp_localgp , doppler_integratedgp , lensing_integratedgp , localgp_integratedgp , integratedgp_integratedgp ,

to which correspond the following functions, respectively:

ξ_LDxGNC_Doppler_Newtonian , ξ_LDxGNC_Lensing_Newtonian , ξ_LDxGNC_LocalGP_Newtonian , ξ_LDxGNC_IntegratedGP_Newtonian , ξ_LDxGNC_Doppler_Doppler , ξ_LDxGNC_Lensing_Doppler , ξ_LDxGNC_LocalGP_Doppler , ξ_LDxGNC_IntegratedGP_Doppler , ξ_LDxGNC_Doppler_Lensing , ξ_LDxGNC_Lensing_Lensing , ξ_LDxGNC_LocalGP_Lensing , ξ_LDxGNC_IntegratedGP_Lensing , ξ_LDxGNC_Doppler_LocalGP , ξ_LDxGNC_Lensing_LocalGP , ξ_LDxGNC_LocalGP_LocalGP , ξ_LDxGNC_IntegratedGP_LocalGP , ξ_LDxGNC_Doppler_IntegratedGP , ξ_LDxGNC_Lensing_IntegratedGP , ξ_LDxGNC_LocalGP_IntegratedGP , ξ_LDxGNC_IntegratedGP_IntegratedGP ,

Internally, the second method recall the first, so all the keyword arguments are in common.

Note that these is an internal conversion of coordiate sistems from (s1, s, μ) to (s1, s2, y) thorugh the functions y and s2. The inverse relations are:

$s = s(s_1, s_2, y) = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$

$\mu = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}} = \frac{y \, s_2 - s_1}{s(s_1, s_2, y)}$.

Inputs

s1: the comoving distance where must be evaluated the integral

s: the comoving distance from s1 where must be evaluated the integral

μ: the cosine between s1 and s where must be evaluated the integral

effect::Function or effect::String : name/function you want to consider, among the ones inside GaPSE.GR_EFFECTS_LDxGNC/GaPSE.VEC_ξs_LDxGNC respectively.

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $\mathcal{F}$

kwargs... : other keyword arguments that will be passed to the selected GR TPCF effect (ξ_LDxGNC_Doppler, ξ_LDxGNC_Lensing, ...)

See also: ξ_LDxGNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, WindowFIntegrated, ϕ, Cosmology, y, s2, GR_EFFECTS_LDxGNC, GaPSE.VEC_ξs_LDxGNC

GaPSE.ξ_LDxGNC_multipole — Function

 ξ_LDxGNC_multipole(
      s1, s, effect::Function, cosmo::Cosmology;
      L::Int = 0, alg::Symbol = :lobatto, 
      use_windows::Bool = true, 
      N_lob::Int = 100, N_trap::Int = 200, 
      atol_quad::Float64 = 0.0, rtol_quad::Float64 = 1e-2,
      enhancer::Float64 = 1e6, 
      kwargs...) ::Float64

 ξ_LDxGNC_multipole(s1, s, effect::String, cosmo::Cosmology; 
      kwargs...) ::Float64

Evaluate the multipole of order L of the chosen two-point cross correlation between a perturbation of the Luminosity Distance (LD) and a Galaxy Number Counts (GNC) effect term i.e. the following function $\xi_L(s_1, s)$:

\[ \xi_L(s_1, s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$s_2 = s_2(s_1, s, \mu) = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}$

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2 = \frac{\mu \, s + s_1}{s_2(s_1, s, \mu)}$

$\xi$ is the corresponding LDxGNC TPCF effect

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

In the former method you have to pass as an input the LDxGNC effect function you want to integrate, while in the (recommended) latter one it's necessary to specify the name of the LDxGNC TPCF term among the following:

doppler_newton , lensing_newton , localgp_newton , integratedgp_newton , doppler_doppler , lensing_doppler , localgp_doppler , integratedgp_doppler , doppler_lensing , lensing_lensing , localgp_lensing , integratedgp_lensing , doppler_localgp , lensing_localgp , localgp_localgp , integratedgp_localgp , doppler_integratedgp , lensing_integratedgp , localgp_integratedgp , integratedgp_integratedgp ,

to which correspond the following functions:

ξ_LDxGNC_Doppler_Newtonian , ξ_LDxGNC_Lensing_Newtonian , ξ_LDxGNC_LocalGP_Newtonian , ξ_LDxGNC_IntegratedGP_Newtonian , ξ_LDxGNC_Doppler_Doppler , ξ_LDxGNC_Lensing_Doppler , ξ_LDxGNC_LocalGP_Doppler , ξ_LDxGNC_IntegratedGP_Doppler , ξ_LDxGNC_Doppler_Lensing , ξ_LDxGNC_Lensing_Lensing , ξ_LDxGNC_LocalGP_Lensing , ξ_LDxGNC_IntegratedGP_Lensing , ξ_LDxGNC_Doppler_LocalGP , ξ_LDxGNC_Lensing_LocalGP , ξ_LDxGNC_LocalGP_LocalGP , ξ_LDxGNC_IntegratedGP_LocalGP , ξ_LDxGNC_Doppler_IntegratedGP , ξ_LDxGNC_Lensing_IntegratedGP , ξ_LDxGNC_LocalGP_IntegratedGP , ξ_LDxGNC_IntegratedGP_IntegratedGP ,

Internally, the second method recall the first, so all the keyword arguments are in common.

Note that these is an internal conversion of coordiate sistems from (s1, s, μ) to (s1, s2, y) thorugh the functions y and s2. The inverse relations are:

$s = s(s_1, s_2, y) = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$

$\mu = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}} = \frac{y \, s_2 - s_1}{s(s_1, s_2, y)}$.

Inputs

s1: the comoving distance where must be evaluated the integral

s: the comoving distance from s1 where must be evaluated the integral

effect::Function or effect::String : name/function you want to consider, among the ones inside GaPSE.GR_EFFECTS_LDxGNC/GaPSE.VEC_ξs_LDxGNC respectively.

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Int = 0: order of the Legendre polynomial to be used

alg::Symbol = :trap : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature. WE RECOMMEND TO USE :quad FOR MONOPOLES AND :lobatto FOR HIGHER ORDER MULTIPOLES!

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $\mathcal{F}$

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

kwargs... : other keyword arguments that will be passed to the selected LDxGNC TPCF effect (ξ_LDxGNC_Doppler, ξ_LDxGNC_Lensing, ...)

See also: integrand_ξ_LDxGNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole WindowFIntegrated, Cosmology, y, s2, GR_EFFECTS_LDxGNC, VEC_ξs_LDxGNC

GaPSE.map_ξ_LDxGNC_multipole — Function

 map_ξ_LDxGNC_multipole(cosmo::Cosmology,
      effect::Union{String,Function}, ss = nothing;
      s1 = nothing, L::Int = 0, alg::Symbol = :lobatto,
      N_lob::Int = 100, N_trap::Int = 50,
      atol_quad::Float64 = 0.0, rtol_quad::Float64 = 1e-2,
      enhancer::Float64=1e6, N_log::Int = 1000, 
      pr::Bool = true,
      kwargs...) ::Tuple{Vector{Float64}, Vector{Float64}}

Evaluate the multipole of order L of the chosen two-point cross correlation between a perturbation of the Luminosity Distance (LD) and a Galaxy Number Counts (GNC) effect term for all the comoving distance values stored inside ss. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). If s1 = nothing, it is set s1 = cosmo.s_eff.

The function evaluated is then the following $\xi_L(s_1, s)$:

\[ \xi_L(s_1, s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$s_2 = s_2(s_1, s, \mu) = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}$

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2 = \frac{\mu \, s + s_1}{s_2(s_1, s, \mu)}$

$\xi$ is the corresponding LDxGNC TPCF effect

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

Note that these is an internal conversion of coordiate sistems from (s1, s, μ) to (s1, s2, y) thorugh the functions y and s2. The inverse relations are:

$s = s(s_1, s_2, y) = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$

$\mu = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}} = \frac{y \, s_2 - s_1}{s(s_1, s_2, y)}$.

Inputs

cosmo::Cosmology : cosmology to be used in this computation

effect::Union{String,Function} : the GR effect TPCF you want to consider; you may specify the name of the effect as one of the following strings (recommended):

doppler_newton , lensing_newton , localgp_newton , integratedgp_newton , doppler_doppler , lensing_doppler , localgp_doppler , integratedgp_doppler , doppler_lensing , lensing_lensing , localgp_lensing , integratedgp_lensing , doppler_localgp , lensing_localgp , localgp_localgp , integratedgp_localgp , doppler_integratedgp , lensing_integratedgp , localgp_integratedgp , integratedgp_integratedgp ,

or directly the name of the function among the following:

ξ_LDxGNC_Doppler_Newtonian , ξ_LDxGNC_Lensing_Newtonian , ξ_LDxGNC_LocalGP_Newtonian , ξ_LDxGNC_IntegratedGP_Newtonian , ξ_LDxGNC_Doppler_Doppler , ξ_LDxGNC_Lensing_Doppler , ξ_LDxGNC_LocalGP_Doppler , ξ_LDxGNC_IntegratedGP_Doppler , ξ_LDxGNC_Doppler_Lensing , ξ_LDxGNC_Lensing_Lensing , ξ_LDxGNC_LocalGP_Lensing , ξ_LDxGNC_IntegratedGP_Lensing , ξ_LDxGNC_Doppler_LocalGP , ξ_LDxGNC_Lensing_LocalGP , ξ_LDxGNC_LocalGP_LocalGP , ξ_LDxGNC_IntegratedGP_LocalGP , ξ_LDxGNC_Doppler_IntegratedGP , ξ_LDxGNC_Lensing_IntegratedGP , ξ_LDxGNC_LocalGP_IntegratedGP , ξ_LDxGNC_IntegratedGP_IntegratedGP ,

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Optional arguments

s1 = nothing : comoving distance from the observer where the TPCF should be evaluated; if s1 = nothing, it is automatically set s1 = cosmo.s_eff from the given input cosmo::Cosmology.

L::Int = 0: order of the Legendre polynomial to be used

alg::Symbol = :trap : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature. WE RECOMMEND TO USE :quad FOR MONOPOLES AND :lobatto FOR HIGHER ORDER MULTIPOLES!

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $\mathcal{F}$

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

kwargs... : other keyword arguments that will be passed to the selected LDxGNC TPCF effect (ξ_LDxGNC_Doppler, ξ_LDxGNC_Lensing, ...)

Returns

A Tuple{Vector{Float64}, Vector{Float64}}, which has as first element the ss vector and as second one the corresponding ξ value evaluated.

See also: integrand_ξ_LDxGNC_multipole, ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, WindowFIntegrated, Cosmology, y, s2, GR_EFFECTS_LDxGNC, GaPSE.VEC_ξs_LDxGNC

GaPSE.print_map_ξ_LDxGNC_multipole — Function

 print_map_ξ_LDxGNC_multipole(
      cosmo::Cosmology, out::String,
      effect::Union{String,Function},
      ss = nothing;
      s1 = nothing, L::Int = 0, alg::Symbol = :lobatto,
      N_lob::Int = 100, N_trap::Int = 50,
      atol_quad::Float64 = 0.0, rtol_quad::Float64 = 1e-2,
      enhancer::Float64=1e6, N_log::Int = 1000, 
      pr::Bool = true,
      kwargs...)

Evaluate the multipole of order L of the of the chosen two-point cross correlation between a perturbation of the Luminosity Distance (LD) and a Galaxy Number Counts (GNC) effect term for all the comoving distance values stored inside ss, and print the results (with all the options used) in a file named out. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). If s1 = nothing, it is set s1 = cosmo.s_eff.

The function evaluated is then the following $\xi_L(s_1, s)$:

\[ \xi_L(s_1, s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$s_2 = s_2(s_1, s, \mu) = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}$

$y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2 = \frac{\mu \, s + s_1}{s_2(s_1, s, \mu)}$

$\xi$ is the corresponding LDxGNC TPCF effect

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

Note that these is an internal conversion of coordiate sistems from (s1, s, μ) to (s1, s2, y) thorugh the functions y and s2. The inverse relations are:

$s = s(s_1, s_2, y) = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$

$\mu = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}} = \frac{y \, s_2 - s_1}{s(s_1, s_2, y)}$.

Inputs

cosmo::Cosmology : cosmology to be used in this computation

effect::Union{String,Function} : the GR effect TPCF you want to consider; you may specify the name of the effect as one of the following strings (recommended):

doppler_newton , lensing_newton , localgp_newton , integratedgp_newton , doppler_doppler , lensing_doppler , localgp_doppler , integratedgp_doppler , doppler_lensing , lensing_lensing , localgp_lensing , integratedgp_lensing , doppler_localgp , lensing_localgp , localgp_localgp , integratedgp_localgp , doppler_integratedgp , lensing_integratedgp , localgp_integratedgp , integratedgp_integratedgp ,

or directly the name of the function among the following:

ξ_LDxGNC_Doppler_Newtonian , ξ_LDxGNC_Lensing_Newtonian , ξ_LDxGNC_LocalGP_Newtonian , ξ_LDxGNC_IntegratedGP_Newtonian , ξ_LDxGNC_Doppler_Doppler , ξ_LDxGNC_Lensing_Doppler , ξ_LDxGNC_LocalGP_Doppler , ξ_LDxGNC_IntegratedGP_Doppler , ξ_LDxGNC_Doppler_Lensing , ξ_LDxGNC_Lensing_Lensing , ξ_LDxGNC_LocalGP_Lensing , ξ_LDxGNC_IntegratedGP_Lensing , ξ_LDxGNC_Doppler_LocalGP , ξ_LDxGNC_Lensing_LocalGP , ξ_LDxGNC_LocalGP_LocalGP , ξ_LDxGNC_IntegratedGP_LocalGP , ξ_LDxGNC_Doppler_IntegratedGP , ξ_LDxGNC_Lensing_IntegratedGP , ξ_LDxGNC_LocalGP_IntegratedGP , ξ_LDxGNC_IntegratedGP_IntegratedGP ,

out::String : name of the file where the results must be stored.

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

Optional arguments

This function recall internally map_ξ_LDxGNC_multipole, so the kwargs are the same; we report them for comfortness:

s1 = nothing : comoving distance from the observer where the TPCF should be evaluated; if s1 = nothing, it is automatically set s1 = cosmo.s_eff from the given input cosmo::Cosmology.

L::Int = 0: order of the Legendre polynomial to be used

alg::Symbol = :trap : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature. WE RECOMMEND TO USE :quad FOR MONOPOLES AND :lobatto FOR HIGHER ORDER MULTIPOLES!

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $\mathcal{F}$

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

kwargs... : other keyword arguments that will be passed to the selected LDxGNC TPCF effect (ξ_LDxGNC_Doppler, ξ_LDxGNC_Lensing, ...)

See also: integrand_ξ_LDxGNC_multipole, ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, WindowFIntegrated, Cosmology, y, s2, GR_EFFECTS_LDxGNC, GaPSE.VEC_ξs_LDxGNC

Two-Point Cross-Correlation Function Sum multipoles

GaPSE.sum_ξ_LDxGNC_multipole — Function

 sum_ξ_LDxGNC_multipole(s1, s, cosmo::Cosmology;
      s1 = nothing, L::Int = 0, alg::Symbol = :lobatto,
      N_lob::Int = 100, N_trap::Int = 50,
      atol_quad::Float64 = 0.0, rtol_quad::Float64 = 1e-2,
      enhancer::Float64=1e6, N_log::Int = 1000, 
      pr::Bool = true,
      kwargs...) ::Tuple{Float64, Vector{Float64}}

Evaluate the multipole of order L of all the two-point cross correlation multipole between the perturbations of the Luminosity Distance (LD) and the Galaxy Number Counts (GNC) and their sum in the comoving distance s1 and a comoving distance s from it for the input cosmo::Cosmology. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

This function makes a for-loop on the GaPSE.GR_EFFECTS_LDxGNC strings, calling ξ_LDxGNC_multipole for each of them. To each string corresponds pretty intuitively one of the 20 LDxGNC effects. They are currently, in order:

doppler_newton , lensing_newton , localgp_newton , integratedgp_newton , doppler_doppler , lensing_doppler , localgp_doppler , integratedgp_doppler , doppler_lensing , lensing_lensing , localgp_lensing , integratedgp_lensing , doppler_localgp , lensing_localgp , localgp_localgp , integratedgp_localgp , doppler_integratedgp , lensing_integratedgp , localgp_integratedgp , integratedgp_integratedgp ,

Inputs

s1: the comoving distance where must be evaluated the integral

s: the comoving distance from s1 where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

This function recall internally ξ_LDxGNC_multipole, so the kwargs are the same; we report them for comfortness:

L::Int = 0: order of the Legendre polynomial to be used

alg::Symbol = :trap : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature. WE RECOMMEND TO USE :quad FOR MONOPOLES AND :lobatto FOR HIGHER ORDER MULTIPOLES!

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $\mathcal{F}$

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

kwargs... : other keyword arguments that will be passed to ALL the LDxGNC TPCF effect (ξ_LDxGNC_Doppler, ξ_LDxGNC_Lensing, ...); if one of them has that keyword argument, it will use the given value, otherwise it will be unaffected.

Returns

A tuple containing:

the sum of all the ξ multipoles as first element

a Vector{Float64} with all the values of each ξ; they are ordered following GR_EFFECTS_LDxGNC

See also: integrand_ξ_LDxGNC_multipole, ξ_LDxGNC_multipole, map_sum_ξ_LDxGNC_multipole, print_map_sum_ξ_LDxGNC_multipole, Cosmology, GR_EFFECTS_LDxGNC

GaPSE.map_sum_ξ_LDxGNC_multipole — Function

 map_sum_ξ_LDxGNC_multipole(
      effect::Union{String,Function}, ss = nothing;
      s1 = nothing, L::Int = 0, alg::Symbol = :lobatto,
      N_lob::Int = 100, N_trap::Int = 50,
      atol_quad::Float64 = 0.0, rtol_quad::Float64 = 1e-2,
      enhancer::Float64=1e6, N_log::Int = 1000, 
      pr::Bool = true,
      kwargs...) ::Tuple{Vector{Float64}, Vector{Float64}, Vector{Vector{Float64}}}

Evaluate the multipole of order L of all the two-point cross correlation multipole between the perturbations of the Luminosity Distance (LD) and the Galaxy Number Counts (GNC) and their sum in the comoving distance s1, for all the comoving distances stored inside ss (representing the comoving distance from s1) for the input cosmo::Cosmology. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). If s1 = nothing, it is set s1 = cosmo.s_eff. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

This function makes a for-loop on the GaPSE.GR_EFFECTS_LDxGNC strings, calling map_ξ_LDxGNC_multipole for each of them. To each string corresponds pretty intuitively one of the 20 LDxGNC effects. They are currently, in order:

doppler_newton , lensing_newton , localgp_newton , integratedgp_newton , doppler_doppler , lensing_doppler , localgp_doppler , integratedgp_doppler , doppler_lensing , lensing_lensing , localgp_lensing , integratedgp_lensing , doppler_localgp , lensing_localgp , localgp_localgp , integratedgp_localgp , doppler_integratedgp , lensing_integratedgp , localgp_integratedgp , integratedgp_integratedgp ,

Inputs

cosmo::Cosmology : cosmology to be used in this computation

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Optional arguments

This function recall internally map_ξ_LDxGNC_multipole, so the kwargs are the same; we report them for comfortness:

s1 = nothing : comoving distance from the observer where the TPCF should be evaluated; if s1 = nothing, it is automatically set s1 = cosmo.s_eff from the given input cosmo::Cosmology.

L::Int = 0: order of the Legendre polynomial to be used

alg::Symbol = :trap : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature. WE RECOMMEND TO USE :quad FOR MONOPOLES AND :lobatto FOR HIGHER ORDER MULTIPOLES!

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $\mathcal{F}$

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

kwargs... : other keyword arguments that will be passed to ALL the LDxGNC TPCF effect (ξ_LDxGNC_Doppler, ξ_LDxGNC_Lensing, ...); if one of them has that keyword argument, it will use the given value, otherwise it will be unaffected.

Returns

A tuple containing:

as first element, the vector ss itself;

as second one, the Vector{Float64} of the sum of all the ξ multipoles;

as third one, a Vector{Vector{Float64}} with all the values of each ξ; they are ordered following GR_EFFECTS_LDxGNC

See also: map_ξ_LDxGNC_multipole, sum_ξ_LDxGNC_multipole, print_map_sum_ξ_LDxGNC_multipole, Cosmology, GR_EFFECTS_LDxGNC

GaPSE.print_map_sum_ξ_LDxGNC_multipole — Function

 print_map_sum_ξ_LDxGNC_multipole(
      cosmo::Cosmology, out::String, ss = nothing;
      s1 = nothing, L::Int = 0, alg::Symbol = :lobatto,
      N_lob::Int = 100, N_trap::Int = 50,
      atol_quad::Float64 = 0.0, rtol_quad::Float64 = 1e-2,
      enhancer::Float64=1e6, N_log::Int = 1000, 
      pr::Bool = true,
      single::Bool = true,
      kwargs...)

Evaluate the multipole of order L of all the two-point cross correlation multipole between the perturbations of the Luminosity Distance (LD) and the Galaxy Number Counts (GNC) and their sum in the comoving distance s1, for all the comoving distances stored inside ss (representing the comoving distance from s1) for the input cosmo::Cosmology; finally, it saves the results inside the file out. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). If s1 = nothing, it is set s1 = cosmo.s_eff. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

This function makes a for-loop on the GaPSE.GR_EFFECTS_LDxGNC strings, calling map_ξ_LDxGNC_multipole for each of them. To each string corresponds pretty intuitively one of the 20 LDxGNC effects. They are currently, in order:

doppler_newton , lensing_newton , localgp_newton , integratedgp_newton , doppler_doppler , lensing_doppler , localgp_doppler , integratedgp_doppler , doppler_lensing , lensing_lensing , localgp_lensing , integratedgp_lensing , doppler_localgp , lensing_localgp , localgp_localgp , integratedgp_localgp , doppler_integratedgp , lensing_integratedgp , localgp_integratedgp , integratedgp_integratedgp ,

Inputs

cosmo::Cosmology : cosmology to be used in this computation

out::String : name of the file where the results must be stored.

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

Optional arguments

This function recall internally map_ξ_LDxGNC_multipole, so the kwargs are the same; we report them for comfortness:

s1 = nothing : comoving distance from the observer where the TPCF should be evaluated; if s1 = nothing, it is automatically set s1 = cosmo.s_eff from the given input cosmo::Cosmology.

L::Int = 0: order of the Legendre polynomial to be used

alg::Symbol = :trap : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature. WE RECOMMEND TO USE :quad FOR MONOPOLES AND :lobatto FOR HIGHER ORDER MULTIPOLES!

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $\mathcal{F}$

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

single::Bool = true : if true, all the CFs are printed inside the file of the sum, in a table with 18 columns (first one for ss, second for their sum an the next 20 for each effect). Otherwise, a new directory "allstandalonesCFs" is created (in the same path given in out) and they are separately saved in files there placed.

kwargs... : other keyword arguments that will be passed to ALL the LDxGNC TPCF effect (ξ_LDxGNC_Doppler, ξ_LDxGNC_Lensing, ...); if one of them has that keyword argument, it will use the given value, otherwise it will be unaffected.

See also: map_ξ_LDxGNC_multipole, sum_ξ_LDxGNC_multipole, map_sum_ξ_LDxGNC_multipole, Cosmology, GR_EFFECTS_LDxGNC