LDxGNC TPCFs

Two-Point Cross-Correlation Functions

GaPSE.ξ_LDxGNC_Doppler_Newtonian — Function

ξ_LDxGNC_Doppler_Newtonian(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler effect arising from the Luminosity Distance (LD) perturbations and the Newtonian one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Newtonian_Doppler; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Newtonian_Doppler

GaPSE.ξ_LDxGNC_Lensing_Newtonian — Function

ξ_LDxGNC_Lensing_Newtonian(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing effect arising from the Luminosity Distance (LD) perturbations and the Newtonian one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Newtonian_Lensing; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Newtonian_Lensing

GaPSE.ξ_LDxGNC_LocalGP_Newtonian — Function

ξ_LDxGNC_LocalGP_Newtonian(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Newtonian one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Newtonian_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Newtonian_LocalGP

GaPSE.ξ_LDxGNC_IntegratedGP_Newtonian — Function

ξ_LDxGNC_IntegratedGP_Newtonian(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Newtonian one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Newtonian_IntegratedGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Newtonian_IntegratedGP

GaPSE.ξ_LDxGNC_Doppler_Doppler — Function

ξ_LDxGNC_Doppler_Doppler(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler effect arising from the Luminosity Distance (LD) perturbations and the Doppler one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Doppler_Doppler; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Doppler_Doppler

GaPSE.ξ_LDxGNC_Lensing_Doppler — Function

ξ_LDxGNC_Lensing_Doppler(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing effect arising from the Luminosity Distance (LD) perturbations and the Doppler one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Doppler_Lensing; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Doppler_Lensing

GaPSE.ξ_LDxGNC_LocalGP_Doppler — Function

ξ_LDxGNC_LocalGP_Doppler(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Doppler one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Doppler_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Doppler_LocalGP

GaPSE.ξ_LDxGNC_IntegratedGP_Doppler — Function

ξ_LDxGNC_IntegratedGP_Doppler(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Doppler one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Doppler_IntegratedGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Doppler_IntegratedGP

GaPSE.ξ_LDxGNC_Doppler_Lensing — Function

ξ_LDxGNC_Doppler_Lensing(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler effect arising from the Luminosity Distance (LD) perturbations and the Lensing one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Lensing_Doppler; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Lensing_Doppler

GaPSE.ξ_LDxGNC_Lensing_Lensing — Function

ξ_LDxGNC_Lensing_Lensing(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing effect arising from the Luminosity Distance (LD) perturbations and the Lensing one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Lensing_Lensing; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Lensing_Lensing

GaPSE.ξ_LDxGNC_LocalGP_Lensing — Function

ξ_LDxGNC_LocalGP_Lensing(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Lensing one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Lensing_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Lensing_LocalGP

GaPSE.ξ_LDxGNC_IntegratedGP_Lensing — Function

ξ_LDxGNC_IntegratedGP_Lensing(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Lensing one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_Lensing_IntegratedGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_Lensing_IntegratedGP

GaPSE.ξ_LDxGNC_Doppler_LocalGP — Function

ξ_LDxGNC_Doppler_LocalGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler effect arising from the Luminosity Distance (LD) perturbations and Local Gravitational Potential (GP) one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_LocalGP_Doppler; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_LocalGP_Doppler

GaPSE.ξ_LDxGNC_Lensing_LocalGP — Function

ξ_LDxGNC_Lensing_LocalGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing effect arising from the Luminosity Distance (LD) perturbations and Local Gravitational Potential (GP) one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_LocalGP_Lensing; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_LocalGP_Lensing

GaPSE.ξ_LDxGNC_LocalGP_LocalGP — Function

ξ_LDxGNC_LocalGP_LocalGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Local GP one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_LocalGP_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_LocalGP_LocalGP

GaPSE.ξ_LDxGNC_IntegratedGP_LocalGP — Function

ξ_LDxGNC_IntegratedGP_LocalGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Local GP one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_LocalGP_IntegratedGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_LocalGP_IntegratedGP

GaPSE.ξ_LDxGNC_Doppler_IntegratedGP — Function

ξ_LDxGNC_Doppler_IntegratedGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Doppler effect arising from the Luminosity Distance (LD) perturbations and Integrated Gravitational Potential (GP) one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_IntegratedGP_Doppler; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_IntegratedGP_Doppler

GaPSE.ξ_LDxGNC_Lensing_IntegratedGP — Function

ξ_LDxGNC_Lensing_IntegratedGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Lensing effect arising from the Luminosity Distance (LD) perturbations and Integrated Gravitational Potential (GP) one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_IntegratedGP_Lensing; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_IntegratedGP_Lensing

GaPSE.ξ_LDxGNC_LocalGP_IntegratedGP — Function

ξ_LDxGNC_LocalGP_IntegratedGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Local Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Integrated GP one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_IntegratedGP_LocalGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_IntegratedGP_LocalGP

GaPSE.ξ_LDxGNC_IntegratedGP_IntegratedGP — Function

ξ_LDxGNC_IntegratedGP_IntegratedGP(s1, s2, y, cosmo::Cosmology;         
    b1=nothing, b2=nothing, s_b1=nothing, s_b2=nothing, 
    𝑓_evo1=nothing, 𝑓_evo2=nothing, s_lim=nothing,
    obs::Union{Bool, Symbol} = :noobsvel ) ::Float64

Return the Two-Point Correlation Function (TPCF) given by the cross correlation between the Integrated Gravitational Potential (GP) effect arising from the Luminosity Distance (LD) perturbations and the Integrated GP one arising from the Galaxy Number Counts (GNC).

It's computed through the symmetric function ξ_GNCxLD_IntegratedGP_IntegratedGP; check its documentation for more details about the analytical expression and the keyword arguments. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Inputs

s1 and s2: comoving distances where the TPCF has to be calculated;

y: the cosine of the angle between the two points P1 and P2 wrt the observer

cosmo::Cosmology: cosmology to be used in this computation; it contains all the splines used for the conversion s -> Point, and all the cosmological parameters $b$, ...

Keyword Arguments

kwargs... : Keyword arguments to be passed to the symmetric TPCF

See also: Point, Cosmology, ξ_GNC_multipole, map_ξ_LDxGNC_multipole, print_map_ξ_LDxGNC_multipole, ξ_GNCxLD_IntegratedGP_IntegratedGP