Mathematical Utilities functions

GaPSE.warning — Function

warning(io::IO, msg::String)
warning(msg::String) = warning(stdout, msg)

It does the following: printstyled(io, "WARNING: " * msg * " "; color=:red, bold=true)

GaPSE.derivate_point — Function

derivate_point(xp, yp, x1, y1, x2, y2)

Return the derivative in (xp, yp), given the neighboor points (x1,y1) and (x2,y2), with x1 < xp < x2. It is not assumed that x2 - xp = xp - x1.

See also: derivate_vector, spectral_index]

GaPSE.derivate_vector — Function

derivate_vector(XS, YS; N::Int=1)

Apply derivate_point to the vector. N::Int=1 is the number of points at each edge that idk.

See also: derivate_point, spectral_index]

GaPSE.spectral_index — Function

spectral_index(xs, ys; N::Int=1, con=false)

Return the spectral index vector for the two input vectors.

The spectral index $S$ of a generic function $f = f(x)$ is defined as:

\[ S = \frac{\partial \log f(x)}{\partial \log x} = \frac{x}{f(x)} \frac{\partial f(x)}{\partial x} \]

See also: derivate_point, derivate_vector

GaPSE.mean_spectral_index — Function

mean_spectral_index(xs, ys; N::Int = 1, con = false)

Assuming that the input ys follow a power law distribution, return the mean spectral index $\langle S \rangle$ of them.

The spectral index $S$ of a generic function $f = f(x)$ is defined as:

\[ S = \frac{\partial \log f(x)}{\partial \log x} = \frac{x}{f(x)} \frac{\partial f(x)}{\partial x} \]

GaPSE.power_law — Function

power_law(x, si, b, a) ::Float64

Return the following $y = f(x)$ "spurious" power-law value:

\[y = f(x) = a + b \, x^s\]

where si is the exponent ($s$), b the coefficient ($b$) and a is the added constant ($a$).

GaPSE.two_power_laws — Function

two_power_laws(x; switch=5.0, si_1=1.0, si_2=2.0, b=1.0, a=0.0)

Return two power laws, depending on the value of switch.

GaPSE.power_law_from_data — Function

power_law_from_data(xs, ys, p0::Vector{Float64},
    fit_min::Number, fit_max::Number; con = false)

power_law_from_data(xs, ys, p0::Vector{Float64}; con = false) = 
    power_law_from_data(xs, ys, p0, xs[begin], xs[end]; con = con)

Returns the "spurious" power-law coefficients $s$, $b$ and $a$ obtained from the fitting of the data vectors xs and ys inside the limits fit_min and fit_max.

If con == false, the returned a is always 0.0, because it is considered the "pure" power-law fitting function:

\[ y = f(x) = b \, x^s\]

while if con == false it is used the spurious one:

\[ y = f(x) = a + b \, x^s\]

(and consequently $a$ may be ≠0).

The fitting is performed through the function curve_fit of the LsqFit Julia package, which is based on the least-squares method.

See also: power_law

GaPSE.polynomial_from_data — Function

polynomial_from_data(xs, ys, p0::Vector{Float64},
    fit_min::Number, fit_max::Number; con = false)

polynomial_from_data(xs, ys, p0::Vector{Float64}; con = false) = 
    polynomial_from_data(xs, ys, p0, xs[begin], xs[end]; con = con)

Returns the 2-degree polynomial coefficients $c$, $b$ and $a$ obtained from the fitting of the data vectors xs and ys inside the limits fit_min and fit_max.

The fitting function is always a second-degree polynomio, but depeding on the starting values vector p0 can assume three different forms:

if length(p0) == 3, the fitting polynomio is

\[ f(x) = a + b \, x + c \, x^2\]

if length(p0) == 2, the fitting polynomio is

\[ f(x) = b \, x + c \, x^2\]

if length(p0) == 1, the fitting polynomio is

\[ f(x) = c \, x^2\]

The fitting is performed through the function curve_fit of the LsqFit Julia package, which is based on the least-squares method.

See also: polynomial

GaPSE.polynomial — Function

polynomia(x, c, b, a) ::Float64

Return the following $y = f(x)$ 2-degree polynomial value:

\[y = f(x) = a + b \, x + c \, x^2\]

where c is the 2-degree coefficient ($c$), b the 1-degree coefficient ($b$) and a is the added constant ($a$).

GaPSE.expand_left_log — Function

expand_left_log(xs, ys;
    lim=1e-8, fit_min=0.05, fit_max=0.5,
    p0::Union{Vector{Float64},Nothing}=nothing,
    con::Bool=false)

GaPSE.expand_right_log — Function

expand_right_log(xs, ys;
    lim=3e3, fit_min=5.0, fit_max=10.0,
    p0::Union{Vector{Float64},Nothing}=nothing,
    con::Bool=false)

GaPSE.expanded_IPS — Function

expanded_IPS(ks, pks; k_in = 1e-8, k_end = 3e3, con = false)

Given the ks and pks of a chosen Power Spectrum, returns the same PS with "longer tails", i.e. it is prolonged for higher and lower ks than the input ones.

GaPSE.expanded_Iln — Function

expanded_Iln(PK, l, n; N = 1024, kmin = 1e-4, kmax = 1e3, s0 = 1e-3,
    fit_left_min = 2.0, fit_left_max = 10.0, p0 = [-1.0, 1.0, 0.0], con = false)

GaPSE.func_I04_tilde — Function

func_I04_tilde(PK, s, kmin, kmax; kwargs...)

Return the following integral:

\[\tilde{I}^4_0 (s) = \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_0(q s) - 1}{(q s)^4}\]

It is brute-force calcuated with quadgk.

Arguments

PK : function that return the Input Power Spectrum

s : value of $s$ whre the integral must be evaluated

kmin, kmax : extremes (min and max) of integration

kwargs... : keyword argruments that must be passed to quadgk, such as rtol or atol

GaPSE.expanded_I04_tilde — Function

expanded_I04_tilde(PK, ss;
    kmin=1e-6, kmax=1e3, kwargs...)

GaPSE.my_interpolation — Function

my_interpolation(x1, y1, x2, y2, x)

GaPSE.EPLs — Type

EPLs(
    l_si::Float64
    l_b::Float64
    l_a::Float64
    left::Float64

    spline::Dierckx.Spline1D

    r_si::Float64
    r_b::Float64
    r_a::Float64
    right::Float64
)

Contains all the information useful in order to return the value of a spline inside the interval left ≤ x ≤ right and the associated power laws for the edges (with the "left" coefficients l_si, l_b and l_a for x < left and the "right" ones r_si, r_b and r_a for x > right)

Arguments

l_si, l_b, l_a :: Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the LEFT edge; when an input value x < left is given, the returned one is obtained from power_law with this coefficients ( where, of course, l_si is the exponent, l_b the coefficient and l_a the spurious adding constant).

left::Float64 : the break between the left power-law (for x <left) and the spline (for x ≥ left); its value is the xs[begin] one.

spline::Dierckx.Spline1D : spline that interpolates between the real values of the integral calculated inside the range left ≤ x ≤ right

right::Float64 : the break between the right power-law (for x ≥ left) and the spline (for x ≤ right); its value is the xs[end] one.

r_si, r_b, r_a :: Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the RIGHT edge; when an input value x > right is given, the returned one is obtained from power_law with this coefficients ( where, of course, r_si is the exponent, r_bthe coefficient andr_athe spurious adding constant). NOTE: for numerical issues, the "pure" power-lawy = f(x) = b + x^s` should be used.

Constructors

EPLs(xs, ys, p0_left::Vector{T1}, p0_right::Vector{T2}; 
    N_left::Int = 15, N_right::Int = 15) where {T1<:Real, T2 <:Real}

xs and ys: the input vector of values.

N_left::Int = 15 : number of points to be used from the left edge for the left power law-fitting. It shouldn't be too low (< 4) or too high (>100).

N_right::Int = 15 : number of points to be used from the right edge for the right power law-fitting. It shouldn't be too low (< 4) or too high (>100).

p0_left::Vector{T1} where T1 <:Real : vector with the initial values for the left power-law fitting; its length must be 2 (if you want to fit with a pure power-law $y = f(x) = b * x^s$, so only l_si and l_b are matter of concern) or 3 (if you want to fit with a spurious power-law $y = f(x) = a + b * x^s$, so you are also interested in l_a); in the first case, the considered l_a will be 0.0. Example:

p0_right::Vector{T1} where T1 <:Real : vector with the initial values for the right power-law fitting; its length must be 2 (if you want to fit with a pure power-law $y = f(x) = b * x^s$, so only r_si and r_b are matter of concern) or 3 (if you want to fit with a spurious power-law $y = f(x) = a + b * x^s$, so you are also interested in r_a); in the first case, the considered r_a will be 0.0. It is recommended to

All the power-law fitting (both "pure" and spurious) are made through the local function power_law_from_data.

Examples

julia> xs = 10 .^ range(0, 2, length=100);

julia> ys = [1.34e2 * x ^ 2.43 for x in xs];

julia> A = EPLs(xs, ys, [1.0, 1.0], [1.0, 1.0]; N_left = 10, N_right = 10)

See also: power_law_from_data

GaPSE.FFTLog.evaluate_FFTLog! — Function

evaluate_FFTLog!(fy, plan::Union{SingleBesselPlan, HankelPlan}, fx)

Given an input plan::Union{SingleBesselPlan, HankelPlan}, evaluate the FFT fy of the fx y-axis data on the basis of the parameters stored in plan. The result is stored both in plan.fy and in the input fy.

See also: SingleBesselPlan, HankelPlan