Other utilities

GaPSE.parameters_used — Function

parameters_used(io::IO, cosmo::Cosmology; logo::Union{Bool,String} = true)
parameters_used(cosmo::Cosmology; kwargs...) = parameters_used(stdout, cosmo; kwargs...)

Writes in the io stream all the information concerning the input Cosmology cosmo. There are currently three options for logo: true, "simple" and false. With true, the logo that is printed to io is the following:

###################################
#   ____       ____  ____  _____  #
#  / ___| __ _|  _ \/ ___|| ____| #
# | |  _ / _` | |_) \___ \|  _|   #
# | |_| | (_| |  __/ ___) | |___  # 
#  \____|\__,_|_|   |____/|_____| #
###################################
# 

With false, the logo that is printed to io is instead the following:

###############
#    GaPSE    #
############### 
#

Finally, with false, not logo is printed.

See also: Cosmology

GaPSE.check_compatible_dicts — Function

check_compatible_dicts(ref::Dict, b::Dict, name::String = "NO-NAME")

Compare the field of two Dict and check if the second one (b) is "compatible" with the first one (ref), i.e.:

checks if each of the key in b is also a key of ref;

for each key of b:

if typeof(ref[key]) <: Real and !(typeof(ref[k]) <: Union{Bool, Int}), checks that typeof(b[k]) <: Real && typeof(b[k]) ≠ Bool

otherwise, checks that typeof(b[k]) == typeof(ref[k])

If someone of the check mentioned is false, raise an AssertionError, otherwise return nothing. The string name is only used inside the AssertionError messages for the correct name of the input b dictionary.

GaPSE.my_println_vec — Function

my_println_vec(io::IO, vec::Vector{T}, name::String; N::Int=5) where {T}
my_println_vec(vec::Vector{T}, name::String; N::Int=5) where {T}
    my_println_vec(stdout, vec, name; N=N)

Print the input vec::Vector{T} as follows:

> vec = [x for x in 1:0.1:4]
> my_println_vec(vec, "vector"; N=8)
vector = [
1.0 , 1.1 , 1.2 , 1.3 , 1.4 , 1.5 , 1.6 , 1.7 , 
1.8 , 1.9 , 2.0 , 2.1 , 2.2 , 2.3 , 2.4 , 2.5 , 
2.6 , 2.7 , 2.8 , 2.9 , 3.0 , 3.1 , 3.2 , 3.3 , 
3.4 , 3.5 , 3.6 , 3.7 , 3.8 , 3.9 , 4.0 , 
];

> my_println_vec(vec, "vector"; N=3)
vector = [
1.0 , 1.1 , 1.2 , 
1.3 , 1.4 , 1.5 , 
1.6 , 1.7 , 1.8 , 
1.9 , 2.0 , 2.1 , 
2.2 , 2.3 , 2.4 , 
2.5 , 2.6 , 2.7 , 
2.8 , 2.9 , 3.0 , 
3.1 , 3.2 , 3.3 , 
3.4 , 3.5 , 3.6 , 
3.7 , 3.8 , 3.9 , 
4.0 , 
];

See also: my_println_dict

GaPSE.my_println_dict — Function

my_println_dict(io::IO, dict::Dict; pref::String="", N::Int=3)
my_println_dict(dict::Dict; pref::String="", N::Int=3) = 
    my_println_dict(stdout, dict; pref=pref, N=N)

Print the input dict::Dict as follows:

> my_println_dict(io, cosmo.params.IPS; pref="#	 ", N=2)
#	 fit_left_min = 1.0e-6 	 fit_right_min = 10.0 	 
#	 fit_left_max = 3.0e-6 	 fit_right_max = 20.0 

See also: my_println_vec

GaPSE.parent_directory — Function

parent_directory(s::String)::String

Return the name of the parent directory of the input filename s::String. Some examples of use:

s = "/Users/username/Downloads/file.txt" => "/Users/username/Downloads/"

s = "/Users/username/Downloads/" => "/Users/username/"

s = "/Users/username/Downloads" => "/Users/username/"

s = "/Users/username/" => "/Users/"

s = "/Users/username" => "/Users/"

s = "/Users/" => "/"

s = "/Users" => "/"

s = "username/Downloads/file.txt" => "username/Downloads/"

s = "username/Downloads/" => "username/"

s = "username/Downloads" => "username/"

s = "username/" => "./"

s = "username" => "./"

s = "file.txt" => "./"

It's used inside the function check_parent_directory.

See also: check_parent_directory, return_namefile check_namefile

GaPSE.check_parent_directory — Function

check_parent_directory(s::String)

Checks if the input namefile s::String is placed in an existing directory (whose name is obtained from s through the function parent_directory); if not, it raises an AssertionError. Some examples of the directory which are checked:

s = "/Users/username/Downloads/file.txt" => "/Users/username/Downloads/"

s = "/Users/username/Downloads/" => "/Users/username/"

s = "/Users/username/Downloads" => "/Users/username/"

s = "/Users/username/" => "/Users/"

s = "/Users/username" => "/Users/"

s = "/Users/" => "/"

s = "/Users" => "/"

s = "username/Downloads/file.txt" => "username/Downloads/"

s = "username/Downloads/" => "username/"

s = "username/Downloads" => "username/"

s = "username/" => "./"

s = "username" => "./"

s = "file.txt" => "./"

See also: parent_directory, return_namefile check_namefile

GaPSE.return_namefile — Function

return_namefile(s::String)::String

Return the namefile of the input s::String, i.e. it removes the path from the name. Internally it uses the function parent_directory. Some examples of use:

s = "/Users/matteofoglieni/Downloads/" => raises an AssertionError

s = "Downloads/" => raises an AssertionError

s = "./Downloads/" => raises an AssertionError

s = "file" => "file"

s = "file.boh" => "file.boh"

s = "/Users/matteo.foglieni/ciao.file" => "ciao.file"

s = "matteo.foglieni/ciao.file.boh" => "ciao.file.boh"

s = "/Users/matteofoglieni/Downloads/file.txt" => "file.txt"

s = "./file.txt" => "file.txt"

s = "./file.dat" => "file.dat"

s = "file.txt" => "file.txt"

s = "file.dat" => "file.dat"

See also: parent_directory check_parent_directory check_namefile

GaPSE.check_namefile — Function

check_namefile(s::String)

Check if the input namefile s::String is a valid name for a file. Internally it uses the function return_namefile. Some examples of use:

s = "/Users/matteofoglieni/Downloads/" => raises an AssertionError

s = "Downloads/" => raises an AssertionError

s = "./Downloads/" => raises an AssertionError

s = "file" => no raises

s = "file.boh" => no raises

s = "/Users/matteo.foglieni/ciao.file" => no raises

s = "matteo.foglieni/ciao.file.boh" => no raises

s = "/Users/matteofoglieni/Downloads/file.txt" => no raises

s = "./file.txt" => no raises

s = "./file.dat" => no raises

s = "file.txt" => no raises

s = "file.dat" => no raises

See also: parent_directory, check_parent_directory, return_namefile

GaPSE.check_group — Function

check_group(s::String; valid_groups::Vector{String}=VALID_GROUPS)

Check if the input s::String belongs to valid groups; if not, it raises an AssertionError. The default VALID_GROUPS is made by the following strings: LD , GNC , GNCxLD , LDxGNC , generic ,

See also: check_fileisingroup

GaPSE.check_fileisingroup — Function

check_fileisingroup(input::String, group::String;
    valid_groups::Vector{String}=VALID_GROUPS, comments::Bool=true)

Check if the filename input::String constains a set of data that are "compatible" with the input group::String. The default VALID_GROUPS is made by the following strings: LD , GNC , GNCxLD , LDxGNC , generic , If the file start with comments (lines starting with #), set comments = true.

Internally it recalls the function check_group.

See also: check_group

GaPSE.number_to_string — Function

number_to_string(x::Number) ::String

Convert a x::Number into a String with the following conventions:

x = 3 -> "3"

x = -2.15 -> "-2.15"

x = 2.15 * im -> "2.15im"

x = 0.0 + 2.15 * im -> "2.15im"

x = - 0.0 - 2.15 * im -> "-2.15im"

x = -3.1415 - 2.15 * im -> "-3.1415-2.15im"

x = 3.1415 + 2.15 * im -> "3.1415+2.15im"

GaPSE.vecstring_to_vecnumbers — Function

vecstring_to_vecnumbers(v; 
    dt::DataType = Float64 ) ::Vector{dt}

Try to convert a vector of String into a Vector of data type dt. If that raises an Exception, it uses parse elementwise.

GaPSE.readxy — Function

readxy(input::String; comments::Bool=true, 
    xdt::DataType = Float64, ydt::DataType = Float64
    ) ::Tuple{Vector{xdt},Vector{ydt}}

Read the file input and return the two data colums with the input types xdt and ydt. If the file start with comments (lines starting with #), set comments = true.

See also: readxall, readxyall, readxchoosey

GaPSE.readxall — Function

readxall(input::String; comments::Bool=true, 
    xdt::DataType = Float64, ydt::DataType = Float64
    ) ::Tuple{Vector{xdt},Vector{Vector{ydt}}}

Read the file input and return a tuple having

as first element the data in the first column (with the input type xdt)

as second element a vector that contains all the following columns (with the input type ydt)

If the file start with comments (lines starting with #), set comments = true.

See also: readxy, readxyall, readxchoosey

GaPSE.readxyall — Function

readxyall(input::String; comments::Bool=true, 
    xdt::DataType = Float64, ydt::DataType = Float64,
    zdt::DataType=Float64
    ) ::Tuple{Vector{xdt},Vector{ydt},Vector{Vector{zdt}}}

Read the file input and return a tuple having

as first element the data in the first column (with the input type xdt)

as second element the data in the second column (with the input type ydt)

as third element a vector that contains all the following columns (with the input type zdt)

If the file start with comments (lines starting with #), set comments = true.

See also: readxy, readxall, readxchoosey

GaPSE.readchoosen — Function

readchoosen(
    input::String, n::Int; 
    comments::Bool = true, dt::DataType = Float64, 
    ) ::Tuple{Vector{xdt},Vector{ydt}}

Read the file input and return the data (with the input type ydt) stored in the column number n.

If the file start with comments (lines starting with #), set comments = true.

See also: readxy, readxall, readxyall, readxchoosey

GaPSE.readxchoosey — Function

readxchoosey(
    input::String, n::Int; 
    comments::Bool = true, xdt::DataType = Float64, ydt::DataType = Float64,
    ) ::Tuple{Vector{xdt},Vector{ydt}}

readxchoosey(
    input::String, effect::String, group::String = VALID_GROUPS[begin+1]; 
    comments::Bool = true, xdt::DataType = Float64, ydt::DataType = Float64,
    ) ::Tuple{Vector{xdt},Vector{ydt}}

Read the file input and return a tuple having

as first element the data in the first column (with the input type xdt)

as second element the data (with the input type ydt)

stored in the column number n (first method)

in the position corresponding to the input effect for the given group (note: also "sum" is valid as effect here, and gives the sum column, i.e. the second one)

If the file start with comments (lines starting with #), set comments = true.

See also: readxy, readxall, readxyall,readchoosen

GaPSE.sample_subdivision_begin — Function

sample_subdivision_begin(x_min, x_stop, x_end; 
    frac_begin::Float64 = 0.5, N::Int = 100, ass::Bool = true)

Return a vector of N+2 points inside the interval x_min ≤ x ≤ x_max linearly distributed with two different sampling:

in x_min ≤ x ≤ x_stop there are around frac_begin * N points;

in x_stop ≤ x ≤ x_max there are around (1.0 - frac_begin) * N points.

frac_begin is then the fraction of the N points that is inside the LEFT INTERVAL. If ass::Bool is set to false the assert checks on the input data will not be performed.

GaPSE.sample_subdivision_middle — Function

sample_subdivision_middle(x_min, x_start, x_stop, x_max; 
    frac_middle::Float64 = 0.5, rel_frac_begin::Union{Float64, Nothing} = nothing, 
    N::Int = 100, ass::Bool = true)

Return a vector of N+3 points inside the interval x_min ≤ x ≤ x_max linearly distributed with three different sampling, depending on the values of frac_middle and rel_frac_begin.

If rel_frac_begin == nothing, defining the relative size of the segment x_min ≤ x ≤ x_start compared to the "masked one" x_min ≤ x ≤ x_start || x_stop ≤ x ≤ x_max as rel_prop_begin = (x_start - x_min) / (x_max - x_min - x_stop + x_start):

in x_min ≤ x ≤ x_start there are around (1.0 - frac_middle) * rel_prop_begin * N points;

in x_start ≤ x ≤ x_stop there are around frac_middle * N points;

in x_stop ≤ x ≤ x_max there are around (1.0 - frac_middle) * (1.0 - rel_prop_begin) * N points.

If rel_frac_begin is instead a float inside the interval 0.0 < rel_frac_begin < 1.0:

in x_min ≤ x ≤ x_start there are around (1.0 - frac_middle) * rel_frac_begin * N points;

in x_start ≤ x ≤ x_stop there are around frac_middle * N points;

in x_stop ≤ x ≤ x_max there are around (1.0 - frac_middle) * (1.0 - rel_frac_begin) * N points.

frac_middle is then the fraction of the N points that is inside the MIDDLE INTERVAL, while rel_frac_begin is the one inside the LEFT INTERVAL COMPARED TO THE MASKED TOTAL ONE. If ass::Bool is set to false the assert checks on the input data will not be performed.