The implications on the local Primordial Non-Gaussianities

GaPSE.TF — Type

TF(
    left_value::Float64
    left::Float64

    spline::Dierckx.Spline1D

    r_si::Float64
    r_b::Float64
    r_a::Float64
    right::Float64
    )

Contains all the information useful in order to return the Transfer Function value from:

a spline inside the interval left ≤ x ≤ right

the associated power law for x > right (with "right" coefficients r_si, r_b and r_a)

the associated constant left value left_value for x < left

Arguments

left_value::Float64 : the constant value that must be returned in case x < left.

left::Float64 : the break between the left power-law (for x <left) and the spline (for x ≥ left); its value is the xs[begin] one.

spline::Dierckx.Spline1D : spline that interpolates between the real values of the integral calculated inside the range left ≤ x ≤ right

right::Float64 : the break between the right power-law (for x ≥ left) and the spline (for x ≤ right); its value is the xs[end] one.

r_si, r_b, r_a :: Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the RIGHT edge; when an input value x > right is given, the returned one is obtained from power_law with this coefficients ( where, of course, r_si is the exponent, r_bthe coefficient andr_athe spurious adding constant). NOTE: for numerical issues, the "pure" power-lawy = f(x) = b + x^s` should be used.

Constructors

TF(ks, Tks) : from a set of (ks, Tks) pairs, take the mean of the first 10 as left value and fit with power_law_from_data the last 15.

See also: power_law_from_data

GaPSE.α_bias — Function

α_bias(k, tf::TF; bf=1.0, D=1.0, Ω_M0=0.29992)

Return the coefficient $\alpha_{\rm bias}$ that relates the Non-Gaussian density fluctiations $\delta_{\rm NG}$ and the Non-Gaussian gravitational potential $\Phi_{\rm NG}$ in Fourier space:

\[\delta_{\rm NG}(k) = \alpha(k, z) \, \Phi_{\rm NG}(k) \; \; , \quad \alpha(k, z) = \frac{2}{3} \frac{k^2 T_m(k) D(z)}{\Omega_{\mathrm{M}0}} \left(\frac{c}{H_0}\right)^2 \; \; , \quad \; \alpha_{\rm bias} = \frac{b_{\phi} f_{\rm NL}}{\alpha(k, z)} \; .\]

bf=1.0 : value of the degenerate product $b_{\phi} f_{\rm NL}$.

D = 1.0 : value of the linear growth factor $D$ at present day; inside this function, it is multiplied for a constant $q = 0.779017$ in order to normalize $D(z)$ to 1/(1+z) in matter domination, i.e. such that $q \, D(z) \, (1+z) = 1$ at $z=20$

See also: TF

GaPSE.IntegralIPSalpha — Type

IntegralIPSalpha(
    l_si::Float64
    l_b::Float64
    l_a::Float64
    left::Float64

    spline::Dierckx.Spline1D

    r_si::Float64
    r_b::Float64
    r_a::Float64
    right::Float64
)

Contains all the information useful in order to return the value of the integral of the Input Power Spectrum weighted with the α_bias function. In other words, return this expression:

\[\int_0^\infty \frac{\mathrm{d} q}{2 \pi^2} \, q^2 \, \frac{j_\ell(qs)}{(qs)^n} \, P(q) \, \alpha_{\mathrm{bias}} \; ,\]

where $P(q)$ is the Input Power Spectrum and

\[\delta_{\rm NG}(k) = \alpha(k, z) \, \Phi_{\rm NG}(k) \; \; , \quad \alpha(k, z) = \frac{2}{3} \frac{k^2 T_m(k) D(z)}{\Omega_{\mathrm{M}0}} \left(\frac{c}{H_0}\right)^2 \; \; , \quad \; \alpha_{\rm bias} = \frac{b_{\phi} f_{\rm NL}}{\alpha(k, z)} \; .\]

Arguments

l_si, l_b, l_a ::Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the LEFT edge; when an input value x < left is given, the returned one is obtained from power_law with this coefficients ( where, of course, l_si is the exponent, l_b the coefficient and l_a the spurious adding constant).

left::Float64 : the break between the left power-law (for x < left) and the spline (for x ≥ left); its value is the fit_min of the used constructor.

spline::Dierckx.Spline1D : spline that interpolates between the real values of the integral calculated inside the range left ≤ x ≤ right

right::Float64 : the break between the right power-law (for x > right) and the spline (for x ≤ right); its value is the fit_max of the used constructor.

r_si, r_b, r_a ::Float64 : coefficient for the spurious power-law $y = f(x) = a + b \, x^s$ for the RIGHT edge; when an input value x > right is given, the returned one is obtained from power_law with this coefficients ( where, of course, r_si is the exponent, r_b the coefficient and r_a the spurious adding constant). NOTE: for numerical issues, only the "pure" power-law $y = f(x) = b + x^s$ can be used. In other words, it always set r_a = 0.0.

Constructors

IntegralIPSalpha(tf::TF, cosmo::Cosmology, l, n=0; D=nothing, bf=1.0,
    N::Int=1024, kmin=1e-6, kmax=1e4, s0=1e-4,
    fit_left_min=nothing, fit_left_max=nothing, p0_left=nothing,
    fit_right_min=nothing, fit_right_max=nothing, p0_right=nothing)

The integral obtained with this constructor is calculated through xicalc, and expanded with power-laws at the edges.

tf::TF: the struct that contains all the data concerning the Transfer Function.

cosmo::Cosmology : cosmology to be used in this computation

l : degree of the spherical Bessel function to be used.

n=0 : degree of the exponent for the denominator. The interesting case is only the default value $0$.

D = nothing : value of the linear growth factor $D$ to be used. If nothing, it will be internally set as $D(z_{\mathrm{eff}})$, where $z_{\mathrm{eff}}$ is the effective redshift for the input cosmology.

bf = 1.0 : value of the degenerate product $b_{\phi} f_{\rm NL}$.

kmin = 1e-6, kmax = 1e4, s0 = 1e-4 : values to be passed to xicalc for the integration

fit_left_min = 2.0, fit_left_max = 10.0 : the limits (min and max) where the integral must be fitted with a power law, for small distances. This operation is necessary, because xicalc, in this context, gives wrong results for too small input distance s; nevertheless, this integral has fixed power-law trends for $s \rightarrow 0$, so this approach gives good results.

p0_left = nothing : vector with the initial values for the left power-law fitting; its length must be 2 (if you want to fit with a pure power-law $y = f(x) = b x^s$, so only l_si and l_b are matter of concern) or 3 (if you want to fit with a spurious power-law $y = f(x) = a + b x^s$, so you are also interested in l_a); if nothing, it will be automatically set p0 = [-1.0, 1.0].

fit_right_min = nothing, fit_right_max = nothing : the limits (min and max) where the integral must be fitted with a power law, for high distances. This integral has fixed power-law trends for $s \rightarrow \infty$, so this approach gives good results. If nothing, the last 15 points returned from xicalc are used for this fitting. NOTE: for numerical issues, only the "pure" power-law $y = f(x) = b + x^s$ can be used.

p0_right = nothing : vector with the initial values for the left power-law fitting; its length must be 2 (to fit with a pure power-law $y = f(x) = b x^s$, so only r_si and r_b are matter of concern); if nothing, it will be automatically set p0 = [-4.0, 1.0].

All the power-law fitting (both "pure" and spurious) are made through the local function power_law_from_data.

See also: power_law_from_data, power_law, Cosmology, α_bias

GaPSE.CosmoPNGParams — Type

CosmoParams(
        D::Float64 
        bf::Float64

        flm_0::Float64 
        flM_0::Float64 
        kmin_0::Float64 
        kmax_0::Float64  
        N_0::Int

        flm_2::Float64  
        flM_2::Float64 
        kmin_2::Float64 
        kmax_2::Float64  
        N_2::Int
    )

Struct that contains all the parameters and options that are matter of concerns for the CosmoPNG we are interested in.

Arguments

D : linear growth factor D to be used for the α_bias function.

bf : value of the degenerate product $b_{\phi} f_{\rm NL}$.

kmin_0, kmax_0, s00: values to be passed toxicalcfor the integration made byIntegralIPSalphaof the term`J0``

flm_0, flM_0 : the limits (min and max) where the integral made by IntegralIPSalpha of the term $J_0$ must be fitted with a power law, for small distances. This operation is necessary, because xicalc, in this context, gives wrong results for too small input distance s; nevertheless, this integral has fixed power-law trends for $s \rightarrow 0$, so this approach gives good results.

kmin_2, kmax_2, s02,flm2,flM2: same as the previous terms, but for the integral`J2``

Constructors

function CosmoPNGParams(D; 
    bf = 1.0,
    flm_0 = 5e-2, flM_0 = 1e-1, s0_0 = 1e-4,
    kmin_0 = 1e-6, kmax_0 = 1e4, N_0::Int = 1024,
    flm_2 = 5e-1, flM_2 = 1e0, s0_2 = 1e-4,
    kmin_2 = 1e-6, kmax_2 = 1e4, N_2::Int = 1024,
    )

The associations are trivials. The only thing to be put attention on is that D is a MANDATORY argument, while all the other ones are keyword arguments with a default value. You should use:

pngparams = CosmoPNGParams(cosmo.D_of_s(cosmo.s_eff); ...)

where cosmo::Cosmology is the Cosmology you are interested in and s_eff is the effective comoving distance (stored on cosmo).

See also: Cosmology, CosmoPNG, IntegralIPSalpha, α_bias

GaPSE.CosmoPNG — Type

CosmoPNG(
    params::CosmoPNGParams
    tf::TF
    file_TF::String

    J0::IntegralIPSalpha
    J2::IntegralIPSalpha
    )

Struct that contains all the information that may be used for the Correlation Function computations of the Primordial Non-Gaussianities (PNG) signal.

Arguments

params::CosmoPNGParams : parameters to be used for this Cosmology. See the docstring of CosmoParams for more information on the possible inputs.

tf::TF : transfer function to be used.

file_TF::String : name of the file where the transfer function was read.

J0 and J2::IntegralIPSalpha : integrals with the following form:

\[J_\ell = \int_0^\infty \frac{\mathrm{d} q}{2 \pi^2} \, q^2 \, j_\ell(qs) \, P(q) \, \alpha_{\mathrm{bias}} \; ,\]

where $P(q)$ is the Input Power Spectrum and

\[\delta_{\rm NG}(k) = \alpha(k, z) \, \Phi_{\rm NG}(k) \; \; , \quad \alpha(k, z) = \frac{2}{3} \frac{k^2 T_m(k) D(z)}{\Omega_{\mathrm{M}0}} \left(\frac{c}{H_0}\right)^2 \; \; , \quad \; \alpha_{\rm bias} = \frac{b_{\phi} f_{\rm NL}}{\alpha(k, z)} \; .\]

Constructor

CosmoPNG(
    pngparams::CosmoPNGParams,
    cosmo::Cosmology, file_TF::String;
    comments::Bool=true
)

pngparams::CosmoParams : parameters to be used for this Cosmology. See the docstring of CosmoParams for more information on the possible inputs.

cosmo::Cosmology : cosmology to be considered, both in terms od Input Power Spectrum and of cosmological parameters.

file_TF::String : name of the file where the Transfer Function to be used is stored.

comments::Bool=true : the file_TF file contains comments at the beginning?

See also: TF, IntegralIPSalpha, Cosmology

GaPSE.ξ_S_L0 — Function

ξ_S_L0(P::Point, cosmo::Cosmology, cosmopng::CosmoPNG; b=nothing, sp::Int64=1)
ξ_S_L0(s1, cosmo::Cosmology, cosmopng::CosmoPNG; kwargs...)

Return the value of the Two-Point Correlation Function (TPCF) monopole of the signal (S) of the local Primordial Non-Gaussianities (PNG) (for the given cosmo::Cosmology and cosmopng::CosmoPNG). In the first method, you should pass the Point where to evaluate that function, while in the second (that internally recalls the first) you must provide the comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\xi^{\mathrm{S}}_0(s) = 2 \left( b + \frac{1}{3}f(z_{\mathrm{eff}})\right) \, D^2(z_{\mathrm{eff}}) \, J_0(s)\]

where:

$b$ is the galaxy bias (stored in cosmo)

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$s_{\mathrm{eff}}$ is the effective comoving distance stored in cosmo (and $z_{\mathrm{eff}}$ its associated effective redshift in that cosmology)

$D$ the linear growth factor and $f$ the linear growth rate (whose splines are stored in cosmo)

$J_\ell$ (stored in cosmopng) is defined as
\[J_\ell(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, j_\ell(qs) \, \alpha_{\mathrm{bias}}(q,z)\]
with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo), $j_\ell$ as spherical Bessel function of order $\ell$ and
\[\alpha_{\rm bias}(q,z) = \frac{b_{\phi} f_{\rm NL}}{\alpha(q, z)} \quad , \quad \alpha(q, z) = \frac{2}{3} \frac{q^2 T_m(q) D(z)}{\Omega_{\mathrm{M}0}} \left(\frac{c}{H_0}\right)^2\]
with $b_{\phi}f_{\rm NL}$ is stored in cosmopng. Check the documentation of α_bias and CosmoPNG for more information.

Keyword arguments

b = nothing : galaxy bias value to be considered in the computation; if set to nothing, its value is inferred from the other keyword argument sp;

sp::Int64=1 : galaxy type to be considered between the two species stored in cosmo: it can be either 1 or 2, and the corresponding set of galaxy, magnification and evolutionary biases stored in cosmo are considered; in case b is set to a value, this parameter is ignored.

See also: Point, Cosmology, CosmoPNG, α_bias, ξ_S_L2, ξ_S, integrand_ξ_S_multipole, ξ_S_multipole map_ξ_S_multipole, print_map_ξ_S_multipole

GaPSE.ξ_S_L2 — Function

ξ_S_L2(P::Point, cosmo::Cosmology, cosmopng::CosmoPNG; b=nothing, sp::Int64=1)
ξ_S_L2(s1, cosmo::Cosmology, cosmopng::CosmoPNG; kwargs...)

Return the value of the Two-Point Correlation Function (TPCF) quadrupole of the signal (S) of the local Primordial Non-Gaussianities (PNG) (for the given cosmo::Cosmology and cosmopng::CosmoPNG). In the first method, you should pass the Point where to evaluate that function, while in the second (that internally recalls the first) you must provide the comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\xi^{\mathrm{S}}_2(s) = - \frac{4}{3} \, f(z_{\mathrm{eff}}) \, D^2(z_{\mathrm{eff}}) \, J_2(s)\]

where:

$b$ is the galaxy bias (stored in cosmo)

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$s_{\mathrm{eff}}$ is the effective comoving distance stored in cosmo (and $z_{\mathrm{eff}}$ its associated effective redshift in that cosmology)

$D$ the linear growth factor and $f$ the linear growth rate (whose splines are stored in cosmo)

$J_\ell$ (stored in cosmopng) is defined as
\[J_\ell(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, j_\ell(qs) \, \alpha_{\mathrm{bias}}(q,z)\]
with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo), $j_\ell$ as spherical Bessel function of order $\ell$ and
\[\alpha_{\rm bias}(q,z) = \frac{b_{\phi} f_{\rm NL}}{\alpha(q, z)} \quad , \quad \alpha(q, z) = \frac{2}{3} \frac{q^2 T_m(q) D(z)}{\Omega_{\mathrm{M}0}} \left(\frac{c}{H_0}\right)^2\]
with $b_{\phi}f_{\rm NL}$ is stored in cosmopng. Check the documentation of α_bias and CosmoPNG for more information.

Keyword arguments

b = nothing : galaxy bias value to be considered in the computation; if set to nothing, its value is inferred from the other keyword argument sp;

sp::Int64=1 : galaxy type to be considered between the two species stored in cosmo: it can be either 1 or 2, and the corresponding set of galaxy, magnification and evolutionary biases stored in cosmo are considered; in case b is set to a value, this parameter is ignored.

See also: Point, Cosmology, CosmoPNG, α_bias, ξ_S_L0, ξ_S, integrand_ξ_S_multipole, ξ_S_multipole map_ξ_S_multipole, print_map_ξ_S_multipole

GaPSE.ξ_S — Function

ξ_S(s, μ, cosmo::Cosmology, cosmopng::CosmoPNG; b=nothing, sp::Int64=1)

Return the value of the Two-Point Correlation Function (TPCF) of the signal (S) of the local Primordial Non-Gaussianities (PNG) in the given comoving distance s and cosine value for the Legendre polynomials μ (for the given cosmo::Cosmology and cosmopng::CosmoPNG). We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\begin{split} \xi^{\mathrm{S}}(s&,\mu) = \xi^{\mathrm{S}}_0(s) + \xi^{\mathrm{S}}_2(s) \mathcal{L}_2(\mu) \\ &\xi^{\mathrm{S}}_0(s) = 2 \left( b + \frac{1}{3}f(z_{\mathrm{eff}})\right) \, D^2(z_{\mathrm{eff}}) \, J_0(s) \\ &\xi^{\mathrm{S}}_2(s) = - \frac{4}{3} \, f(z_{\mathrm{eff}}) \, D^2(z_{\mathrm{eff}}) \, J_2(s) \end{split}\]

where:

$b$ is the galaxy bias (stored in cosmo)

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$s_{\mathrm{eff}}$ is the effective comoving distance stored in cosmo (and $z_{\mathrm{eff}}$ its associated effective redshift in that cosmology)

$D$ the linear growth factor and $f$ the linear growth rate (whose splines are stored in cosmo)

$\mathcal{L}_\ell$ the Legendre polynomial of order $\ell$

$J_\ell$ (stored in cosmopng) is defined as
\[J_\ell(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, j_\ell(qs) \, \alpha_{\mathrm{bias}}(q,z)\]
with $P(q)$ as the matter Power Spectrum at $z=0$ (stored in cosmo), $j_\ell$ as spherical Bessel function of order $\ell$ and
\[\alpha_{\rm bias}(q,z) = \frac{b_{\phi} f_{\rm NL}}{\alpha(q, z)} \quad , \quad \alpha(q, z) = \frac{2}{3} \frac{q^2 T_m(q) D(z)}{\Omega_{\mathrm{M}0}} \left(\frac{c}{H_0}\right)^2\]
with $b_{\phi}f_{\rm NL}$ is stored in cosmopng. Check the documentation of α_bias and CosmoPNG for more information.

Keyword arguments

b = nothing : galaxy bias value to be considered in the computation; if set to nothing, its value is inferred from the other keyword argument sp;

sp::Int64=1 : galaxy type to be considered between the two species stored in cosmo: it can be either 1 or 2, and the corresponding set of galaxy, magnification and evolutionary biases stored in cosmo are considered; in case b is set to a value, this parameter is ignored.

See also: Point, Cosmology, CosmoPNG, α_bias, ξ_S_L0, ξ_S_L2, integrand_ξ_S_multipole, ξ_S_multipole map_ξ_S_multipole, print_map_ξ_S_multipole

GaPSE.integrand_ξ_S_multipole — Function

integrand_ξ_S_multipole(s, μ, cosmo::Cosmology, cosmopng::CosmoPNG;
    L::Int=0, use_windows::Bool=true, b=nothing, sp::Int64=1 )::Float64

Return the integrand on $\mu = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}$ of the Two-Point Correlation Function (TPCF) concerning the signal (S) of the local Primordial Non-Gaussianities (PNG), i.e. the following function $f(s, \mu)$:

\[ f_L(s, \mu) = \xi^{\mathrm{S}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{S}}$ is the TPCF of the PNG signal, computed from ξ_S.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

Inputs

s: the comoving distance where must be evaluated the integral

μ: the cosine between s1 and s where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

cosmopng::CosmoPNG: struct that contains all the information that may be used for the TPCF computations of the PNG signal.

Keyword arguments

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

b = nothing : galaxy bias value to be considered in the computation; if set to nothing, its value is inferred from the other keyword argument sp;

sp::Int64=1 : galaxy type to be considered between the two species stored in cosmo: it can be either 1 or 2, and the corresponding set of galaxy, magnification and evolutionary biases stored in cosmo are considered; in case b is set to a value, this parameter is ignored.

See also:ξ_S, ξ_S_multipole, map_ξ_S_multipole, print_map_ξ_S_multipole WindowFIntegrated, Cosmology, CosmoPNG,

GaPSE.ξ_S_multipole — Function

ξ_S_multipole(
    s, cosmo::Cosmology, cosmopng::CosmoPNG;;
    L::Int = 0, use_windows::Bool = true,
    atol_quad::Float64 = 0.0,
    rtol_quad::Float64 = 1e-2
    enhancer::Float64 = 1e6,
    b=nothing, sp::Int64=1 ) ::Float64

Evaluate the multipole of order L of the Two-Point Correlation Function (TPCF) concerning the signal (S) of the local Primordial Non-Gaussianities (PNG), i.e. the following function $\xi^{\mathrm{S}} (s)$:

\[ \xi^{\mathrm{S}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{S}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{S}}(s,\mu)$ is the TPCF of the PNG signal with the angular dependence, computed from ξ_S.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

s: the comoving distance where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

cosmopng::CosmoPNG: struct that contains all the information that may be used for the TPCF computations of the PNG signal.

Keyword arguments

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

b = nothing : galaxy bias value to be considered in the computation; if set to nothing, its value is inferred from the other keyword argument sp;

sp::Int64=1 : galaxy type to be considered between the two species stored in cosmo: it can be either 1 or 2, and the corresponding set of galaxy, magnification and evolutionary biases stored in cosmo are considered; in case b is set to a value, this parameter is ignored.

See also: ξ_S, integrand_ξ_S_multipole, map_ξ_S_multipole, print_map_ξ_S_multipole WindowFIntegrated, Cosmology, CosmoPNG,

GaPSE.map_ξ_S_multipole — Function

map_ξ_S_multipole(
    cosmo::Cosmology, cosmopng::CosmoPNG,
    ss = nothing;
    L::Int = 0, use_windows::Bool = true,
    atol_quad::Float64 = 0.0,
    rtol_quad::Float64 = 1e-2,
    enhancer::Float64 = 1e6,
    pr::Bool = true,
    N_log::Int = 1000,
    kwargs...) ::Tuple{Vector{Float64}, Vector{Float64}}

Evaluate the multipole of order L of the Two-Point Correlation Function (TPCF) concerning the signal (S) of the local Primordial Non-Gaussianities (PNG), for all the comoving distance values stored inside ss. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

The function evaluated is then the following $\xi^{\mathrm{S}} (s)$:

\[ \xi^{\mathrm{S}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{S}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{S}}(s,\mu)$ is the TPCF of the PNG signal with the angular dependence, computed from ξ_S.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

cosmo::Cosmology: cosmology to be used in this computation

cosmopng::CosmoPNG: struct that contains all the information that may be used for the TPCF computations of the PNG signal.

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Keyword arguments

This function recall internally ξ_S_multipole, so the kwargs of the latter are valid also for the former; we report them for comfortness:

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

b = nothing : galaxy bias value to be considered in the computation; if set to nothing, its value is inferred from the other keyword argument sp;

sp::Int64=1 : galaxy type to be considered between the two species stored in cosmo: it can be either 1 or 2, and the corresponding set of galaxy, magnification and evolutionary biases stored in cosmo are considered; in case b is set to a value, this parameter is ignored.

Returns

A Tuple{Vector{Float64}, Vector{Float64}}, which has as first element the ss vector and as second one the corresponding ξ value evaluated.

See also: ξ_S, integrand_ξ_S_multipole, ξ_S_multipole, print_map_ξ_S_multipole WindowFIntegrated, Cosmology, CosmoPNG,

GaPSE.print_map_ξ_S_multipole — Function

print_map_ξ_S_multipole(
    cosmo::Cosmology, cosmopng::CosmoPNG, 
    out::String, ss = nothing;
    L::Int = 0, use_windows::Bool = true,
    atol_quad::Float64 = 0.0,
    rtol_quad::Float64 = 1e-2,
    enhancer::Float64 = 1e6,
    pr::Bool = true,
    N_log::Int = 1000,
    kwargs...)

Evaluate the multipole of order L of the Two-Point Correlation Function (TPCF) concerning the signal (S) of the local Primordial Non-Gaussianities (PNG), for all the comoving distance values stored inside ss, and print the results (with all the options used) in a file named out. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

The function evaluated is then the following $\xi^{\mathrm{S}} (s)$:

\[ \xi^{\mathrm{S}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{S}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{S}}(s,\mu)$ is the TPCF of the PNG signal with the angular dependence, computed from ξ_S.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

cosmo::Cosmology: cosmology to be used in this computation

cosmopng::CosmoPNG: struct that contains all the information that may be used for the TPCF computations of the PNG signal.

out::String : name of the file where the results must be stored.

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Keyword arguments

This function recall internally map_ξ_S_multipole, so the kwargs of the latter are valid also for the former; we report them for comfortness:

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

b = nothing : galaxy bias value to be considered in the computation; if set to nothing, its value is inferred from the other keyword argument sp;

sp::Int64=1 : galaxy type to be considered between the two species stored in cosmo: it can be either 1 or 2, and the corresponding set of galaxy, magnification and evolutionary biases stored in cosmo are considered; in case b is set to a value, this parameter is ignored.

See also: ξ_S, integrand_ξ_S_multipole, ξ_S_multipole, map_ξ_S_multipole WindowFIntegrated, Cosmology, CosmoPNG