The GNC Doppler TPCF in the Plane Parallel Approximation

GaPSE.ξ_PPDoppler_L0 — Function

 ξ_PPDoppler_L0(P::Point, cosmo::Cosmology)
 ξ_PPDoppler_L0(s, cosmo::Cosmology)

Return the value of the Two-Point Correlation Function (TPCF) monopole of the Doppler effect in the Plane-Parallel approximation. In the first method, you should pass the Point where to evaluate that function, while in the second (that internally recalls the first) you must provide the comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\xi^{\mathrm{pp}, v_{\parallel}}_0(s) = \frac{1}{3} \, f^2(s_{\mathrm{eff}}) \, \mathscr{H}^2(s_{\mathrm{eff}}) \, \mathscr{R}^2_{\mathrm{LD}}(s_{\mathrm{eff}}) \, D^2(s_{\mathrm{eff}}) \, s^2 \, I_0^2(s)\]

where:

$s_{\mathrm{eff}}$ is the effective comoving distance stored in cosmo::Cosmology

$D$ the linear growth factor and $f$ the linear growth rate

$\mathscr{H}(z)$ is the comoving Hubble parameter

$\mathscr{R}_{\mathrm{LD}}$ is defined as : $\mathscr{R}_{\mathrm{LD}}(z)= 1 - \frac{1}{\mathscr{H}(z) \, s(z)}$

$I_\ell^n$ is defined as
\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]
with $P(q)$ as the matter Power Spectrum at $z=0$ and $j_\ell$ as spherical Bessel function of order $\ell$.

All the cosmological data needed for this computation are taken from the input struct cosmo::Cosmology.

See also: Point, Cosmology, func_ℛ_LD, ξ_PPDoppler_L2, ξ_PPDoppler, integrand_ξ_PPDoppler_multipole, ξ_PPDoppler_multipole map_ξ_PPDoppler_multipole, print_map_ξ_PPDoppler_multipole

GaPSE.ξ_PPDoppler_L2 — Function

ξ_PPDoppler_L0(P::Point, cosmo::Cosmology)
ξ_PPDoppler_L0(s, cosmo::Cosmology)

Return the value of the Two-Point Correlation Function (TPCF) quadrupole of the Doppler effect in the Plane-Parallel approximation. In the first method, you should pass the Point where to evaluate that function, while in the second (that internally recalls the first) you must provide the comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\xi^{\mathrm{pp}, v_{\parallel}}_2(s) = -\frac{2}{3} \, f^2(z_{\mathrm{eff}}) \, \mathscr{H}^2(z_{\mathrm{eff}}) \, \mathscr{R}^2_{\mathrm{LD}}(z_{\mathrm{eff}}) \, D^2(z_{\mathrm{eff}}) \, s^2 \, I_2^2(s)\]

where:

$z_{\mathrm{eff}}$ is the effective redshift stored in cosmo::Cosmology ( and $z_{\mathrm{eff}}$ is the effective comoving distance associated in this cosmology)

$D$ the linear growth factor and $f$ the linear growth rate

$\mathscr{H}(z)$ is the comoving Hubble parameter

$\mathscr{R}_{\mathrm{LD}}(z)$ is defined as : $\mathscr{R}_{\mathrm{LD}}(z)= 1 - \frac{1}{\mathscr{H}(z) \, s(z)}$

$I_\ell^n$ is defined as
\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]
with $P(q)$ as the matter Power Spectrum at $z=0$ and $j_\ell$ as spherical Bessel function of order $\ell$.

All the cosmological data needed for this computation are taken from the input struct cosmo::Cosmology.

See also: Point, Cosmology, func_ℛ_LD, ξ_PPDoppler_L0, ξ_PPDoppler, integrand_ξ_PPDoppler_multipole, ξ_PPDoppler_multipole map_ξ_PPDoppler_multipole, print_map_ξ_PPDoppler_multipole

GaPSE.ξ_PPDoppler — Function

 ξ_PPDoppler(s, μ, cosmo::Cosmology)

Return the value of the Two-Point Correlation Function (TPCF) of the Doppler effect in the Plane-Parallel approximation in the given comoving distance s and cosine value for the Legendre polynomials μ (for the given cosmo::Cosmology).

The analytical expression of such TPCF is the following:

\[\begin{split} \xi^{\mathrm{pp, v_{\parallel}}}(s,&z,\mu) = \xi^{\mathrm{pp, v_{\parallel}}}_0(s,z) + \xi^{\mathrm{pp, v_{\parallel}}}_2(s) \mathcal{L}_2(\mu) \\ &\xi^{\mathrm{pp}, v_{\parallel}}_0(s) = \frac{1}{3} \, f^2(s_{\mathrm{eff}}) \, \mathscr{H}^2(s_{\mathrm{eff}}) \, \mathscr{R}^2_{\mathrm{LD}}(s_{\mathrm{eff}}) \, D^2(s_{\mathrm{eff}}) \, s^2 \, I_0^2(s) \\ &\xi^{\mathrm{pp}, v_{\parallel}}_2(s) = -\frac{2}{3} \, f^2(z_{\mathrm{eff}}) \, \mathscr{H}^2(z_{\mathrm{eff}}) \, \mathscr{R}^2_{\mathrm{LD}}(z_{\mathrm{eff}}) \, D^2(z_{\mathrm{eff}}) \, s^2 \, I_2^2(s) \\ \end{split}\]

where:

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$D$ the linear growth factor and $f$ the linear growth rate

$\mathcal{L}_\ell$ the Legendre polynomial of order $\ell$

$\mathscr{H}(z)$ is the comoving Hubble parameter

$\mathscr{R}_{\mathrm{LD}}(z)$ is defined as : $\mathscr{R}_{\mathrm{LD}}(z)= 1 - \frac{1}{\mathscr{H}(z) \, s(z)}$

$I_\ell^n$ is defined as
\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]
with $P(q)$ as the matter Power Spectrum at $z=0$ and $j_\ell$ as spherical Bessel function of order $\ell$.

All the cosmological data needed for this computation are taken from the input struct cosmo::Cosmology.

See also: Point, Cosmology, integrand_ξ_PPDoppler_multipole, ξ_PPDoppler_multipole map_ξ_PPDoppler_multipole, print_map_ξ_PPDoppler_multipole

GaPSE.integrand_ξ_PPDoppler_multipole — Function

 integrand_ξ_PPDoppler_multipole(s, μ, cosmo::Cosmology;
      L::Int=0, use_windows::Bool=true)

Return the integrand on $\mu = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}$ of the Doppler effect Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) approximation, i.e. the following function $f(s, \mu)$:

\[ f_L(s, \mu) = \xi^{\mathrm{pp}, v_{\parallel}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp}, v_{\parallel}}(s,\mu)$ is the TPCF of the Doppler term in the PP approximation with the angular dependence, computed from ξ_PPDoppler.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

Inputs

s: the comoving distance where must be evaluated the integral

μ: the cosine between s1 and s where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

See also:ξ_PPDoppler, ξ_PPDoppler_multipole, map_ξ_PPDoppler_multipole, print_map_ξ_PPDoppler_multipole WindowFIntegrated, Cosmology,

GaPSE.ξ_PPDoppler_multipole — Function

 ξ_PPDoppler_multipole(
      s, cosmo::Cosmology;
      L::Int = 0, use_windows::Bool = true,
      atol_quad::Float64 = 0.0,
      rtol_quad::Float64 = 1e-2,
      enhancer::Float64 = 1e6 ) ::Float64

Evaluate the multipole of order L of the Doppler effect Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) term i.e. the following function $\xi^{\mathrm{pp}, v_{\parallel}} (s)$:

\[ \xi^{\mathrm{pp}, v_{\parallel}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{pp}, v_{\parallel}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp}, v_{\parallel}}(s,\mu)$ is the TPCF of the Doppler term in the PP approximation with the angular dependence, computed from ξ_PPDoppler.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

s: the comoving distance where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

See also: ξ_PPDoppler, integrand_ξ_PPDoppler_multipole, map_ξ_PPDoppler_multipole, print_map_ξ_PPDoppler_multipole WindowFIntegrated, Cosmology,

GaPSE.map_ξ_PPDoppler_multipole — Function

 map_ξ_PPDoppler_multipole(
      cosmo::Cosmology, ss = nothing;
      L::Int = 0, use_windows::Bool = true,
      atol_quad::Float64 = 0.0,
      rtol_quad::Float64 = 1e-2,
      enhancer::Float64 = 1e6,
      pr::Bool = true,
      N_log::Int = 1000,
      kwargs...) ::Tuple{Vector{Float64}, Vector{Float64}}

Evaluate the multipole of order L of the Doppler effect Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) term for all the comoving distance values stored inside ss. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

The function evaluated is then the following $\xi^{\mathrm{pp, g}} (s)$:

\[ \xi^{\mathrm{pp}, v_{\parallel}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{pp}, v_{\parallel}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp}, v_{\parallel}}(s,\mu)$ is the TPCF of the Doppler term in the PP approximation with the angular dependence, computed from ξ_PPDoppler.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

cosmo::Cosmology: cosmology to be used in this computation

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Optional arguments

This function recall internally ξ_PPDoppler_multipole, so the kwargs of the latter are valid also for the former; we report them for comfortness:

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

Returns

A Tuple{Vector{Float64}, Vector{Float64}}, which has as first element the ss vector and as second one the corresponding ξ value evaluated.

See also: ξ_PPDoppler, integrand_ξ_PPDoppler_multipole, ξ_PPDoppler_multipole, print_map_ξ_PPDoppler_multipole WindowFIntegrated, Cosmology,

GaPSE.print_map_ξ_PPDoppler_multipole — Function

 print_map_ξ_PPDoppler_multipole(
      cosmo::Cosmology, out::String,
      ss = nothing;
      L::Int = 0, use_windows::Bool = true,
      atol_quad::Float64 = 0.0,
      rtol_quad::Float64 = 1e-2,
      enhancer::Float64 = 1e6,
      pr::Bool = true,
      N_log::Int = 1000,
      kwargs...)

Evaluate the multipole of order L of the Doppler effect Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) term for all the comoving distance values stored inside ss, and print the results (with all the options used) in a file named out. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

The function evaluated is then the following $\xi^{\mathrm{pp}, v_{\parallel}} (s)$:

\[ \xi^{\mathrm{pp}, v_{\parallel}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{pp}, v_{\parallel}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp}, v_{\parallel}}(s,\mu)$ is the TPCF of the Doppler term in the PP approximation with the angular dependence, computed from ξ_PPDoppler.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

cosmo::Cosmology: cosmology to be used in this computation

out::String : name of the file where the results must be stored.

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Optional arguments

This function recall internally map_ξ_PPDoppler_multipole, so the kwargs of the latter are valid also for the former; we report them for comfortness:

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

See also: ξ_PPDoppler, integrand_ξ_PPDoppler_multipole, ξ_PPDoppler_multipole, map_ξ_PPDoppler_multipole WindowFIntegrated, Cosmology,

The Galaxy TPCF term in the Plane Parallel Approximation

GaPSE.ξ_PPGalaxies_L0 — Function

 ξ_PPGalaxies_L0(P::Point, cosmo::Cosmology)
 ξ_PPGalaxies_L0(s, cosmo::Cosmology)

Return the value of the Two-Point Correlation Function (TPCF) monopole of the Galaxies in the Plane-Parallel approximation. In the first method, you should pass the Point where to evaluate that function, while in the second (that internally recalls the first) you must provide the comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\xi^{\mathrm{pp, g}}_0(s) = D^2(z) \, I_0^0(s)\, \left(b^2 + \frac{2}{3} \, b \, f(z) + \frac{1}{5} \, f^2(z)\right)\]

where:

$b$ is the galaxy bias

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$D$ the linear growth factor and $f$ the linear growth rate

$I_\ell^n$ is defined as
\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]
with $P(q)$ as the matter Power Spectrum at $z=0$ and $j_\ell$ as spherical Bessel function of order $\ell$.

All the cosmological data needed for this computation are taken from the input struct cosmo::Cosmology.

See also: Point, Cosmology, ξ_PPGalaxies_L2, ξ_PPGalaxies_L4, integrand_ξ_PPGalaxies_multipole, ξ_PPGalaxies_multipole map_ξ_PPGalaxies_multipole, print_map_ξ_PPGalaxies_multipole

GaPSE.ξ_PPGalaxies_L2 — Function

 ξ_PPGalaxies_L2(P::Point, cosmo::Cosmology)
 ξ_PPGalaxies_L2(s, cosmo::Cosmology)

Return the value of the Two-Point Correlation Function (TPCF) quadrupole of the Galaxies in the Plane-Parallel approximation. In the first method, you should pass the Point where to evaluate that function, while in the second (that internally recalls the first) you must provide the comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\xi^{\mathrm{pp, g}}_2(s) = - D^2(z) \, I_2^0(s)\, \left(\frac{4}{3} \, b \, f(z) + \frac{4}{7} \, f^2(z)\right)\]

where:

$b$ is the galaxy bias

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$D$ the linear growth factor and $f$ the linear growth rate

$I_\ell^n$ is defined as
\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]
with $P(q)$ as the matter Power Spectrum at $z=0$ and $j_\ell$ as spherical Bessel function of order $\ell$.

All the cosmological data needed for this computation are taken from the input struct cosmo::Cosmology.

See also: Point, Cosmology, ξ_PPGalaxies_L0, ξ_PPGalaxies_L4, integrand_ξ_PPGalaxies_multipole, ξ_PPGalaxies_multipole map_ξ_PPGalaxies_multipole, print_map_ξ_PPGalaxies_multipole

GaPSE.ξ_PPGalaxies_L4 — Function

 ξ_PPGalaxies_L4(P::Point, cosmo::Cosmology)
 ξ_PPGalaxies_L4(s, cosmo::Cosmology)

Return the value of the Two-Point Correlation Function (TPCF) hexadecapole of the Galaxies in the Plane-Parallel approximation. In the first method, you should pass the Point where to evaluate that function, while in the second (that internally recalls the first) you must provide the comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of such TPCF monopole is the following:

\[\xi^{\mathrm{pp, g}}_4(s,z) = D^2(z) \, I_4^0(s)\, \left(\frac{8}{35} \, f^2(z)\right)\]

where:

$b$ is the galaxy bias

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$D$ the linear growth factor and $f$ the linear growth rate

$I_\ell^n$ is defined as
\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]
with $P(q)$ as the matter Power Spectrum at $z=0$ and $j_\ell$ as spherical Bessel function of order $\ell$.

All the cosmological data needed for this computation are taken from the input struct cosmo::Cosmology.

See also: Point, Cosmology, ξ_PPGalaxies_L0, ξ_PPGalaxies_L2, integrand_ξ_PPGalaxies_multipole, ξ_PPGalaxies_multipole map_ξ_PPGalaxies_multipole, print_map_ξ_PPGalaxies_multipole

GaPSE.ξ_PPGalaxies — Function

 ξ_PPGalaxies(s, μ, cosmo::Cosmology)

Return the value of the Two-Point Correlation Function (TPCF) of the Galaxies in the Plane-Parallel approximation in the given comoving distance s and cosine value for the Legendre polynomials μ (for the given cosmo::Cosmology).

The analytical expression of such TPCF is the following:

\[\begin{split} \xi^{\mathrm{pp, g}}(s&,\mu) = \xi^{\mathrm{pp, g}}_0(s) + \xi^{\mathrm{pp, g}}_2(s) \mathcal{L}_2(\mu) + \xi^{\mathrm{pp, g}}_4(s) \mathcal{L}_4(\mu) \\ &\xi^{\mathrm{pp, g}}_0(s) = D^2(z) \, I_0^0(s)\, \left(b^2 + \frac{2}{3} \, b \, f(z) + \frac{1}{5} \, f^2(z)\right) \\ &\xi^{\mathrm{pp, g}}_2(s) = - D^2(z) \, I_2^0(s)\, \left(\frac{4}{3} \, b \, f(z) + \frac{4}{7} \, f^2(z)\right) \\ &\xi^{\mathrm{pp, g}}_4(s) = D^2(z) \, I_4^0(s)\, \left(\frac{8}{35} \, f^2(z)\right) \end{split}\]

where:

$b$ is the galaxy bias

$z$ is the redshift associated to the comoving distance $s$ in this cosmology

$D$ the linear growth factor and $f$ the linear growth rate

$\mathcal{L}_\ell$ the Legendre polynomial of order $\ell$

$I_\ell^n$ is defined as
\[I_\ell^n(s) = \int_0^{+\infty} \frac{\mathrm{d}q}{2\pi^2} \, q^2 \, P(q) \, \frac{j_\ell(qs)}{(qs)^n}\]
with $P(q)$ as the matter Power Spectrum at $z=0$ and $j_\ell$ as spherical Bessel function of order $\ell$.

All the cosmological data needed for this computation are taken from the input struct cosmo::Cosmology.

See also: Point, Cosmology, ξ_PPGalaxies_L0, ξ_PPGalaxies_L2, ξ_PPGalaxies_L4 integrand_ξ_PPGalaxies_multipole, ξ_PPGalaxies_multipole map_ξ_PPGalaxies_multipole, print_map_ξ_PPGalaxies_multipole

GaPSE.integrand_ξ_PPGalaxies_multipole — Function

 integrand_ξ_PPGalaxies_multipole(s, μ, cosmo::Cosmology;
      L::Int=0, use_windows::Bool=true)

Return the integrand on $\mu = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}$ of the Galaxies Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) approximation, i.e. the following function $f(s, \mu)$:

\[ f_L(s, \mu) = \xi^{\mathrm{pp, g}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp, g}}(s,\mu)$ is the TPCF of the Galaxies in the PP approximation with the angular dependence, computed from ξ_PPGalaxies.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

Inputs

s: the comoving distance where must be evaluated the integral

μ: the cosine between s1 and s where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

See also:ξ_PPGalaxies, ξ_PPGalaxies_multipole, map_ξ_PPGalaxies_multipole, print_map_ξ_PPGalaxies_multipole WindowFIntegrated, Cosmology,

GaPSE.ξ_PPGalaxies_multipole — Function

 ξ_PPGalaxies_multipole(
      s, cosmo::Cosmology;
      L::Int = 0, use_windows::Bool = true,
      atol_quad::Float64 = 0.0,
      rtol_quad::Float64 = 1e-2,
      enhancer::Float64 = 1e6 ) ::Float64

Evaluate the multipole of order L of the Galaxies Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) term i.e. the following function $\xi^{\mathrm{pp, g}} (s)$:

\[ \xi^{\mathrm{pp, g}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{pp, g}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp, g}}(s,\mu)$ is the TPCF of the Galaxies in the PP approximation with the angular dependence, computed from ξ_PPGalaxies.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

s: the comoving distance where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

See also: ξ_PPGalaxies, integrand_ξ_PPGalaxies_multipole, map_ξ_PPGalaxies_multipole, print_map_ξ_PPGalaxies_multipole WindowFIntegrated, Cosmology,

GaPSE.map_ξ_PPGalaxies_multipole — Function

 map_ξ_PPGalaxies_multipole(
      cosmo::Cosmology, ss = nothing;
      L::Int = 0, use_windows::Bool = true,
      atol_quad::Float64 = 0.0,
      rtol_quad::Float64 = 1e-2,
      enhancer::Float64 = 1e6,
      pr::Bool = true,
      N_log::Int = 1000,
      kwargs...) ::Tuple{Vector{Float64}, Vector{Float64}}

Evaluate the multipole of order L of the Galaxies Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) term for all the comoving distance values stored inside ss. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

The function evaluated is then the following $\xi^{\mathrm{pp, g}} (s)$:

\[ \xi^{\mathrm{pp, g}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{pp, g}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp, g}}(s,\mu)$ is the TPCF of the Galaxies in the PP approximation with the angular dependence, computed from ξ_PPGalaxies.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

cosmo::Cosmology: cosmology to be used in this computation

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Optional arguments

This function recall internally ξ_PPGalaxies_multipole, so the kwargs of the latter are valid also for the former; we report them for comfortness:

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

Returns

A Tuple{Vector{Float64}, Vector{Float64}}, which has as first element the ss vector and as second one the corresponding ξ value evaluated.

See also: ξ_PPGalaxies, integrand_ξ_PPGalaxies_multipole, ξ_PPGalaxies_multipole, print_map_ξ_PPGalaxies_multipole WindowFIntegrated, Cosmology,

GaPSE.print_map_ξ_PPGalaxies_multipole — Function

 print_map_ξ_PPGalaxies_multipole(
      cosmo::Cosmology, out::String,
      ss = nothing;
      L::Int = 0, use_windows::Bool = true,
      atol_quad::Float64 = 0.0,
      rtol_quad::Float64 = 1e-2,
      enhancer::Float64 = 1e6,
      pr::Bool = true,
      N_log::Int = 1000,
      kwargs...)

Evaluate the multipole of order L of the Galaxies Two-Point Correlation Function (TPCF) in the Plane Parallel (PP) term for all the comoving distance values stored inside ss, and print the results (with all the options used) in a file named out. If ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log).

The function evaluated is then the following $\xi^{\mathrm{pp, g}} (s)$:

\[ \xi^{\mathrm{pp, g}} (s) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi^{\mathrm{pp, g}} \left(s, \mu\right) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \frac{1}{\mathcal{N}}\mathcal{F}(s, \mu) \quad \mathrm{use\_windows == true} \\ 1 \quad\quad \mathrm{use\_windows == false} \end{cases}\]

where:

$\xi^{\mathrm{pp, g}}(s,\mu)$ is the TPCF of the Galaxies in the PP approximation with the angular dependence, computed from ξ_PPGalaxies.

$\mathcal{L}_L(\mu)$ is the Legendre polynomial of order $L$

$\mathcal{F}(s, \mu)$ is the integrated window function stored in cosmo::Cosmology (check the documentation of WindowFIntegrated)

$\mathcal{N}$ is the integrated window function norm (check the documentation of WindowFIntegrated)

The integration over $\mu$ is preformed through the Julia function quadgk from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

Inputs

cosmo::Cosmology: cosmology to be used in this computation

out::String : name of the file where the results must be stored.

ss : vector/range of s values where the function must be evaluated; if ss = nothing, it is set ss = 10 .^ range(0, log10(2 * cosmo.s_max), length=N_log). This is why it is returned also the vector of the "input" values.

Optional arguments

This function recall internally map_ξ_PPGalaxies_multipole, so the kwargs of the latter are valid also for the former; we report them for comfortness:

L::Int = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider $\mathcal{F}$ or not.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

N_log::Int = 1000 : number of points to be used in the default logaritmically-spaced range for ss, i.e. range(0, log10(2 * cosmo.s_max), length=N_log); it is ignored if ss ≠ nothing

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

See also: ξ_PPGalaxies, integrand_ξ_PPGalaxies_multipole, ξ_PPGalaxies_multipole, map_ξ_PPGalaxies_multipole WindowFIntegrated, Cosmology,