Power Spectrum for a generic window fuction

Multipoles Q_l1 of the toy-model window function

GaPSE.WindowFIntegrated_multipole — Function

WindowFIntegrated_multipole(
        s, windowfint::GaPSE.WindowFIntegrated;
        s_min, s_max,
        L::Int=0, alg::Symbol=:lobatto,
        N_lob::Int=100, N_trap::Int=200,
        atol_quad::Float64=0.0, rtol_quad::Float64=1e-2,
        enhancer::Float64=1e6,
        )

Evaluate the multipole of order L of the input Integrated Window Function windowfint in the input comoving distance s. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

The analytical expression of the Integrated Window Function multipole $Q_{\ell_1}$ is the following:

\[Q_{\ell_1}(s) = \int_{-1}^{1} \mathrm{d}\mu \; \mathcal{L}_{\ell_1}(\mu) \; \mathcal{F}(s, \mu)\]

where $\mathcal{L}_{\ell_1}$ is the Legendre polynomial of order $\ell_1$, $\mu$ the cosine angle,

\[\mathcal{F}(s, \mu) = \int_0^\infty \mathrm{d}s_1 \, \phi(s_1) \, \phi\left(\sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}\right) \, F\left(\frac{s}{s_1}, \mu \right)\]

the integrated window function associated to the window function $F\left(\frac{s}{s_1}, \mu \right)$ (check the docstring of WindowF for its definition) and $\phi$ the radial window function, obtained by ϕ.

Optional arguments

s_min and s_max (mandatory keyword arguments) : min and max comoving distance of the survey; their values will be internally used by

L::Int = 0: order of the Legendre polynomial to be used

alg::Symbol = :lobatto : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature.

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

See also: WindowFIntegrated, WindowF, ϕ

GaPSE.print_map_WindowFIntegrated_multipole — Function

print_map_WindowFIntegrated_multipole(
    ss::Vector{Float64},
    windowFint::Union{String,GaPSE.WindowFIntegrated}, out::String;
    s_min, s_max,
    pr::Bool=true, L_max::Int=4, alg::Symbol=:lobatto,
    N_lob::Int=100, N_trap::Int=200,
    atol_quad::Float64=0.0, rtol_quad::Float64=1e-2,
    enhancer::Float64=1e6)

print_map_WindowFIntegrated_multipole(
    s_zs::Vector{Float64},
    windowFint::Union{String,GaPSE.WindowFIntegrated}, out::String,
    file_data::String; z_min, z_max,
    names_bg=GaPSE.NAMES_BACKGROUND, h_0=0.7, kwargs...))

print_map_WindowFIntegrated_multipole(
    windowFint::Union{String,GaPSE.WindowFIntegrated}, out::String,
    file_data::String; z_min, z_max,
    names_bg=GaPSE.NAMES_BACKGROUND, h_0=0.7, N::Int=100, 
    m::Float64=2.1, st::Float64=0.0, kwargs...)

Evaluate the integrated window function multipoles $Q_{\ell_1}(s)$ in a vector of $s$ values for all the multipoles $0 \leq \ell_1 \leq L_\mathrm{max}$, and print the results in the out file. The computation of the multipole is performed through WindowFIntegrated_multipole.

The first method takes as input:

ss::Vector{Float64} : the vector of s points where to sample the integrated window function multipoles $Q_{\ell_1}$.ss must be a float vector of increasing comoving distance values (so each element must be ≥ 0); the first and last values ARE NOT RELATED to s_min and s_max.

windowFint::Union{String,WindowFIntegrated}, i.e. the integrated window function itself; it can be passed as the namefile where the integrated window is stored in (that will be opened with WindowFIntegrated) or as a WindowFIntegrated struct directly.

out::String : the name of the output file

s_min and s_max (keyword arguments) : min and max comoving distance of the survey; their values will be internally used by

The second method takes as input the min and max redshifts of the survey (z_minand z_max), the vector of redshifts zs::Vector{Float64} for the integrated window function sampling, windowFint as before and the file_data where can be found the association $z \rightarrow s(z)$. Such file must have the structure of the background data produced by the CLASS (link: https://github.com/lesgourg/class_public) code. Note that also zs musyt be a float vector of increasing redshift values (so each element must be ≥ 0). This method internally recalls the first one, so the other kwargs... are in common.

The third method takes as input the min and max redshifts of the survey (z_minand z_max) and the same input as the second method (windowF, out and file_data) but NOT THE REDSHIFT SAMPLING VECTOR zs. The sampling will be internally made linearly from $s = \mathrm{st}$(where st::Float64 = 0.0 is a keyword argument) to $s = m \, s_{\mathrm{max}}$, where s_max is the comoving distance associated to z_max (for the data stored in file_data) and m::Float64 = 2.1 a coefficient that we suggest to set equals to 2 < m < 3. N::Int = 100 is the number of s values used for the sampling in the interval $[0, m \, s_{\mathrm{max}}]$. This method internally recalls the first one, so the other kwargs... are in common.

The analytical expression for the integrated window function is the following:

\[Q_{\ell_1}(s) = \int_{-1}^{1} \mathrm{d}\mu \; \mathcal{L}_{\ell_1}(\mu) \; \mathcal{F}(s, \mu)\]

where $s$ is the comoving distance, $\mu$ the cosine angle, $\mathcal{L}_{\ell_1}$ the Legendre polynomial of order $\ell_1$ and $\mathcal{F}(x, μ)$ the integrated window function. Check the documentation of WindowFIntegrated for its definition. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Example

julia> windowfint = GaPSE.WindowFIntegrated(PATHTOGAPSE"data/IntegrFREFERENCEpi2z115.txt"); julia> GaPSE.printmapWindowFIntegratedmultipole(windowfint, "myQlmultipoles.txt", PATHTOGAPSE"data/WideAZAbackground.dat"; zmin = 1.0, zmax=1.5, st = 1.0, N=500, pr=false) julia> run(head -n 20 $(DIR*"my_Ql_multipoles.txt")) ###############

GaPSE

###############

This is an integration map of the Q{l1} multipoles, defined as:

Q{l1}(s1, s \mu) = \int{-1}^{+1} \mathrm{d}\mu \mathcal{L}{l1}(\mu) \mathcal{F}(s, \mu)

\mathcal{F}(s, \mu) = \int0^{\infty} \mathrm{d}s1 s1^2 \phi(s1) \phi(\sqrt(s1^2 + s^2 + 2 s1 s \mu)) F(s/s_1, \mu)

where \mathcal{L}{l1}(\mu) is tre Legendre polynomial if order l1 and

F(x, \mu) is the window function considered (for its analytical definition, check the code).

Time needed for this computation [in s]: 27.256186962127686

The keyword arguments were:

s [h0^{-1} Mpc] Q{l1=0} Q{l1=1} Q{l1=2} Q{l1=3} Q{l_1=4}

1.0 4.1857800000750543e11 -4.377435879373042e7 -5.084259164821501e8 1.2380785453994218e6 -3.641597411371149e8 13.852533751787348 4.1473071900503394e11 -6.063857839848524e8 -1.342493986435839e9 1.7150523594728626e7 -2.0225033264194965e8 ... ... ... ... ... ...

Optional arguments

As optional arguments of the first method:

s_min and s_max (mandatory keyword arguments) : min and max comoving distance of the survey; their values will be internally used by

pr::Bool = true : do you want to see the progress-bar of the computation?

L_max::Int64 = 4 : maximum multipole order to be computed

alg::Symbol = :lobatto : algorithm to be used for the integration; the valid options are (other values will lead to AssertionError):

:quad -> the integration over $\mu$ will be preformed through the Julia function quadgk

from the QuadGK.jl Julia package, that uses an adaptive Gauss-Kronrod quadrature.

:trap -> the integration over $\mu$ will be preformed through the Julia function trapz

from the Trapz.jl Julia package, that uses the simple trapezoidal rulae.

:lobatto -> the integration over $\mu$ will be preformed through the Julia function gausslobatto

from the FastGaussQuadrature.jl Julia package, that uses the Gauss-Lobatto quadrature.

N_lob::Int = 100 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

N_trap::Int = 200 : number of points to be used in the sampling made by the function trapz. Note that these options will have an effect only if you se alg = :quad.

atol_quad::Float64 = 0.0 and rtol_quad::Float64 = 1e-2: absolute and relative tolerance to be passed to the function quadgk; it's recommended not to set rtol_quad < 1e-2 because the time for evaluation increase quickly. Note that these options will have an effect only if you se alg = :quad.

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

The optional arguments given to the second method will be directly given to the first one. The only two exceptions are:

s_min and s_max (mandatory keyword arguments) : min and max redshift of the survey; their values will be internally coverted to comoving distances and passed to the first method

names = NAMES_BACKGROUND : the column names of the file_data. If the colum order change from the default one NAMES_BACKGROUND, you must set as input the vector of string with the correct one, with the SAME names. They are, with the default order:

["z", "proper time [Gyr]", "conf. time [Mpc]", "H [1/Mpc]", "comov. dist.", "ang.diam.dist.", "lum. dist.", "comov.snd.hrz.", "(.)rhog", "(.)rhob", "(.)rhocdm", "(.)rholambda", "(.)rhour", "(.)rhocrit", "gr.fac. D", "gr.fac. f"]

h = 0.7 : the adimensional hubble constant. By default, CLASS background data are measured with it numerically expressed (so distances are measured in Mpc, for example), while this code works with h in the unit of measure (so distances are measured in Mpc/h, for example). Change this value to 1.0 if the input data do not have this issue, or to your value of interest (0.67, 0.5, ...).

The optional arguments given to the third method will be directly given to the first one again. The only two exceptions are:

names_bg=GaPSE.NAMES_BACKGROUND and h_0=0.7 : same as for the second method

N::Int=100 : number of points to be used in the liearly spaced comoving distance vector

st::Float64=0.0 : starting comoving distance of the vector

m:Float64 = 2.1 : coefficient that set the maximum comoving distance of the vector, equals to $m * s_max$, where s_max is the comoving distance associated to the redhsift z_max

See also: WindowFIntegrated_multipole, WindowFIntegrated, WindowF, ϕ, BackgroundData

Creating a XiMultipole struct

GaPSE.XiMultipoles — Type

XiMultipoles(
    comdist::Vector{Float64}
    multipoles::Vector{Vector{Float64}}
)

Stores the multipoles of a generic Two-Point Correlation Function.

Constructors

XiMultipoles(file::String; comments::Bool=true, xdt::DataType=Float64, ydt::DataType=Float64)

Read the file file and create a XiMultipoles struct having

as first element the comoving distances stored in the first column (with the input type xdt)

as second element a vector that contains all the following columns (with the input type ydt), which are the multipoles L=0,1,2,... of the TPCF considered

If the file start with comments (lines starting with #), set comments = true.

GaPSE.create_file_for_XiMultipoles — Function

create_file_for_XiMultipoles(out::String, names::Vector{String}, 
    effect::Union{String, Integer}, group::String="GNC"; 
    comments::Bool=true, xdt::DataType=Float64, ydt::DataType=Float64)

Read the column number effect (if is an Integer) or the one corresponding to the GR effect effect for the input group group (if is a String) from all the filenames stored in the Vector names, and save them in a file named out.

The first column of out will be the same as the first column of the first filename in names; it is however checked internally if the first column of all the other files coincides with this one. The following columns of out follow the order in names. Note that effect, if passed as Integer, must be > 1 (because 1 is the index of the first column, used as x-axis).

group must be one among the following: ["LD", "GNC", "GNCxLD", "LDxGNC", "generic"] If group=generic, then effect must be an integer (because you are not selecting a specific effect in one of the native GaPSE groups).

xdt and ydt are the data types to be used for respectively the first column and the 2-3-4-... columns of out. Set comments=true if the files in names start with a header that must be skipped (its lines must start with #, otherwise they will not be recognised as comments).

Example

julia> run(cat file_1.txt)

Generic comment line

of the file_1.txt

1.0 0.999999 0.34545 0.00991 ... ... ... ...

julia> run(cat file_2.txt)

same, for file_2.txt

1.0 0.58244 0.12123 0.000154 ... ... ... ...

julia> createfileforXiMultipoles("mix.txt", ["file1.txt", "file_2.txt"], 3, "generic"); julia> run(cat mix.txt) ###############

GaPSE

###############

This is a table containing the multipoles of the Two-Point Correlation Function (TPCF)

for a generic group effect [not given, provied only the index 2] taken from the files:

- L = 0 : file_1.txt

- L = 1 : file_2.txt

s [Mpc/h0] xi{L=0} xi_{L=1}

1.0 0.34545 0.12123 ... ... ...

The Power Spectrum for a GenWin

GaPSE.GenericWindow — Type

GenericWindow(
    comdist::Vector{Float64}
    multipoles::Vector{Vector{Float64}}
    splines::Vector{Dierckx.Spline1D}
)

Stores the multipoles of a generic window function, computed as:

\[ Q_{\ell_1} = \int_0^{\infty} \mathrm{d}s_1 \, s_1^2 \, \phi(s_1) \, F_{\ell_1}(s_1,s)\]

with some FFT algorithm. See Eq. (2.13) of Castorina, Di Dio (2021) for more details.

Constructors

GenericWindow(file::String; comments::Bool=true, xdt::DataType=Float64, ydt::DataType=Float64)

Read the file file and create a GenericWindow struct having

as first element the comoving distances stored in the first column (with the input type xdt)

as second element a vector that contains all the following columns (with the input type ydt), which are the multipoles L=0,1,2,... of the Window FUnction considered

as second element a vector that contains all the splines of that multipoles

If the file start with comments (lines starting with #), set comments = true.

GaPSE.PS_multipole_GenWin — Function

PS_multipole_GenWin(
    ximult::Union{XiMultipoles,String}, genwin::Union{GenericWindow,String};
    alg::Symbol=:fftlog, L::Int=0,
    cut_first_n::Int=0, cut_last_n::Int=0,
    kwargs...)

Return the L-order Power Spectrum (PS) multipole for a generic window function, through the following Fast Fourier Transform and the effective redshift approximation:

\[ \left\langle \hat{P}_L(k) \right\rangle = \frac{2 L + 1}{A} (-i)^L \sum_{\ell=0}^{\infty} \sum_{\ell_1=0}^{\infty} \begin{pmatrix} L & \ell & \ell_1 \\ 0 & 0 & 0 \end{pmatrix}^2 \int_0^{\infty}\!\mathrm{d} s \, s^2 \, \xi_\ell(s, s_{\rm eff}) \, j_L(k s) \, Q_{\ell_1}(s) \, ,\]

where:

$\left\langle \hat{P}_L(k) \right\rangle$ is the order $L$ Power Spectrum of the effect we are interested in; we are basing this expression on the Yamamoto estimator (see Yamamoto (2000) and Yamamoto (2006))

$A$ is a normalization constant

the 2x3 matrix represents the Wigner-3j symbols

$\xi_\ell$ is the order $\ell$ multipole of the Two-Point Correlation Function (TPCF)

$j_L$ is the spherical Bessel function of order $L$

$s_{\mathrm{eff}}$ is the comoving distance associated with the effective redshift (see the TUTORIAL.ipynb notebook)

$Q_{\ell_1}$ can be easily estimated with FFT methods:

\[\begin{split} Q_{\ell_1}(s) &:= \int_0^{\infty}\mathrm{d}s_1 \; s_1^2 \; \phi(s_1) \; F_{\ell_1}(s_1, s) \\ F_{\ell_1} (s_1 , s) &:= \int_{4\pi} \mathrm{d}\Omega_{\mathbf{\hat{s}}} \, \int_{4\pi} \mathrm{d}\Omega_{\mathbf{\hat{s}}_1} \, \phi(s_2) \, W(\mathbf{\hat{s}}_1) \, W(\mathbf{\hat{s}}_2) \, \mathcal{L}_{\ell_1}(\mathbf{\hat{s}} \cdot \mathbf{\hat{s}}_1) \, . \end{split}\]

where:

$\mathcal{L}_{\ell_1}$ is the Legendre polynomial of order $\ell_1$

$\mathrm{d}\Omega_{\mathbf{\hat{s}}}$ is the infinitesimal solid angle pointing in the direction of the versor $\mathbf{\hat{s}}$

$\phi(s)$ and $W(\mathbf{\hat{s}})$ are respectively the radial and angular part of your window function (we remember that we assumed that such separability of the window function is possible)

Check Eq.(2.13) of Castorina, Di Dio for the theoretical explanation of this formula.

Currenlty, there are two algorithms you can choose in order to perform the computation; you can choose which one to use through the keyword value alg:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

IMPORTANT: no matter which algorithm you choose, you will need to give the input data in a LOGARITHMICALLY DISTRIBUTED scale. A linear distribution does not fit for the algorithms to apply.

Inputs

ximult::Union{XiMultipoles,String} : Two-Point Correlation functions to be used in the computation. You can provide either a XiMultipoles struct containing them or the String filename where they are stored (that will be internally open with XiMultipoles too). The file must have a structure analogous to the genwin one (see the next Inputs item). You can use create_file_for_XiMultipoles to produce such a file.

genwin::Union{GenericWindow,String} : multipoles $Q_{\ell_1}$ of the generic window function you want to consider. You can provide either a GenericWindow struct containing them or the String filename where they are stored (that will be internally open with GenericWindow too). The file must have the following structure

  $ cat Ql_multipoles.txt 
  # Any comment line in the file must start with a #
  # you can have how many comment lines you want in the header; they 
  # will be all skipped.
  # Then you must provide in blank space separated columns:
  # - as first column, the comoving distance values, measured in Mpc/h_0
  # - from the second column onwards, all the Q_{ll_1} multipoles you want;
  #   they must be ordered followinf the ascending multipole order (so ll_1 = 0
  #   must be the 2 column), and you can go as further as you want in the multipole
  #   order
  # 
  # s [Mpc/h_0]      Q_{l1=0}      Q_{l1=1}      Q_{l1=2}      ...
  1.0     0.9999999999999999      1.445269850978461e-7      0.000011917268941324522    ...
  21.0    0.9832914433168294      -0.0025537781362117177  -0.0033199998619560947        ...
  41.0    0.9669175943095181      -0.004923364937797496      -0.006463561496567318        ...     
  ...     ...                  ...                      ...

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

Returns

A Tuple{Vector{Float64}, Vector{Float64}} with:

the k values vector as first element;

the correspoding PS pk values vector as second one.

See also: create_file_for_XiMultipoles, XiMultipoles, GenericWindow, V_survey, A, A_prime, EPLs, print_PS_multipole

GaPSE.print_PS_multipole_GenWin — Function

print_PS_multipole_GenWin(
    ximult::Union{XiMultipoles,String}, genwin::Union{GenericWindow,String},
    out::String; alg::Symbol=:fftlog, L::Int=0, 
    cut_first_n::Int=0, cut_last_n::Int=0,
    kwargs...)

Compute and save in the file out the L-order Power Spectrum (PS) multipole for a generic window function, through the following Fast Fourier Transform and the effective redshift approximation:

\[ \left\langle \hat{P}_L(k) \right\rangle = \frac{2 L + 1}{A} (-i)^L \sum_{\ell=0}^{\infty} \sum_{\ell_1=0}^{\infty} \begin{pmatrix} L & \ell & \ell_1 \\ 0 & 0 & 0 \end{pmatrix}^2 \int_0^{\infty}\!\mathrm{d} s \, s^2 \, \xi_\ell(s, s_{\rm eff}) \, j_L(k s) \, Q_{\ell_1}(s) \, ,\]

where:

$\left\langle \hat{P}_L(k) \right\rangle$ is the order $L$ Power Spectrum of the effect we are interested in; we are basing this expression on the Yamamoto estimator (see Yamamoto (2000) and Yamamoto (2006))

$A$ is a normalization constant

the 2x3 matrix represents the Wigner-3j symbols

$\xi_\ell$ is the order $\ell$ multipole of the Two-Point Correlation Function (TPCF)

$j_L$ is the spherical Bessel function of order $L$

$s_{\mathrm{eff}}$ is the comoving distance associated with the effective redshift (see the TUTORIAL.ipynb notebook)

$Q_{\ell_1}$ can be easily estimated with FFT methods:

\[\begin{split} Q_{\ell_1}(s) &:= \int_0^{\infty}\mathrm{d}s_1 \; s_1^2 \; \phi(s_1) \; F_{\ell_1}(s_1, s) \\ F_{\ell_1} (s_1 , s) &:= \int_{4\pi} \mathrm{d}\Omega_{\mathbf{\hat{s}}} \, \int_{4\pi} \mathrm{d}\Omega_{\mathbf{\hat{s}}_1} \, \phi(s_2) \, W(\mathbf{\hat{s}}_1) \, W(\mathbf{\hat{s}}_2) \, \mathcal{L}_{\ell_1}(\mathbf{\hat{s}} \cdot \mathbf{\hat{s}}_1) \, . \end{split}\]

where:

$\mathcal{L}_{\ell_1}$ is the Legendre polynomial of order $\ell_1$

$\mathrm{d}\Omega_{\mathbf{\hat{s}}}$ is the infinitesimal solid angle pointing in the direction of the versor $\mathbf{\hat{s}}$

$\phi(s)$ and $W(\mathbf{\hat{s}})$ are respectively the radial and angular part of your window function (we remember that we assumed that such separability of the window function is possible)

Check Eq.(2.13) of Castorina, Di Dio for the theoretical explanation of this formula.

Currenlty, there are two algorithms you can choose in order to perform the computation; you can choose which one to use through the keyword value alg:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

IMPORTANT: no matter which algorithm you choose, you will need to give the input data in a LOGARITHMICALLY DISTRIBUTED scale. A linear distribution does not fit for the algorithms to apply.

Inputs

ximult::Union{XiMultipoles,String} : Two-Point Correlation functions to be used in the computation. You can provide either a XiMultipoles struct containing them or the String filename where they are stored (that will be internally open with XiMultipoles too). The file must have a structure analogous to the genwin one (see the next Inputs item). You can use create_file_for_XiMultipoles to produce such a file.

genwin::Union{GenericWindow,String} : multipoles $Q_{\ell_1}$ of the generic window function you want to consider. You can provide either a GenericWindow struct containing them or the String filename where they are stored (that will be internally open with GenericWindow too). The file must have the following structure

  $ cat Ql_multipoles.txt 
  # Any comment line in the file must start with a #
  # you can have how many comment lines you want in the header; they 
  # will be all skipped.
  # Then you must provide in blank space separated columns:
  # - as first column, the comoving distance values, measured in Mpc/h_0
  # - from the second column onwards, all the Q_{ll_1} multipoles you want;
  #   they must be ordered followinf the ascending multipole order (so ll_1 = 0
  #   must be the 2 column), and you can go as further as you want in the multipole
  #   order
  # 
  # s [Mpc/h_0]      Q_{l1=0}      Q_{l1=1}      Q_{l1=2}      ...
  1.0     0.9999999999999999      1.445269850978461e-7      0.000011917268941324522    ...
  21.0    0.9832914433168294      -0.0025537781362117177  -0.0033199998619560947        ...
  41.0    0.9669175943095181      -0.004923364937797496      -0.006463561496567318        ...     
  ...     ...                  ...                      ...

out::String : name of the file where the results must be stored.

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

See also: create_file_for_XiMultipoles, XiMultipoles, GenericWindow, V_survey, A, A_prime, EPLs, print_PS_multipole, PS_multipole_GenWin