The Power Spectrum with FFTLog

GaPSE.FFTLog_PS_multipole — Function

 FFTLog_PS_multipole(ss, xis;
      pr::Bool=true,
      L::Int=0, ν::Union{Float64,Nothing}=nothing,
      n_extrap_low::Int=500,
      n_extrap_high::Int=500, n_pad::Int=500,
      )

Computes the Power Spectrum through the FFTLog algorithm. More precisely, it computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and xis.

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

See also: PS_multipole

GaPSE.FFTLog_all_PS_multipole — Function

 FFTLog_all_PS_multipole(input::String,
      group::String=VALID_GROUPS[end];
      L::Int=0, pr::Bool=true,
      ν::Union{Float64,Nothing,Vector{Float64}}=nothing,
      n_extrap_low::Int=500,
      n_extrap_high::Int=500, n_pad::Int=500
 )

Computes the Power Spectrum through the FFTLog algorithm for a set of TPCFs. More precisely, it read the input file input, taking the first column as the x-axis ss vector and the following columns as the y-axis ones, and computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and each y-axis xis.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

pr::Bool=true : want to print the automatic messages to the screen?

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

See also: FFTLog_PS_multipole, PS_multipole

The Power Spectrum with TwoFAST

GaPSE.TwoFAST_PS_multipole — Function

 TwoFAST_PS_multipole(f_in;
      int_s_min::Float64 = 1e-1, int_s_max::Float64 = 1e3,
      L::Int = 0, N::Int = 1024, pr::Bool = true,
      k0::Union{Nothing,Float64} = nothing,
      right::Union{Float64,Nothing} = nothing
 ) ::Tuple{Vector{Float64}, Vector{Float64}}

Computes the Power Spectrum from the input spline f_in through the TwoFAST xicalc function of the TwoFAST Julia package. More precisely, it computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the inpunt spline.

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

N::Int = 1024 : number of points to be used in Fourier transform

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / int_s_max

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

See also: PS_multipole

 TwoFAST_PS_multipole(ss, fs;
      int_s_min::Float64 = 1e-1, int_s_max::Float64 = 1e3,
      epl::Bool = true, pr::Bool = true, L::Int = 0,
      N_left::Int = 12, N_right::Int = 12,
      p0_left = [-2.0, 1.0], p0_right = [-2.0, 1.0],
      k0::Union{Nothing,Float64} = nothing
 ) ::Tuple{Vector{Float64}, Vector{Float64}}

Takes the input data vector ss and fs and creates a spline from them, passing it as input tho the other TwoFAST_PS_multipole method. Depending on the options, it may create also a power law epansions on the edges.

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

See also: PS_multipole

GaPSE.TwoFAST_all_PS_multipole — Function

 TwoFAST_all_PS_multipole(input::String,
      group::String=VALID_GROUPS[end];
      L::Int = 0, pr::Bool = true, 
      kwargs...)

Computes the Power Spectrum through the TwoFAST xicalc function of the TwoFAST Julia package for a set of TPCFs. More precisely, it read the input file input, taking the first column as the x-axis ss vector and the following columns as the y-axis ones, and computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and each y-axis xis.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

See also: TwoFAST_PS_multipole, PS_multipole

The Power Spectrum multipole computation

GaPSE.PS_multipole — Function

 PS_multipole(ss, fs; 
      pr::Bool = true, L::Int = 0, 
      alg::Symbol = :fftlog, 
      cut_first_n::Int = 0, cut_last_n::Int = 0, 
      kwargs...
      ) ::Tuple{Vector{Float64}, Vector{Float64}}

 PS_multipole(input::String; 
      kwargs...)

Return the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation:

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi} The second method reads the input file, takes the first column as `ss` and the second as `fs` and recalls the first method. Currenlty, there are two algorithms you can choose in order to perform the computation; you can choose which one to use through the keyword value `alg`: - `alg = :fftlog` (default and recommended option) will employ the [FFTLog](https://github.com/marcobonici/FFTLog.jl) algorithm. - `alg = :twofast` will employ the TwoFAST `xicalc` function of the [TwoFAST](https://github.com/hsgg/TwoFAST.jl) Julia package. Note that in the computation the integration range ``0\leq s \leq \infty`` is reduced to `int_s_min ≤ s ≤ int_s_max`. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round. IMPORTANT: no matter which algorithm you choose, you will need to give the input data in a LOGARITHMICALLY DISTRIBUTED scale. A linear distribution does not fit for the algorithms to apply. ## Optional arguments Depending on the algorithm you choose, the options would change. The options in common are: - `pr::Bool=true` : want to print the automatic messages to the screen? - `L::Int=0` : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF! - `cut_first_n::Int=0` and `cut_last_n::Int=0` : you can cout the first and/or last n elements of the input data, if they are highly irregular. The specific ones for `alg = :fftlog` are: - `ν::Union{Float64,Nothing} = nothing` : bias parameter, i.e. exponent used to "balance" the curve; if `nothing`, will be set automatically to `1.5` - `n_extrap_low::Int = 500` and `n_extrap_high::Int = 500` : number of points to concatenate on the left/right of the input x-axis `ss` vector, logarithmically distributed with the same ratio of the left/right-edge elements of `ss`. - `n_pad::Int = 500` : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm. The specific ones for `alg = :twofast` are: - `epl::Bool=true` : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from `EPLs`) - `N_left::Int = 12` and `N_right::Int = 12` : number of points from left right edges to be used for the power law fitting in `EPLs`. They matters only if in the given input file ξ is not defined until the extremes of integration `int_s_min` and `int_s_max`. - `int_s_min::Float64 = 1e-1` and `int_s_max::Float64 = 1e3`: extremes of integration; if `epl` is set to `false`, their values will be automatically set to `min(ss...)` and `max(ss...)` respectively. Their values do matter only if `epl=true`. - `p0_left=[-2.0, 1.0]` and `p0_right=[-2.0, 1.0]`: vectors with the initial values for the left/right power-law fitting of `EPLs`; the power-law is in the form ``y = f(x) = b * x^s``, so the first vector value is the initial value of ``s`` (and of course the second is the one of ``b``). - `k0::Union{Nothing,Float64} = nothing` : starting point for the `xicalc` function; if `nothing`, it will be set `k0 = 1.0 / max(ss...)` - `right::Union{Float64,Nothing} = nothing` : do you want to cut the output elements with `ks .> right`? if set to `nothing`, no cut will be done. - `N::Int = 1024` : number of points to be used in Fourier transform ## Analytical derivation The analytical expression previously showed can be easily obtained from the standard one:\]

math \begin{split} PL(k) = &\frac{2 L + 1}{A} (-i)^L \, \int0^\infty \mathrm{d} s1 \; s1^2 \int0^\infty \mathrm{d} s \; s^2 \int{-1}^{+1} \mathrm{d} \mu \; jL(ks) \, \xi(s1, s, \mu) \, \phi(s1) \, \phi(s2) \, \mathcal{L}L(\mu) F\left(\frac{s}{s1}, \mu \right) \ &\mathrm{with} \; \;s2 = s2(s1, s, μ) = \sqrt{s1^2 + s^2 + 2s1s\mu} \; , \quad A(s\mathrm{max}, s\mathrm{min}, \theta\mathrm{max}) \frac{ V(s\mathrm{max}, s\mathrm{min}, \theta_\mathrm{max}) }{4 \, \pi^2} \end{split}

with the definition

math f\mathrm{in}(s1, s) = \int{-1}^{+1} \mathrm{d} \mu \; \xi(s1, s, \mu) \, \phi(s2) \, \mathcal{L}L(\mu) \, F\left(\frac{s}{s_1}, \mu \right) ``` and the application of the effective redshift approximation.

Returns

A Tuple{Vector{Float64}, Vector{Float64}} with:

the k values vector as first element;

the correspoding PS pk values vector as second one.

See also: V_survey, A, A_prime, EPLs, print_PS_multipole

GaPSE.print_PS_multipole — Function

 print_PS_multipole(ss, fs, out::String;
      L::Int=0, pr::Bool=true, alg::Symbol=:fftlog, kwargs...)
 print_PS_multipole(input::String, out::String;
      kwargs...)

Takes in input a filename input where is stored a TPCF multipole, calculate the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and xis, and save it in the file out, together with the options used for the computation.

The second method reads the input file, takes the first column as ss and the second as fs and recalls the first method.

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

alg::Symbol = :fftlog : algorithm to be used for the computation. Currenlty, there are two algorithms you can coose in order to perform the computation:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

See also: V_survey, A, A_prime, EPLs, PS_multipole

GaPSE.all_PS_multipole — Function

 function all_PS_multipole(input::String,
      group::String=VALID_GROUPS[end];
      L::Int = 0, pr::Bool = true, 
      alg::Symbol=:fftlog, kwargs...
      ) ::Tuple{Vector{Float64}, Vector{Vector{Float64}}}

Given an input file where the first column is the x-axis data one and all the following columns are the corresponding y-data ones, this function computes all the Power Spectra of each y-data column and return a Tuple containing

as first element, the ks values, common to all the PS

as second element, a vector where in each position there is the Power Spectra corresponding to the associated inputy y-data.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

alg::Symbol = :fftlog : algorithm to be used for the computation. Currenlty, there are two algorithms you can coose in order to perform the computation:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

See also: EPLs, PS_multipole

GaPSE.print_all_PS_multipole — Function

 print_all_PS_multipole(input::String, out::String,
      group::String = VALID_GROUPS[end]; 
      L::Int = 0, pr::Bool = true, 
      alg::Symbol = :fftlog,
      kwargs...)

Given an inputfile where the first column is the x-axis data one and all the following columns are the corresponding y-data ones, this function computes all the Power Spectra of each y-data column and print in a file namedout`

as first column, the ks values, common to all the PS

as folowing columns, a vector where in each position there is the Power Spectra corresponding to the associated inputy y-data.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

alg::Symbol = :fftlog : algorithm to be used for the computation. Currenlty, there are two algorithms you can coose in order to perform the computation:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

See also: EPLs, PS_multipole