The Window Function F

GaPSE.DEFAULT_FMAP_OPTS_hcub — Constant

DEFAULT_FMAP_OPTS_hcub = Dict(
    :θ_max => π / 2.0::Float64, 
    :tolerance => 1e-10::Float64, 
    :rtol => 1e-2::Float64, 
    :atol => 1e-3::Float64,
    :pr => true::Bool,
)

The default values to be used for the F function when you want to perform the computation with hcubature.

See also: integrand_F, F_hcub, print_map_F

GaPSE.DEFAULT_FMAP_OPTS_trap — Constant

DEFAULT_FMAP_OPTS_trap = Dict(
    :θ_max => π / 2.0::Float64, 
    :tolerance => 1e-10::Float64, 
    :N => 300::Int64, 
    :en => 1.0::Float64,
    :pr => true::Bool,
)

The default values to be used for the F function when you want to perform the computation with trap.

See also: integrand_F, F_trap, print_map_F

GaPSE.integrand_F — Function

integrand_F(θ_1, θ, x, μ, θ_max; tolerance=1e-8) ::Float64

Return the integrand of the function $F(x,\mu; \theta_\mathrm{max})$, i.e the function $f(x,\mu, \theta, \theta_1; \theta_\mathrm{max})$:

\[\begin{split} f(x,\mu, \theta, \theta_1; \theta_\mathrm{max}) = \; &\Theta\left( \frac {x \cos \theta + \cos \theta_1}{\sqrt{x^1+2+2x\mu}} - \cos(\theta_\mathrm{max}) \right) \; \times \; \Theta(\mu-\cos(\theta+\theta_1)) \; \, \times \\ & \quad \Theta(\cos(\theta - \theta_1)-\mu) \; \times \; \frac{4\pi \sin\theta\sin\theta_1} {\sqrt{(\sin\theta\sin\theta_1)^2-(\cos\theta\cos\theta_1-\mu)^2}} \end{split}\]

\[\begin{equation} F(x,\mu; \theta_\mathrm{max}) = \int_0^{\theta_\mathrm{max}} \mathrm{d}\theta_1 \int_0^\pi \mathrm{d} \theta \; f(x,\mu, \theta, \theta_1; \theta_\mathrm{max}) \end{equation}\]

tolerance is a parameter needed in case of small negative denominator: the Heaviside theta function mathematically prevent that $\mathrm{den}=(\sin\theta\sin\theta_1)^2-(\cos\theta\cos\theta_1-\mu)^2$ becomes negative, but computationally might happen that $\mathrm{den}$ results as a very small negative number (for instance -1.2368946523-18); in this case tolerance solve the problem, returning 0 if $0<-\mathrm{den}< \mathrm{tolerance}$

See also: print_map_F

GaPSE.F_hcub — Function

F_hcub(x, μ; θ_max = π/2, tolerance = 1e-10, 
    atol = 1e-2, rtol = 1e-5, 
    kwargs...) ::Tuple{Float64, Float64}

Computes with hcubature the value of $F(x,\mu; \theta_\mathrm{max})$, defined as follows:

\[\begin{split} F(x,\mu; \theta_\mathrm{max}) = & \;4\pi \int_0^{\theta_\mathrm{max}} \mathrm{d}\theta_1 \int_0^\pi \mathrm{d} \theta \; \, \Theta\left(\frac {x \cos \theta + \cos \theta_1}{\sqrt{x^1+2+2x\mu}} - \cos(\theta_\mathrm{max}) \right) \, \Theta(\mu-\cos(\theta+\theta_1)) \\ &\Theta(\cos(\theta - \theta_1)-\mu) \; \frac{\sin\theta\sin\theta_1} {\sqrt{(\sin\theta\sin\theta_1)^2-(\cos\theta\cos\theta_1-\mu)^2}} \end{split}\]

tolerance is a parameter needed in case of small negative denominator: the Heaviside theta function mathematically prevent that $\mathrm{den}=(\sin\theta\sin\theta_1)^2-(\cos\theta\cos\theta_1-\mu)^2$ becomes negative, but computationally might happen that $\mathrm{den}$ results as a very small negative number (for instance -1.2368946523-18); in this case tolerance solve the problem, returning 0 if $0<-\mathrm{den}< \mathrm{tolerance}$.

The double integral is performed with hcubature() function from the Julia Package HCubature.jl; rtol, atol and all the kwargs insert into F are directly transferred to hcubature.

The output of this function is a Tuple{Float64, Float64}, containing respectively the value of the integral ad its error.

PAY ATTENTION: do not set too small atol and rtol, or the computation can easily become overwhelming!

NOTE: for computational efficiency and stability, it is highly recommended to use the other functionF_trap, based on the trapezoidal rule, in order to compute this function.

See also: F_trap, print_map_F, integrand_F, check_compatible_dicts

GaPSE.F_trap — Function

F_trap(x, μ; θ_max = π/2, tolerance = 1e-10, 
    atol = 1e-2, rtol = 1e-5, 
    kwargs...) ::Float64

Computes with trap the value of $F(x,\mu; \theta_\mathrm{max})$, defined as follows:

\[\begin{split} F(x,\mu; \theta_\mathrm{max}) = & \;4\pi \int_0^{\theta_\mathrm{max}} \mathrm{d}\theta_1 \int_0^\pi \mathrm{d} \theta \; \, \Theta\left(\frac {x \cos \theta + \cos \theta_1}{\sqrt{x^1+2+2x\mu}} - \cos(\theta_\mathrm{max}) \right) \, \Theta(\mu-\cos(\theta+\theta_1)) \\ &\Theta(\cos(\theta - \theta_1)-\mu) \; \frac{\sin\theta\sin\theta_1} {\sqrt{(\sin\theta\sin\theta_1)^2-(\cos\theta\cos\theta_1-\mu)^2}} \end{split}\]

tolerance is a parameter needed in case of small negative denominator: the Heaviside theta function mathematically prevent that $\mathrm{den}=(\sin\theta\sin\theta_1)^2-(\cos\theta\cos\theta_1-\mu)^2$ becomes negative, but computationally might happen that $\mathrm{den}$ results as a very small negative number (for instance -1.2368946523-18); in this case tolerance solve the problem, returning 0 if $0<-\mathrm{den}< \mathrm{tolerance}$.

The double integral is performed with trapz() function from the Julia Package Trapz.jl, that is based on the trapezoidal rule. N is the number of point to be used to sample INDIPENDENTLY θ_1 and θ, so consider that there is a N^2 time dependence. It's recommended to set 100 < N < 1000.

NOTE: there is another function, called F_hcub, that performs this calculus. Nevertheless, for computational efficiency and stability, it is highly recommended to use to use this one.

See also: F_hcub, print_map_F, integrand_F, check_compatible_dicts

GaPSE.print_map_F — Function

print_map_F(out::String, x_step::Float64 = 0.01, μ_step::Float64 = 0.01;
    alg::Symbol = :trap, x1 = 0, x2 = 3, μ1 = -1, μ2 = 1, 
    Fmap_opts::Dict = Dict{Symbol,Any}(), 
    kwargs...)

print_map_F(out::String, xs::Vector{Float64}, μs::Vector{Float64};
    alg::Symbol = :trap, Fmap_opts::Dict = Dict{Symbol,Any}(),
    kwargs...)

Evaluate the window function $F(x,\mu; \theta_\mathrm{max})$ in a rectangual grid of $\mu$ and $x$ values, and print the results in the out file.

In the first method you have to specify manually, both for x and μ, start (x1 and μ1), stop (x2 and μ2), and step (x_step and μ_step). In the second one, you need to pass the values you want to calculate the function in, through the vectors xs and μs.

The Symbol keyword argument alg tells if you want to perform the computation of F with F_trap (if alg = :trap) or with F_hcub (if alg = :hcub). Other symbols will lead to an AssertionError. Both for computational efficiency and stability, it's highly recommended to use the former (i.e. the default one).

Fmap_opts is instead the way you should exploit in order to pass to F_trap/F_hcub other options you are interested in. You may pass only the key and the value you are focused on, and all the other default ones will be considered. A common key to both the algortithm is :pr => true or :pr => false, that tells if you want to se the progress-bar of the computation.

For example, if you set trap == false and:

Fmap_opts = Dict(:tolerance => 1e-5, :θ_max => 2.0)

then the dictionary with all the options that will be passed to F will be:

Fmap_dict = merge(DEFAULT_FMAP_OPTS_hcub, Fmap_opts) = :θ_max => 2.0, # CHANGED VALUE :tolerance => 1e-5, # CHANGED VALUE :rtol => 1e-2, # default :atol => 1e-3, # default :pr => true, # default )

Check the documentation of DEFAULT_FMAP_OPTS_hcub and DEFAULT_FMAP_OPTS_trap for more information about these default values.

See also: DEFAULT_FMAP_OPTS_hcub, DEFAULT_FMAP_OPTS_trap F_trap, F_hcub, integrand_F

GaPSE.WindowF — Type

WindowF(
    xs::Vector{Float64}
    μs::Vector{Float64}
    Fs::Matrix{Float64}
    )

Struct containing xs, μs and Fs values of the window function $F(x, μ)$. xs and μs are 1D vectors containing each value only once, while Fs values are contained in a matrix of size (length(xs), length(μs)), so:

along a fixed column the changing value is x

along a fixed row the changing value is μ

The analytical definition of the window function is the following (see Eq. A.10 of Castorina, Di Dio, 2021):

\[\begin{split} F(x,\mu; \theta_\mathrm{max}) = & \;4\pi \int_0^{\theta_\mathrm{max}} \mathrm{d}\theta_1 \int_0^\pi \mathrm{d} \theta \; \, \Theta\left(\frac {x \cos \theta + \cos \theta_1}{\sqrt{x^1+2+2x\mu}} - \cos(\theta_\mathrm{max}) \right) \, \Theta(\mu-\cos(\theta+\theta_1)) \\ &\Theta(\cos(\theta - \theta_1)-\mu) \; \frac{\sin\theta\sin\theta_1} {\sqrt{(\sin\theta\sin\theta_1)^2-(\cos\theta\cos\theta_1-\mu)^2}} \end{split}\]

where the $\Theta$s are Heaviside theta functions.

Constructors

WindowF(file::String) : read the F map from the file file. Such a file might be produced by print_map_F, check its docstring.

It does not matter if the pattern is

# xs      μs      Fs
0.0       -1.0       ...
0.0       -0.9       ...
0.0       -0.8       ...
...       ...      ...

or

# xs      μs      Fs
0.0       -1.0       ...
0.1       -1.0       ...
0.2       -1.0       ...
...       ...      ...

because the constructor will recognise it. What does matter is the columns order: xs first, then μs and finally Fs.

See also: print_map_F, F_trap, spline_F

GaPSE.spline_F — Function

spline_F(x, μ, str::WindowF)) ::Float64

Return the 2-dim spline value of $F$ in the given (x,μ), where $F$ is defined in the input WindowF. The spline is obtained through the interpolate function of the GridInterpolations Julia package.

See also: WindowF