The F window function

GaPSE.integrated_F_quadgk — Function

 integrated_F_quadgk(s, μ, s_min, s_max, windowF::WindowF;
      llim=0.0, rlim=Inf, rtol=1e-2, atol=0.0)

Computes the Integrated Window Function fron the input Window Function windowF, through the quadgk function of the QuadGK Julia pagkage, in the point (s,μ) (where s is the comoving distance and μ the angle cosine). s_min and s_max are the min and max value for the radial part of the survey window function.

The analytical expression of the Integrated window function is the following:

\[\mathcal{F}(s, \mu) = \int_0^\infty \mathrm{d}s_1 \, \phi(s_1) \, \phi\left(\sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}\right) \, F\left(\frac{s}{s_1}, \mu \right)\]

where $s$ is the comoving distance, $\mu$ the cosine angle, $\phi$ is the angular part of the survey window function and $F(x, μ)$ is the window function. Check the documentation of WindowF for its definition. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Optional arguments

llim=nothing and rlim=nothing : integration limits for quadgk; if llim=nothing it will be set to 0.95 * s_min; if rlim=nothing, it will be set to 1.05*s_max, while if rlim=Inf it will be set to 3 * s_max

rtol=1e-2 and atol=0.0 : relative and absoute tolerance for quadgk

See also: [WindowF], ϕ

GaPSE.integrated_F_trapz — Function

 integrated_F_trapz(s, μ, s_min, s_max, windowF::WindowF;
      llim=nothing, rlim=nothing, N::Int=1000)

Computes the Integrated Window Function fron the input Window Function windowF, through the trapz function of the Trapz Julia pagkage, in the point (s,μ) (where s is the comoving distance and μ the angle cosine). s_min and s_max are the min and max value for the radial part of the survey window function.

The analytical expression of the Integrated window function is the following:

\[\mathcal{F}(s, \mu) = \int_0^\infty \mathrm{d}s_1 \, \phi(s_1) \, \phi\left(\sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}\right) \, F\left(\frac{s}{s_1}, \mu \right)\]

where $s$ is the comoving distance, $\mu$ the cosine angle, $\phi$ is the angular part of the survey window function and $F(x, μ)$ is the window function. Check the documentation of WindowF for its definition. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Optional arguments

llim=nothing and rlim=nothing : limits of the sampling interval to be used for trapz; if llim=nothing it will be set to 0.95 * s_min; if rlim=nothing, it will be set to 1.05*s_max, while if rlim=Inf it will be set to 3 * s_max

N::Int = 1000 : number of points to be used for the sampling

See also: [WindowF], ϕ

GaPSE.print_map_IntegratedF — Function

 print_map_IntegratedF(
      s_min, s_max, 
      ss::Vector{Float64}, μs::Vector{Float64}, 
      windowF::Union{String,WindowF}, out::String;
      alg::Symbol=:trap, llim=nothing, rlim=nothing,
      rtol=1e-2, atol=0.0, N::Int=1000, pr::Bool=true)

 print_map_IntegratedF(
      z_min, z_max, 
      zs::Vector{Float64}, μs::Vector{Float64}, 
      windowF::Union{String,WindowF}, out::String,
      file_data::String; 
      names_bg = NAMES_BACKGROUND, h_0 = 0.7, kwargs...)

 print_map_IntegratedF(
      z_min, z_max,
      μs::Vector{Float64}, 
      windowF::Union{String,WindowF}, out::String,
      file_data::String;
      names_bg = NAMES_BACKGROUND, h_0 = 0.7, N_ss::Int = 100, 
      m::Float64 = 2.1, kwargs...)

Evaluate the integrated window function $\mathcal{F}(s,\mu)$ in a rectangual grid of $\mu$ and $s$ values, and print the results in the out file.

The first method takes as input:

s_min and s_max : min and max comoving distance of the survey; their values will be internally used by the radial function ϕ

ss::Vector{Float64} and μs::Vector{Float64} : the vector of s and μ points where to sample the integrated window function $\mathcal{F}$. They must be a float vector of increasing values; more precisely:

ss must be a float vector of increasing comoving distance values (so each element must be ≥ 0); the first and last values ARE NOT RELATED to s_min and s_max.

μs must be a float vector of increasing cosine values (so each element x must be -1 ≤ x ≤ 1).

windowF::Union{String,WindowF}, i.e. the window function itself; it can be passed as the namefile where the window is stored in (that will be opened with WindowF) or as a WindowF struct directly.

out::String : the name of the output file

The second method takes as input the min and max redshifts of the survey (z_minand z_max), the vector of redshifts zs::Vector{Float64} for the integrated window function sampling, μs and windowF as before and the file_data where can be found the association $z \rightarrow s(z)$. Such file must have the structure of the background data produced by the CLASS code. Note that also zs musyt be a float vector of increasing redshift values (so each element must be ≥ 0). This method internally recalls the first one, so the other kwargs... are in common.

The third method takes as input the min and max redshifts of the survey (z_minand z_max) and the same input as the second method (μs, widnowF, out and file_data) but NOT THE REDSHIFT SAMPLING VECTOR zs. The sampling will be internally made linearly from $s = 0$ to $s = m \, s_{\mathrm{max}}$, where s_max is the comoving distance associated to z_max (for the data stored in file_data) and m::Float64 = 2.1 a coefficient that we suggest to set equals to 2 < m < 3. N_ss::Int = 100 is the number of s values used for the sampling in the interval $[0, m \, s_{\mathrm{max}}]$. This method internally recalls the first one, so the other kwargs... are in common.

The analytical expression for the integrated window function is the following:

\[\mathcal{F}(s, \mu) = \int_0^\infty \mathrm{d}s_1 \, \phi(s_1) \, \phi\left(\sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}\right) \, F\left(\frac{s}{s_1}, \mu \right)\]

where $s$ is the comoving distance, $\mu$ the cosine angle, $\phi$ is the angular part of the survey window function and $F(x, μ)$ is the window function. Check the documentation of WindowF for its definition. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Optional arguments

As optional arguments of the first method:

alg::Symbol = :trap : algorithm to be used for the integration; the valid options are :quad (that will recall integrated_F_quadgk) and :trap (that will recall integrated_F_trapz); other values will lead to AssertionError

llim=nothing and rlim=nothing : integration limits for quad/trap; if llim=nothing it will be set to 0.95 * s_min; if rlim=nothing, it will be set to 1.05*s_max, while if rlim=Inf it will be set to 3 * s_max.

N::Int = 1000 : number of points to be used for the sampling of trapz; it's useless if you set alg = :quad;

rtol=1e-2 and atol=0.0 : relative and absoute tolerance for quadgk; they are useless if you set alg = :trap;

pr::Bool = true : do you want to see the progress-bar of the computation?

The optional arguments given to the second method will be directly given to the first one. The only two exceptions are options relative to the background data, managed internally by the struct BackgroundData:

names = NAMES_BACKGROUND : the column names of the file_data. If the colum order change from the default one NAMES_BACKGROUND, you must set as input the vector of string with the correct one, with the SAME names. They are, with the default order:

["z", "proper time [Gyr]", "conf. time [Mpc]", "H [1/Mpc]", "comov. dist.", "ang.diam.dist.", "lum. dist.", "comov.snd.hrz.", "(.)rhog", "(.)rhob", "(.)rhocdm", "(.)rholambda", "(.)rhour", "(.)rhocrit", "gr.fac. D", "gr.fac. f"]

h = 0.7 : the adimensional hubble constant. By default, CLASS background data are measured with it numerically expressed (so distances are measured in Mpc, for example), while this code works with h in the unit of measure (so distances are measured in Mpc/h, for example). Change this value to 1.0 if the input data do not have this issue, or to your value of interest (0.67, 0.5, ...).

See also: integrated_F_quadgk, integrated_F_trapz, ϕ, WindowF, WindowFIntegrated, BackgroundData

GaPSE.WindowFIntegrated — Type

 WindowFIntegrated(
      ss::Vector{Float64}
      μs::Vector{Float64}
      IFs::Matrix{Float64}
      )

Struct containing ss, μs and IFs values of the integrated window function $\mathcal{F}(s, μ)$. ss and μs are 1D vectors containing each value only once, while IFs values are contained in a matrix of size (length(ss), length(μs)), so:

along a fixed column the changing value is s

along a fixed row the changing value is μ

The analytical expression for the integrated window function is the following:

\[\mathcal{F}(s, \mu) = \int_0^\infty \mathrm{d}s_1 \, \phi(s_1) \, \phi\left(\sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}\right) \, F\left(\frac{s}{s_1}, \mu \right)\]

where $s$ is the comoving distance, $\mu$ the cosine angle, $\phi$ is the angular part of the survey window function and $F(x, μ)$ is the window function. Check the documentation of WindowF for its definition. We remember that all the distances are measured in $h_0^{-1}\mathrm{Mpc}$.

Constructors

 WindowFIntegrated(file::String)

Read the IF map from the file file. Such a file might be produced by print_map_IntegratedF, check its docstring.

It does not matter if the pattern is

# ss      μs      IFs
0.0       -1.0       ...
0.0       -0.9       ...
0.0       -0.8       ...
...       ...      ...

or

# ss      μs      IFs
0.0       -1.0       ...
0.1       -1.0       ...
0.2       -1.0       ...
...       ...      ...

because the constructor will recognise it. What does matter is the columns order: ss first, then μs and finally IFs.

See also: integrated_F_trapz, integrated_F_quadgk, spline_integrF, WindowF, ϕ, print_map_IntegratedF

Missing docstring.

Missing docstring for GaPSE.spline_IntegratedF. Check Documenter's build log for details.