The ξ multipoles computation

GaPSE.integrand_ξ_multipole — Function

 integrand_ξ_multipole(s1, s, μ, effect::Function, cosmo::Cosmology;
      L::Integer = 0, 
      use_windows::Bool = true, 
      kwargs...) ::Float64

 integrand_ξ_multipole(s1, s, μ, effect::String, cosmo::Cosmology; 
      L::Integer = 0, 
      use_windows::Bool = true, 
      kwargs...) ::Float64

Return the integrand on $\mu = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}$ of the of the chosen correlation function term, i.e. the following function $f(s_1, s, \mu)$:

\[ f(s_1, s, \mu) = \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \, \phi(s_2) \, F\left(\frac{s}{s_1}, \mu \right)\]

where $y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$, $s = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$ and $\xi$ is the corresponding CF effect.

In the former method you have to pass as an input the effect function you want to integrate, while in the (recommended) latter one it's necessary to specify the name of the CF term among the following:

auto_doppler , auto_lensing , auto_localgp , auto_integratedgp , lensing_doppler , doppler_lensing , doppler_localgp , localgp_doppler , doppler_integratedgp , integratedgp_doppler , lensing_localgp , localgp_lensing , lensing_integratedgp , integratedgp_lensing , localgp_integratedgp , integratedgp_localgp ,

to which correspond the following functions:

ξ_Doppler , ξ_Lensing , ξ_LocalGP , ξ_IntegratedGP , ξ_Lensing_Doppler , ξ_Doppler_Lensing , ξ_Doppler_LocalGP , ξ_LocalGP_Doppler , ξ_Doppler_IntegratedGP , ξ_IntegratedGP_Doppler , ξ_Lensing_LocalGP , ξ_LocalGP_Lensing , ξ_Lensing_IntegratedGP , ξ_IntegratedGP_Lensing , ξ_LocalGP_IntegratedGP , ξ_IntegratedGP_LocalGP ,

The window functions $F(x, \mu)$ and $\phi(s)$ are calculated for the given Cosmology cosmo through the functions spline_F and ϕ respectivelly.

In case use_windows is set to false, the window functions $\phi$ and $F$ are removed, i.e is returned the following function $f^{'}(s_1, s, \mu)$:

\[ f^{'}(s_1, s, \mu) = \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \]

The function $\xi(s_1, s_2, \cos{\theta})$ is calculated from, depending on the value of effect:

effect == auto_doppler => ξ_Doppler

effect == auto_lensing => ξ_Lensing

effect == auto_localgp => ξ_LocalGP

effect == auto_integratedgp => ξ_IntegratedGP

effect == doppler_lensing => ξ_Doppler_lensing

effect == lensing_doppler => ξ_Lensing_Doppler

effect == doppler_localgp => ξ_Doppler_LocalGP

effect == localgp_doppler => ξ_LocalGP_Doppler

effect == doppler_integratedgp => ξ_Doppler_IntegratedGP

effect == integratedgp_doppler => ξ_IntegratedGP_Doppler

effect == lensing_localgp => ξ_Lensing_LocalGP

effect == localgp_lensing => ξ_LocalGP_Lensing

effect == lensing_integratedgp => ξ_Lensing_IntegratedGP

effect == integratedgp_lensing => ξ_IntegratedGP_Lensing

effect == localgp_integratedgp => ξ_LocalGP_IntegratedGP

effect == integratedgp_localgp => ξ_IntegratedGP_LocalGP

Note that these is an internal conversion of coordiate sistems from (s1, s, μ) to (s1, s2, y) thorugh the functions y and s2

Inputs

s1: the comoving distance where must be evaluated the integral

s: the comoving distance from s1 where must be evaluated the integral

μ: the cosine between s1 and s where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Integer = 0: order of the Legendre polynomial to be used

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $F$

kwargs... : other keyword arguments that will be passed to the selected GR TPCF effect (ξ_Doppler, ξ_Lensing, ...)

See also: ξ_multipole, map_ξ_multipole, print_map_ξ_multipole, spline_F, ϕ, Cosmology, y, s2, IMPLEMENTED_GR_EFFECTS

GaPSE.ξ_multipole — Function

 ξ_multipole(s1, s, effect::Function, cosmo::Cosmology; 
      L::Integer = 0, 
      enhancer::Float64 = 1e6,
      use_windows::Bool = true,
      μ_atol::Float64 = 1e-4,
      μ_rtol::Float64 = 1e-1, 
      kwargs...) ::Float64

 ξ_multipole(s1, s, effect::String, cosmo::Cosmology; 
      L::Integer = 0, 
      enhancer::Float64 = 1e6,
      use_windows::Bool = true,
      μ_atol::Float64 = 1e-4,
      μ_rtol::Float64 = 1e-1, 
      kwargs...) ::Float64

Evaluate the multipole of order L of the chosen correlation function term, through the integrand_ξ_multipole function.

In the former method you have to pass as an input the effect function you want to integrate, while in the (recommended) latter one it's necessary to specify the name of the CF term among the following:

auto_doppler , auto_lensing , auto_localgp , auto_integratedgp , lensing_doppler , doppler_lensing , doppler_localgp , localgp_doppler , doppler_integratedgp , integratedgp_doppler , lensing_localgp , localgp_lensing , lensing_integratedgp , integratedgp_lensing , localgp_integratedgp , integratedgp_localgp ,

to which correspond the following functions:

ξ_Doppler , ξ_Lensing , ξ_LocalGP , ξ_IntegratedGP , ξ_Lensing_Doppler , ξ_Doppler_Lensing , ξ_Doppler_LocalGP , ξ_LocalGP_Doppler , ξ_Doppler_IntegratedGP , ξ_IntegratedGP_Doppler , ξ_Lensing_LocalGP , ξ_LocalGP_Lensing , ξ_Lensing_IntegratedGP , ξ_IntegratedGP_Lensing , ξ_LocalGP_IntegratedGP , ξ_IntegratedGP_LocalGP ,

The function evaluated is then the following:

\[\xi_L(s_1, s, \mu) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \phi(s_2) \, F\left(\frac{s}{s_1}, \mu \right) \;, \quad \mathrm{use_windows == true}\\ 1\;, \quad \mathrm{use_windows == false} \end{cases}\]

where $y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$ and $\xi$ is the chosen CF effect.

The window functions $F(x, \mu)$ and $\phi(s)$ are calculated for the given Cosmology cosmo through the functions spline_F and ϕ respectivelly.

The function $\xi(s_1, s_2, \cos{\theta})$ is calculated from, depending on the value of effect:

effect == auto_doppler => ξ_Doppler

effect == auto_lensing => ξ_Lensing

effect == auto_localgp => ξ_LocalGP

effect == auto_integratedgp => ξ_IntegratedGP

effect == doppler_lensing => ξ_Doppler_lensing

effect == lensing_doppler => ξ_Lensing_Doppler

effect == doppler_localgp => ξ_Doppler_LocalGP

effect == localgp_doppler => ξ_LocalGP_Doppler

effect == doppler_integratedgp => ξ_Doppler_IntegratedGP

effect == integratedgp_doppler => ξ_IntegratedGP_Doppler

effect == lensing_localgp => ξ_Lensing_LocalGP

effect == localgp_lensing => ξ_LocalGP_Lensing

effect == lensing_integratedgp => ξ_Lensing_IntegratedGP

effect == integratedgp_lensing => ξ_IntegratedGP_Lensing

effect == localgp_integratedgp => ξ_LocalGP_IntegratedGP

effect == integratedgp_localgp => ξ_IntegratedGP_LocalGP

Note that these is an internal conversion of coordiate sistems from (s1, s, μ) to (s1, s2, y) thorugh the functions y and s2

Inputs

s1: the comoving distance where must be evaluated the integral

s: the comoving distance from s1 where must be evaluated the integral

cosmo::Cosmology: cosmology to be used in this computation

Optional arguments

L::Integer = 0: order of the Legendre polynomial to be used

enhancer::Float64 = 1e6: just a float number used in order to deal better with small numbers; the returned value is NOT modified by this value, because after a multiplication the internal result is divided by enhancer.

use_windows::Bool = false: tells if the integrand must consider the two window function $\phi$ and $F$

μ_atol::Float64 = 1e-3 and μ_rtol::Float64 = 1e-3: absolute and relative tolerance to be passed to quadgk; it's recommended not to set μ_rtol < 1e-2 because of long time for evaluations

kwargs... : other keyword arguments that will be passed to the selected GR TPCF effect (ξ_Doppler, ξ_Lensing, ...)

See also: integrand_ξ_multipole, map_ξ_multipole, print_map_ξ_multipole spline_F, ϕ, Cosmology, y, s2, IMPLEMENTED_GR_EFFECTS

GaPSE.map_ξ_multipole — Function

 map_ξ_multipole(cosmo::Cosmology,
      effect::Union{String,Function},
      v_ss = nothing;
      s1 = nothing,
      pr::Bool = true,
      N_log::Integer = 1000,
      L::Integer = 0,
      kwargs... ) ::Tuple{Vector{Float64}, Vector{Float64}}

Evaluate the multipole of order L of the chosen correlation function term, through the ξ_multipole function, for all the s values stored inside v_ss. If v_ss = nothing, it is set v_ss = 10 .^ range(0, 3, length = N_log). If s1 = nothing, it is set s1 = cosmo.s_eff.

The function evaluated is then the following:

\[\xi_L(s_1, s, \mu) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \phi(s_2) \, F\left(\frac{s}{s_1}, \mu \right) \;, \quad \mathrm{use_windows == true}\\ 1\;, \quad \mathrm{use_windows == false} \end{cases}\]

where $y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$ and $\xi$ is the chosen CF effect, for all the comoving distances s inside v_ss.

The window functions $F(x, \mu)$ and $\phi(s)$ are calculated for the given Cosmology cosmo through the functions spline_F and ϕ respectivelly.

Inputs

cosmo::Cosmology : cosmology to be used in this computation

effect::Union{String,Function} : the GR effect TPCF you want to consider; you may specify the name of the effect as one of the following strings (recommended):

auto_doppler , auto_lensing , auto_localgp , auto_integratedgp , lensing_doppler , doppler_lensing , doppler_localgp , localgp_doppler , doppler_integratedgp , integratedgp_doppler , lensing_localgp , localgp_lensing , lensing_integratedgp , integratedgp_lensing , localgp_integratedgp , integratedgp_localgp ,

or directly the name of the function among the following:

ξ_Doppler , ξ_Lensing , ξ_LocalGP , ξ_IntegratedGP , ξ_Lensing_Doppler , ξ_Doppler_Lensing , ξ_Doppler_LocalGP , ξ_LocalGP_Doppler , ξ_Doppler_IntegratedGP , ξ_IntegratedGP_Doppler , ξ_Lensing_LocalGP , ξ_LocalGP_Lensing , ξ_Lensing_IntegratedGP , ξ_IntegratedGP_Lensing , ξ_LocalGP_IntegratedGP , ξ_IntegratedGP_LocalGP ,

`v_ss : vector/range of s values where the function must be evaluated; if v_ss = nothing, it is set v_ss = 10 .^ range(0, 3, length = N_log). This is why it is returned also the vector of the "input" values.

Optional arguments

s1 = nothing : comoving distance from the observer where the TPCF should be evaluated; if s1 = nothing, it is automatically set s1 = cosmo.s_eff from the given input Cosmology.

L::Integer = 0: order of the Legendre polynomial to be used

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

N_log::Integer = 1000 : number of points to be used in the default logaritmically-spaced range for v_ss, i.e. range(0, 3, N_log); it is ignored if v_ss ≠ nothing

kwargs... : other keyword arguments that will be passed to ξ_multipole

Returns

A Tuple{Vector{Float64}, Vector{Flaot64}}, which has as first element the v_ss vector and as second one the corresponding ξ value evaluated.

See also: integrand_ξ_multipole, ξ_multipole, print_map_ξ_multipole, spline_F, ϕ, Cosmology, y, s2, IMPLEMENTED_GR_EFFECTS

GaPSE.print_map_ξ_multipole — Function

 print_map_ξ_multipole(
      cosmo::Cosmology,
      out::String,
      effect::Union{String,Function},
      v_ss = nothing;
      s1 = nothing,
      kwargs...)

Evaluate the multipole of order L of the chosen correlation function term, through the ξ_multipole function, for all the s values stored inside v_ss, and print the results (with all the options used) in a file named out. If v_ss = nothing, it is set v_ss = 10 .^ range(0, 3, length = N_log). If s1 = nothing, it is set s1 = cosmo.s_eff.

The function evaluated is then the following:

\[\xi_L(s_1, s, \mu) = \frac{2 L + 1}{2} \int_{-1}^{+1} \mathrm{d}\mu \; \xi (s_1, s_2, \cos{\theta}) \, \mathcal{L}_L(\mu) \, \times \begin{cases} \phi(s_2) \, F\left(\frac{s}{s_1}, \mu \right) \;, \quad \mathrm{use_windows == true}\\ 1\;, \quad \mathrm{use_windows == false} \end{cases}\]

where $y = \cos{\theta} = \hat{\mathbf{s}}_1 \cdot \hat{\mathbf{s}}_2$ and $\xi$ is the chosen CF effect, for all the comoving distances s inside v_ss.

The window functions $F(x, \mu)$ and $\phi(s)$ are calculated for the given Cosmology cosmo through the functions spline_F and ϕ respectivelly.

Inputs

cosmo::Cosmology : cosmology to be used in this computation

effect::Union{String,Function} : the GR effect TPCF you want to consider; you may specify the name of the effect as one of the following strings (recommended):

auto_doppler , auto_lensing , auto_localgp , auto_integratedgp , lensing_doppler , doppler_lensing , doppler_localgp , localgp_doppler , doppler_integratedgp , integratedgp_doppler , lensing_localgp , localgp_lensing , lensing_integratedgp , integratedgp_lensing , localgp_integratedgp , integratedgp_localgp ,

or directly the name of the function among the following:

ξ_Doppler , ξ_Lensing , ξ_LocalGP , ξ_IntegratedGP , ξ_Lensing_Doppler , ξ_Doppler_Lensing , ξ_Doppler_LocalGP , ξ_LocalGP_Doppler , ξ_Doppler_IntegratedGP , ξ_IntegratedGP_Doppler , ξ_Lensing_LocalGP , ξ_LocalGP_Lensing , ξ_Lensing_IntegratedGP , ξ_IntegratedGP_Lensing , ξ_LocalGP_IntegratedGP , ξ_IntegratedGP_LocalGP ,

out::String : name of the file where the results must be stored.

`v_ss : vector/range of s values where the function must be evaluated; if v_ss = nothing, it is set v_ss = 10 .^ range(0, 3, length = N_log).

Optional arguments

s1 = nothing : comoving distance from the observer where the TPCF should be evaluated; if s1 = nothing, it is automatically set s1 = cosmo.s_eff from the given input Cosmology.

kwargs... : other keyword arguments that will be passed to map_ξ_multipole

See also: integrand_ξ_multipole, ξ_multipole, map_ξ_multipole, spline_F, ϕ, Cosmology, y, s2, IMPLEMENTED_GR_EFFECTS