Background functions

GaPSE.f0 — Constant

 const f0 :: Float64

Linear growth rate at present time. Its value is equal to:

\[ f_0 \simeq 0.5126998572951\]

GaPSE.D0 — Constant

 const D0 :: Float64

Linear growth factor at present time. Its value is equal to:

\[ D_0 = 1.0\]

GaPSE.ℋ0 — Constant

 const ℋ0 :: Float64

Comoving Hubble constant at present time. Its value is, in natural system (where the speed of light c=1): $\mathcal{H}_0 \simeq 3.335641\times10^{-4} \; h_0^{-1}\mathrm{Mpc}$

GaPSE.BackgroundData — Type

 BackgroundData(
      z::Vector{Float64}
      conftime::Vector{Float64}
      comdist::Vector{Float64}
      angdist::Vector{Float64}
      lumdist::Vector{Float64}
      D::Vector{Float64}
      f::Vector{Float64}
      ℋ::Vector{Float64}
      ℋ_p::Vector{Float64})

Struct that contains all the relevant cosmological information for future computations. The data are stored with increasing distance values (so the first ones are associated to z=0). It is internally used in Cosmology.

Arguments

z::Vector{Float64} : redshifts (adimensionals).

conftime::Vector{Float64} : conformal times, measured in [Mpc/h].

comdist::Vector{Float64} : comoving distances, measured in [Mpc/h].

angdist::Vector{Float64} : angular diameter distances, measured in [Mpc/h].

lumdist::Vector{Float64} : luminosity distances, measured in [Mpc/h].

D::Vector{Float64} : linear growth factors, normalized to 1.0 at the present day (adimensional).

f::Vector{Float64} : linear growth rates (adimensional).

ℋ::Vector{Float64} : comoving Hubble parameters, measured in [h/Mpc].

ℋ_p::Vector{Float64} : derivatives of the comoving Hubble parameter wrt the conformal time. It is here manually computed with the Dierckx function derivative.

Constructors

BackgroundData(file::String, z_max; names = NAMES_BACKGROUND, h = 0.7)

file::string : input file where the data are stored; it is expected that such file is a background output of the CLASS program (link: https://github.com/lesgourg/class_public)

z_max : the maximum redhsift we are interested in our analysis. The constructor will store the data necessary for a study only in 0 < z < z_max, for optimisation purposes (More precisely, the maximum distance stored will be 3*z_max).

names = NAMES_BACKGROUND : the column names of the file. If the colum order change from the default one NAMES_BACKGROUND, you must set as input the vector of string with the correct one, with the SAME names. They are, with the default order: nothing

h = 0.7 : the adimensional hubble constant. By default, CLASS background data are measured with it nuymerically expressed (so distances are measured in Mpc, for example), while this code works with h in the unit of measure (so distances are measured in Mpc/h, for example). Change this value to 1.0 if the input data do not have this issue, or to your value of interest (0.67, 0.5, ...).

See also: CosmoParams, Cosmology

GaPSE.CosmoParams — Type

 CosmoParams(
      z_min::Float64
      z_max::Float64
      θ_max::Float64

      k_min::Float64
      k_max::Float64

      Ω_b::Float64
      Ω_cdm::Float64
      Ω_M0::Float64
      h_0::Float64

      N::Integer
      fit_min::Float64
      fit_max::Float64
      con::Bool
      s_lim::Float64)

Struct that contains all the parameters and options that are matter of concerns for the Cosmology we are interested in.

Arguments

z_min::Float64 and z_max::Float64 : the minimum and maximum redshifts of the survey we want to study.

θ_max::Float64 : Angular maximum value of the survey. It is implicitly assumed an azimutal simmetry of the survey.

k_min::Float64 and k_max::Float64 : extremes of integration for the σ_i integrals in IPSTools.

Ω_b::Float64, Ω_cdm::Float64 and Ω_M0::Float64 : barionic, cold-dark-matter and total matter density parameters.

h_0::Float64 : today's Hubble adimensional parameter (H_0 = h_0 * 100 km/(s * Mpc)).

N::Integer : number of points to be used in the Sperical Bessel Fourier Transform made by xicalc in IPSTools.

fit_min::Float64 and fit_max::Float64 : the limits (min and max) where the integral $I_\ell^n$ in Cosmology must be fitted with a power law, for small distances. This operation is necessary, because xicalc, in this context, gives wrong results for too small input distance s; nevertheless, all these $I_\ell^n$ integrals have fixed power-law trends for $s \rightarrow 0$, so this approach gives good results.

con::Bool : do you want that the fit of all the $I_\ell^n$ in IPSTools for the LEFT edge is not a simple power-law $y = f(x) = b \, x^s$, but also consider a constant $a$, such that $y = f(x) = a + b \, x^s$?

s_lim::Float64 : the lower-bound value for the function func_ℛ; it is necessary, because ℛ blows up for $s \rightarrow 0^{+}$. Consequently, if the func_ℛ input value is 0 ≤ s < s_lim, the returned value is always func_ℛ(s_lim).

Constructors

CosmoParams(z_min, z_max, θ_max; k_min = 1e-8, k_max = 10.0, Ω_b = 0.0489, Ω_cdm = 0.251020, h_0 = 0.70, N::Integer = 1024, fit_min = 0.05, fit_max = 0.5, con::Bool = true, s_lim = 1e-2) The associations are trivials, with Ω_M0 = Ω_cdm + Ω_b.

See also: Cosmology, IPSTools, func_ℛ

GaPSE.func_z_eff — Function

 func_z_eff(s_min, s_max, z_of_s) :: Float64

Return the effective redshift $z_\mathrm{eff}$, calcuated as follows:

\[\begin{split} z_\mathrm{eff} := \frac{ \int \mathrm{d}^3\mathbf{s} \, \phi^2(\mathbf{s}) \, z(s) }{ \int \mathrm{d}^3\mathbf{s}\, \phi^2(\mathbf{s}) } &= \frac{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s) \, z(s) \times \int_{4\pi}\mathrm{d}^2\hat{\mathbf{s}} \, W^2(\hat{\mathbf{s}}) }{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s)\times \int_{4\pi}\mathrm{d}^2\hat{\mathbf{s}} \, W^2(\hat{\mathbf{s}}) } \\[5pt] &= \frac{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s) \, z(s) }{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s) } \\[4pt] &= \frac{3}{s_\mathrm{max}^3 - s_\mathrm{min}^3} \, \int_{s_\mathrm{min}}^{s_\mathrm{max}} \mathrm{d}s \, s^2 \, z(s) \end{split}\]

where we have used our assuption on separability of the window function

\[ \phi(\mathbf{s}) = \phi(s) \, W(\hat{s})\]

and their definitions.

See also: ϕ, W

GaPSE.s — Function

 s(s1, s2, y) :: Float64

Return the value $s = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$

See also: μ, s2, y

GaPSE.μ — Function

 μ(s1, s2, y) :: Float64

Return the value $\mu=\hat{\mathbf{s}}_1\dot\hat{\mathbf{s}}$, defined as:

\[\mu = \mu(s_1, s_2, y) = \frac{y \, s_2 - s_1}{s(s_1, s_2, y)} \;, \quad s(s_1, s_2, y) = \sqrt{s_1^2 + s^2 - 2 \, s_1 \, s_2 \, y}\]

with $y=\cos\theta=\hat{\mathbf{s}}_1\dot\hat{\mathbf{s}}$ and where $s$ is obtained from the function s

See also: s, s2, y

GaPSE.s2 — Function

 s2(s1, s, μ) :: Float64

Return the value $s_2 = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}$

See also: s, μ, y

GaPSE.y — Function

 y(s1, s, μ) :: Float64

Return the value $y=\cos\theta$, defined as:

\[y = y(s_1, s, \mu) = \frac{\mu \, s + s_1}{s2(s_1, s, \mu)} \;, \quad s_2 = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}\]

with $\mu=\hat{\mathbf{s}}_1\dot\hat{\mathbf{s}}_2$ and where $s_2$ is btained from the function s2

See also: s, μ, s2