The Cosmology Parameters

GaPSE.DEFAULT_IPS_OPTS — Constant

 const DEFAULT_IPS_OPTS = Dict(
      :fit_left_min => 1e-6::Float64, 
      :fit_left_max => 3e-6::Float64,
      :fit_right_min => 1e1::Float64, 
      :fit_right_max => 2e1::Float64,
      )

The default values to be stored in CosmoParams concerning the Input Power Spectrum. In the Cosmology that will have such CosmoParams as input, they will be used in its InputPS.

See also: CosmoParams, Cosmology, InputPS

GaPSE.DEFAULT_IPSTOOLS_OPTS — Constant

 const DEFAULT_IPSTOOLS_OPTS = Dict(
      :N => 1024::Integer,
      :fit_min => 0.05::Float64,
      :fit_max => 0.5::Float64,
      :con => true::Bool,
      :k_min => 1e-6::Float64,
      :k_max => 10.0::Float64,
 )

The default values to be stored in CosmoParams concerning the Input Power Spectrum Tools. In the Cosmology that will have such CosmoParams as input, they will be used in its IPSTools.

See also: CosmoParams, Cosmology, IPSTools

GaPSE.CosmoParams — Type

 CosmoParams(
      z_min::Float64
      z_max::Float64
      θ_max::Float64

      Ω_b::Float64
      Ω_cdm::Float64
      Ω_M0::Float64
      h_0::Float64

      s_lim::Float64

      IPS::Dict{Symbol,T1} where {T1}
      IPSTools::Dict{Symbol,T2} where {T2}
 )

Struct that contains all the parameters and options that are matter of concerns for the Cosmology we are interested in.

Arguments

z_min::Float64 and z_max::Float64 : the minimum and maximum redshifts of the survey we want to study.

θ_max::Float64 : Angular maximum value of the survey. It is implicitly assumed an azimutal simmetry of the survey.

Ω_b::Float64, Ω_cdm::Float64 and Ω_M0::Float64 : barionic, cold-dark-matter and total matter density parameters.

h_0::Float64 : today's Hubble adimensional parameter (H_0 = h_0 * 100 km/(s * Mpc)).

s_lim::Float64 : the lower-bound value for the function func_ℛ; it is necessary, because ℛ blows up for $s \rightarrow 0^{+}$. Consequently, if the func_ℛ input value is 0 ≤ s < s_lim, the returned value is always func_ℛ(s_lim).

IPS::Dict{Symbol,T1} where {T1} : dictionary concerning all the options that should be passed to InputPS in the contruction of a Cosmology. The allowed keys, with their default values, are stored in DEFAULT_IPS_OPTS, and are the following:

:fit_left_min => 1e-6 and :fit_left_max => 3e-6 : the limits (min and max) where the PS must be fitted with a (pure) power law, for small wavenumbers.

:fit_right_min => 1e1 and :fit_right_max => 2e1 : the limits (min and max) where the PS must be fitted with a (pure) power law, for high wavenumbers.

IPSTools::Dict{Symbol,T2} where {T2} : dictionary concerning all the options that should be passed to IPSTools in the contruction of a Cosmology. The allowed keys, with their default values, are stored in DEFAULT_IPSTOOLS_OPTS, and are the following:

:fit_min => 0.05 and :fit_max => 0.5 : the limits (min and max) where the integral $I_\ell^n$ in Cosmology must be fitted with a power law, for small distances. This operation is necessary, because xicalc, in this context, gives wrong results for too small input distance s; nevertheless, all these $I_\ell^n$ integrals have fixed power-law trends for $s \rightarrow 0$, so this approach gives good results.

:N => 1024 : number of points to be used in the Sperical Bessel Fourier Transform made by xicalc in IPSTools.

:k_min => 1e-6 and :k_max => 10.0 : extremes of integration for the σ_i integrals in IPSTools.

:con => true : do you want that the fit of all the $I_\ell^n$ in IPSTools for the LEFT edge is not a simple power-law $y = f(x) = b \, x^s$, but also consider a constant $a$, such that $y = f(x) = a + b \, x^s$?

Constructors

CosmoParams(z_min, z_max, θ_max; Ω_b = 0.0489, Ω_cdm = 0.251020, h_0 = 0.70, s_lim = 1e-2, IPS_opts::Dict = Dict{Symbol,Any}(), IPSTools_opts::Dict = Dict{Symbol,Any}() ) The associations are trivials, with Ω_M0 = Ω_cdm + Ω_b. For the two dictionary, you may pass only the key and the value you are interested in, and all the other default ones will be considered. For example, if you set:

IPSTools_opts = Dict(:N => 150, :con => false, :k_max => 30.0)

then the dictionary with all the options that will be passed to IPSTools will be:

IPSTools = merge(DEFAULT_IPSTOOLS_OPTS, IPSTools_opts) = :fit_min => 0.05, # default :fit_max => 0.5, # default :N => 150, # CHANGED VALUE :con => false, # CHANGED VALUE :k_min => 1e-6, # default :k_max => 30.0, # CHANGED VALUE )

and similar for IPS_opts.

See also: Cosmology, IPSTools, InputPS, func_ℛ, DEFAULT_IPSTOOLS_OPTS, DEFAULT_IPS_OPTS, check_compatible_dicts