The Cosmology Utilities

GaPSE.func_z_eff — Function

 func_z_eff(s_min, s_max, z_of_s) :: Float64

Return the effective redshift $z_\mathrm{eff}$, calcuated as follows:

\[\begin{split} z_\mathrm{eff} := \frac{ \int \mathrm{d}^3\mathbf{s} \, \phi^2(\mathbf{s}) \, z(s) }{ \int \mathrm{d}^3\mathbf{s}\, \phi^2(\mathbf{s}) } &= \frac{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s) \, z(s) \times \int_{4\pi}\mathrm{d}^2\hat{\mathbf{s}} \, W^2(\hat{\mathbf{s}}) }{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s)\times \int_{4\pi}\mathrm{d}^2\hat{\mathbf{s}} \, W^2(\hat{\mathbf{s}}) } \\[5pt] &= \frac{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s) \, z(s) }{ \int_0^\infty \mathrm{d}s \, s^2 \, \phi^2(s) } \\[4pt] &= \frac{3}{s_\mathrm{max}^3 - s_\mathrm{min}^3} \, \int_{s_\mathrm{min}}^{s_\mathrm{max}} \mathrm{d}s \, s^2 \, z(s) \end{split}\]

where we have used our assuption on separability of the window function

\[ \phi(\mathbf{s}) = \phi(s) \, W(\hat{s})\]

and their definitions.

See also: ϕ, W

GaPSE.s — Function

 s(s1, s2, y) :: Float64

Return the value $s = \sqrt{s_1^2 + s_2^2 - 2 \, s_1 \, s_2 \, y}$

See also: μ, s2, y

GaPSE.μ — Function

 μ(s1, s2, y) :: Float64

Return the value $\mu=\hat{\mathbf{s}}_1\cdot\hat{\mathbf{s}}$, defined as:

\[\mu = \mu(s_1, s_2, y) = \frac{y \, s_2 - s_1}{s(s_1, s_2, y)} \;, \quad s(s_1, s_2, y) = \sqrt{s_1^2 + s^2 - 2 \, s_1 \, s_2 \, y}\]

with $y=\cos\theta=\hat{\mathbf{s}}_1\cdot\hat{\mathbf{s}}$ and where $s$ is obtained from the function s

See also: s, s2, y

GaPSE.s2 — Function

 s2(s1, s, μ) :: Float64

Return the value $s_2 = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}$

See also: s, μ, y

GaPSE.y — Function

 y(s1, s, μ) :: Float64

Return the value $y=\cos\theta$, defined as:

\[y = y(s_1, s, \mu) = \frac{\mu \, s + s_1}{s2(s_1, s, \mu)} \;, \quad s_2 = \sqrt{s_1^2 + s^2 + 2 \, s_1 \, s \, \mu}\]

with $\mu=\hat{\mathbf{s}}_1\cdot\hat{\mathbf{s}}_2$ and where $s_2$ is btained from the function s2

See also: s, μ, s2

GaPSE.ϕ — Function

 ϕ(s, s_min, s_max) :: Float64

Radial part of the survey window function. Return 1.0 if is true that $s_\mathrm{min} < s < s_\mathrm{max}$ and 0.0 otherwise.

In this software we made the assuption that the survey window function can be separated into a radial and angular part, i.e.:

\[ \phi(\mathbf{s}) = \phi(s) \, W(\hat{s})\]

See also: W

GaPSE.W — Function

 W(θ, θ_max) :: Float64

Angular part of the survey window function. Return 1.0 if is true that $0.0 \leq \theta < \theta_\mathrm{max}$ and 0.0 otherwise. It is implicitly assumed an azimutal simmetry of the survey.

In this software we made the assuption that the survey window function can be separated into a radial and angular part, i.e.:

\[ \phi(\mathbf{s}) = \phi(s) \, W(\hat{s})\]

See also: ϕ

GaPSE.V_survey — Function

 V_survey(s_min, s_max, θ_max) :: Float64

Return the volume of a survey with azimutal simmetry, i.e.:

\[\begin{split} V(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max}) &= \; C_\mathrm{up} - C_\mathrm{down} + TC \\ &C_\mathrm{up} = \frac{\pi}{3} s_\mathrm{max}^3 \, (1 - \cos\theta_\mathrm{max})^2 \, (2 + \cos\theta_\mathrm{max}) \\ &C_\mathrm{down} = \frac{\pi}{3} s_\mathrm{min}^3 \, (1 - \cos\theta_\mathrm{max})^2 \, (2 + \cos\theta_\mathrm{max}) \\ &TC = \frac{\pi}{3} (s_\mathrm{max}^2 + s_\mathrm{min}^2 + s_\mathrm{max} \,s_\mathrm{min}) \, (s_\mathrm{max} - s_\mathrm{min})\, \cos\theta_\mathrm{max}\, \sin^2\theta_\mathrm{max} \end{split}\]

GaPSE.A — Function

 A(s_min, s_max, θ_max) :: Float64

Return the Power Spectrum multipole normalization coefficient A, i.e.:

\[ A(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max})= \frac{ V(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max}) }{4 \, \pi^2}\]

where $V(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max})$ is the survey volume.

Pay attention: this is NOT used for the normalization of PS, see instead A_prime

See also: V_survey

GaPSE.A_prime — Constant

 A_prime :: Float64

It's the Power Spectrum multipole normalization coefficient $A^{'}$, i.e.:

\[ A^{'} = \frac{3 \, A}{ (s_\mathrm{max}^3 - s_\mathrm{min}^3)} = \frac{1}{4\,\pi}\]

See also: A, V_survey