IMPLEMENTEDGREFFECTS and so on

GaPSE.IMPLEMENTED_GR_EFFECTS — Constant

 const IMPLEMENTED_GR_EFFECTS = [
      "auto_doppler", "auto_lensing",
      "auto_localgp", "auto_integratedgp", 
      
      "lensing_doppler", "doppler_lensing",
      "doppler_localgp", "localgp_doppler",
      "doppler_integratedgp", "integratedgp_doppler",
      "lensing_localgp", "localgp_lensing",
      "lensing_integratedgp", "integratedgp_lensing",
      "localgp_integratedgp", "integratedgp_localgp",
 ]

The names of the GR effects implemented. Their order is associated with the one in IMPLEMENTED_ξs, so be careful to change it.

See also: IMPLEMENTED_ξs

GaPSE.IMPLEMENTED_ξs — Constant

 const IMPLEMENTED_ξs = [
      ξ_Doppler, ξ_Lensing, ξ_LocalGP, ξ_IntegratedGP, 
      ξ_Lensing_Doppler, ξ_Doppler_Lensing,
      ξ_Doppler_LocalGP, ξ_LocalGP_Doppler,
      ξ_Doppler_IntegratedGP, ξ_IntegratedGP_Doppler,
      ξ_Lensing_LocalGP, ξ_LocalGP_Lensing,
      ξ_Lensing_IntegratedGP, ξ_IntegratedGP_Lensing,
      ξ_LocalGP_IntegratedGP, ξ_IntegratedGP_LocalGP
 ]

The names of the GR effect TPCFs implemented. Their order is associated with the one in IMPLEMENTED_GR_EFFECTS, so be careful to change it.

See also: IMPLEMENTED_GR_EFFECTS

GaPSE.DICT_GR_ξs — Constant

 const DICT_GR_ξs::Dict{String,Function}

For an input key string effect from IMPLEMENTED_GR_EFFECTS, return the associated TPCF DICT_GR_ξs[effect] from IMPLEMENTED_ξs.

Example

julia> GaPSE.DICT_GR_ξs["auto_doppler"]
ξ_Doppler

See also: IMPLEMENTED_GR_EFFECTS, IMPLEMENTED_ξs

GaPSE.INDEX_GR_EFFECT — Constant

 const INDEX_GR_EFFECT::Dict{String,Integer}

For an input key string effect from IMPLEMENTED_GR_EFFECTS, return the associated index position in that vector.

Example

julia> GaPSE.INDEX_GR_EFFECT["auto_doppler"]
1

See also: IMPLEMENTED_GR_EFFECTS

GaPSE.GR_EFFECT_INDEXED — Constant

 const GR_EFFECT_INDEXED::Dict{Integer,String}

For an input index position i of IMPLEMENTED_GR_EFFECTS, return the associated key string effect.

Example

julia> GaPSE.DICT_GR_ξs[1]
"auto_doppler"

See also: IMPLEMENTED_GR_EFFECTS