Mathematical Utilities functions

Missing docstring.

Missing docstring for GaPSE.warning. Check Documenter's build log for details.

GaPSE.derivate_point — Function

 derivate_point(xp, yp, x1, y1, x2, y2)

Return the derivative in (xp, yp), given the neighboor points (x1,y1) and (x2,y2), with x1 < xp < x2. It is not assumed that x2 - xp = xp - x1.

Missing docstring.

Missing docstring for GaPSE.derivate_vector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.spectral_index. Check Documenter's build log for details.

GaPSE.mean_spectral_index — Function

 mean_spectral_index(xs, ys; N::Integer = 1, con = false)

Assuming that the input ys follow a power law distribution, return the mean spectral index $\langle S \rangle$ of them.

The spectral index $S$ of a generic function $f = f(x)$ is defined as:

\[ S = \frac{\partial \log f(x)}{\partial \log x} = \frac{x}{f(x)} \frac{\partial f(x)}{\partial x} \]

GaPSE.power_law — Function

 power_law(x, si, b, a) ::Float64

Return the following $y = f(x)$ "spurious" power-law value:

\[y = f(x) = a + b \, x^s\]

where si is the exponent ($s$), b the coefficient ($b$) and a is the added constant ($a$).

Missing docstring.

Missing docstring for GaPSE.two_power_laws. Check Documenter's build log for details.

GaPSE.power_law_from_data — Function

 power_law_from_data(xs, ys, p0::Vector{Float64},
      fit_min::Number, fit_max::Number; con = false)

 power_law_from_data(xs, ys, p0::Vector{Float64}; con = false) = 
      power_law_from_data(xs, ys, p0, xs[begin], xs[end]; con = con)

Returns the "spurious" power-law coefficients $s$, $b$ and $a$ obtained from the fitting of the data vectors xs and ys inside the limits fit_min and fit_max.

If con == false, the returned a is always 0.0, because it is considered the "pure" power-law fitting function:

\[ y = f(x) = b \, x^s\]

while if con == false it is used the spurious one:

\[ y = f(x) = a + b \, x^s\]

(and consequently $a$ may be ≠0).

See also: power_law

Missing docstring.

Missing docstring for GaPSE.expanded_left_log. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.expanded_right_log. Check Documenter's build log for details.

GaPSE.expanded_IPS — Function

 expanded_IPS(ks, pks; k_in = 1e-8, k_end = 3e3, con = false)

Given the ks and pks of a chosen Power Spectrum, returns the same PS with "longer tails", i.e. it is prolonged for higher and lower ks than the input ones.

GaPSE.expanded_Iln — Function

 expanded_Iln(PK, l, n; N = 1024, kmin = 1e-4, kmax = 1e3, s0 = 1e-3,
      fit_left_min = 2.0, fit_left_max = 10.0, p0 = [-1.0, 1.0, 0.0], con = false)

GaPSE.func_I04_tilde — Function

 func_I04_tilde(PK, s, kmin, kmax; kwargs...)

Return the following integral:

\[\tilde{I}^4_0 (s) = \int_0^\infty \frac{\mathrm{d}q}{2\pi^2} q^2 \, P(q) \, \frac{j_0(q s) - 1}{(q s)^4}\]

It is brute-force calcuated with quadgk.

Arguments

PK : function that return the Input Power Spectrum

s : value of $s$ whre the integral must be evaluated

kmin, kmax : extremes (min and max) of integration

kwargs... : keyword argruments that must be passed to quadgk, such as rtol or atol

Missing docstring.

Missing docstring for GaPSE.expanded_I04_tilde. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.my_interpolation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GaPSE.my_println_vec. Check Documenter's build log for details.