Fast Fourier Transforms and Power Spectra

The FFTLog package

GaPSE.FFTLog.AbstractPlan — Type

 abstract type AbstractPlan end

The abstract type of all the Plan to be used in the code. At the moment, they are:

SingleBesselPlan

HankelPlan

GaPSE.FFTLog._c_window — Function

 _c_window(N::AbstractArray, NCut::Int)

Returns the smoothing window function $W(x, N_\mathrm{cut})$ as defined in Eq. (C.1) of McEwen et al. (2016):

\[W(x) = \begin{cases} \displaystyle \frac{x - x_{\mathrm{min}}}{x_{\mathrm{left}} - x_{\mathrm{min}}} - \frac{1}{2\pi}\sin\left( 2\pi \frac{x - x_{\mathrm{min}}}{x_{\mathrm{left}} - x_{\mathrm{min}}}\right) \; , \quad\quad x < x_{\mathrm{left}}\\[12pt] \displaystyle 1 \; , \quad\quad\quad\quad\quad\quad\quad\quad \quad\quad\quad\quad\quad\quad\quad\; \, x_{\mathrm{left}} \leq x \leq x_{\mathrm{right}}\\[8pt] \displaystyle \frac{x_{\mathrm{max}} - x}{x_{\mathrm{max}} - x_{\mathrm{right}}} - \frac{1}{2\pi}\sin\left( 2\pi \frac{x_{\mathrm{max}} - x}{x_{\mathrm{max}} - x_{\mathrm{right}}} \right) \; , \quad x > x_{\mathrm{right}} \end{cases}\]

GaPSE.FFTLog._logextrap — Function

 _logextrap(x::Vector, 
      n_extrap_low::Int, n_extrap_high::Int) ::Vector

Given an input LOGARITHMICALLY SPACED vector of values x, expands that vector adding n_extrap_low point on the left and n_extrap_high on the right. Consequently, for an input x of N values, it returns a vector X with length N + n_extrap_low + n_extrap_right.

It is not assumed that the spacing is the same in the two edges of the data.

GaPSE.FFTLog._zeropad — Function

 _zeropad(x::Vector, n_pad::Int)::Vector

Concatenates n_pad zeros both on the left and on the right of the input vector x. Consequently, for an input x of N values, it returns a vector X with length N + 2 * n_pad.

GaPSE.FFTLog._eval_cm! — Function

 _eval_cm!(plan::AbstractPlan, fx)

Given a plan::AbstractPlan, compute the power-law expansion coefficients $c_m$ of the input data vector fx. It is assumed that fx contains the y-axis values corresponding to the x-axis ones plan.x, and consequently their length must be the same. The computed cm vector is stored in plan.cm, and nothing is returned.

For a function f evaluated the N x-axis values x, the c_m coefficients are

\[c_m = W_m \sum_{q=0}^{N-1} \frac{f(x_q)}{x_q^\nu} e^{-\frac{2\pi}{N}i m q}\]

where $W_m$ is the smoothing window function computed via _c_window and $\nu$ is the bias parameter stored in plan.ν

See also: _c_window, AbstractPlan

GaPSE.FFTLog._eval_ηm! — Function

 _eval_ηm!(plan::AbstractPlan)

Given an input plan::AbstractPlan, compute all the $\eta_m$ coefficients, defined as follows:

\[\eta_m = \frac{2 \pi m}{N \, \Delta_{\ln x}} \, \]

where $N$, $\Delta_{\ln x}$ and the $m$ vector are respectively plan.N, plan.d_ln_x and plan.m.

The computed ηm vector is stored in plan.cm, and nothing is returned.

See also: AbstractPlan

GaPSE.FFTLog.SingleBesselPlan — Type

 mutable struct SingleBesselPlan{T,C} <: AbstractPlan

This struct contains all the elements necessary to evaluate the integral with one Bessel function. All the arguments of this struct are keyword arguments. Here we show the compelte list and their default values:

x::Vector{T} : the LOGARITHMICALLY SPACED vector of x-axis values. You need always to provide this vector.

y::Matrix{T} = zeros(10, 10) : the logarithmically spaced vector of the values where the transformed function will be evaluated. It has the same length of x

fy::Matrix{T} = zeros(10, 10) : the y-axis of the transformed function; it is a vector if only one Bessel function order is provided in the functions

hm::Matrix{C} = zeros(ComplexF64, 10, 10) : matrix of the coefficients $h_m = c_m \, h_{m, \mathrm{corr}} \, g_\ell$, where $c_m$s, $h_{m, \mathrm{corr}}$s and $g_\ell$ are respectively stored in plan.cm, plan.hm_corr and plan.gl. Each column contains all the $h_m$s for a given spherical Bessel order $\ell$.

hm_corr::Matrix{C} = zeros(ComplexF64, 10, 10) : matrix of the coefficients $h_{m, \mathrm{corr}} = (x_0 y_0)^{- i \eta_m}$, where $\eta_m = \frac{2 \pi m}{N \, \Delta_{\ln x}}$ and $x_0$ and $y_0$ are the smallest values of plan.x and plan.y, respectively. Each column contains all the $h_{m, \mathrm{corr}}$s for a given spherical Bessel order $\ell$.

d_ln_x::T = log(x[2] / x[1]) : the spacing between the x elements.

fy_corr::Matrix{T} = zeros(10, 10) : matrix of the coefficients $K(y) = \frac{\sqrt{\pi}}{4 y^{\nu}}$, where $\nu$ is the bias paremeter stored in plan.ν. Each column contains all the $h_{m, \mathrm{corr}}$s for a given spherical Bessel order $\ell$.

original_length::Int = length(x) : the original inpout length of the x vector; it is stored because, for numerical stability purposes, during the computation this vector is expanded at the edged, and so the input function ones.

gl::Matrix{C} = zeros(ComplexF64, 100, 100) : vector with the $g_\ell$ values for all the input spherical Bessel order.

ν::T = 1.01 : bias parameter.

n_extrap_low::Int = 0 : number of points to concatenate on the left of x, logarithmically distributed with the same ratio of the left-edge elements of x

n_extrap_high::Int = 0 : number of points to concatenate on the right of x, logarithmically distributed with the same ratio of the right-edge elements of x

c_window_width::T = 0.25 : position where the tapering by the window function begins; by default c_window_width= 0.25, so is begins when $m = \pm 0.75 \times N/2$, where $N$ is the size of the input array.

n_pad::Int = 0 : number of zeros to be concatenated both on the left and on the right of the input function.

n::Int = 0 : the derivative order for the spherical Bessel function.

N::Int = original_length + n_extrap_low + n_extrap_high + 2 * n_pad : number of points where the input function is known; are considered both the "true values" and the fake ones, added for a more numerically stable fft.

m::Vector{T} = zeros(N) : vector with all the indexes that will be used for the power-law expansion of the input function

cm::Vector{C} = zeros(ComplexF64, N) : vector containing all the input function power-law exapnsion $c_m$ coefficients.

ηm::Vector{T} = zeros(N) : vector of all the $\eta_m = \frac{2 \pi m}{N \, \Delta_{\ln x}}$ coefficients.

plan_rfft::FFTW.rFFTWPlan = plan_rfft(randn(1024)) : a random initialized fft plan of FFTW

plan_irfft = plan_irfft( randn(Complex{Float64}, 2, Int((original_length + n_extrap_low + n_extrap_high + 2 * n_pad) / 2) + 1), original_length + n_extrap_low + n_extrap_high + 2 * n_pad, 2 ) :

See also: AbstractPlan

GaPSE.FFTLog.HankelPlan — Type

 mutable struct HankelPlan{T,C} <: AbstractPlan

A specific type of FFTLogPlan designed for the Hankel transform. Its arguments are the same of SingleBesselPlan, checks its documentation for more information.

See also: SingleBesselPlan, AbstractPlan

GaPSE.FFTLog._eval_gl — Function

 _eval_gl(ell, z::Vector, n::Int )::Vector

Evaluate the $g_{\ell}$ coefficients, defined as

\[g_{\ell}^{(n)}(z) = (-1)^n \, 2^{z-n} \, \frac{ \Gamma\left(\frac{\ell + z - n}{2}\right) }{ \Gamma\left(\frac{3 + \ell + n - z}{2}\right) }\]

GaPSE.FFTLog._eval_y! — Function

 _eval_y!(plan::Union{SingleBesselPlan, HankelPlan}, ell::Vector)

Given an input plan::Union{SingleBesselPlan, HankelPlan}, compute the y values where the output function will be evaluated and the coefficient $K(y)$ outside the IFFT. They are, respectively:

\[y = \frac{\ell + 1}{x} \; , \quad\quad K(y) = \frac{\sqrt{\pi}}{4 y^{\nu}}\]

The vector of their values are stored respectively in plan.y and plan.fy_corr, and nothing is returned.

See also: SingleBesselPlan, HankelPlan

GaPSE.FFTLog._eval_gl_hm! — Function

 _eval_gl_hm!(plan::Union{SingleBesselPlan, HankelPlan}, ell::Vector)

Given an input plan::Union{SingleBesselPlan, HankelPlan}, compute the $g_{\ell}$ values and the $h_{m, \mathrm{corr}}$ coefficents inside the IFFT. They are, respectively:

\[g_{\ell}^{(n)}(z) = (-1)^n \, 2^{z-n} \, \frac{ \Gamma\left(\frac{\ell + z - n}{2}\right) }{ \Gamma\left(\frac{3 + \ell + n - z}{2}\right) } \; , \quad\quad h_{m, \mathrm{corr}} = (x_0 y_0)^{- i \eta_m}\]

where $\eta_m = \frac{2 \pi m}{N \, \Delta_{\ln x}}$, and $x_0$ $y_0$ are the smallest values of plan.x and plan.y, respectively.

The vector of their values are stored in plan.gl and plan.hy_corr, and nothing is returned.

See also: SingleBesselPlan, HankelPlan

GaPSE.FFTLog.prepare_FFTLog! — Function

 prepare_FFTLog!(plan::Union{SingleBesselPlan, HankelPlan}, ell::Vector)

Given an input plan::Union{SingleBesselPlan, HankelPlan}, pre-plan an optimized real-input FFT for all the Bessel function orders stored in the vector ell. In other words, it computes:

the y vector of values where the transformed will be evaluated (stored in plan.y).

the corresponding gl vector of $g_{\ell}$ values (stored in plan.gl).

the m vector of indexes for the $c_m$ coefficents (stored in plan.m).

the corresponding ηm and hm_corr vector of $\eta_m$ and $h_{m, \mathrm{corr}}$ values (stored in plan.ηm and plan.hm_corr).

See also: SingleBesselPlan, HankelPlan

GaPSE.FFTLog.prepare_Hankel! — Function

 prepare_Hankel!(plan::HankelPlan, ell::Vector)

Given an input plan::HankelPlan, pre-plan an optimized real-input FFT for all the Bessel function orders stored in the vector ell concerning an Hankel transform. Same as prepare_FFTLog, checks its documentation for more information.

See also: HankelPlan, prepare_FFTLog!

GaPSE.FFTLog.get_y — Function

 get_y(plan::Union{SingleBesselPlan, HankelPlan})::Vector

Return the computed y vector, containing the values where the transformed function will be evaluated.

See also: SingleBesselPlan, HankelPlan

GaPSE.FFTLog.evaluate_FFTLog — Function

evaluate_FFTLog(plan::AbstractPlan, fx)::Union{Vector, Matrix}

Given an input plan::AbstractPlan, evaluate the FFT fy of the fx y-axis data on the basis of the parameters stored in plan. The result is both stored in plan.fy and retuned as output.

See also: AbstractPlan

evaluate_FFTLog(plan::Union{SingleBesselPlan, HankelPlan}, fx)::Union{Vector, Matrix}

Given an input plan::Union{SingleBesselPlan, HankelPlan}, evaluate the FFT fy of the fx y-axis data on the basis of the parameters stored in plan. The result is both stored in plan.fy and retuned as output.

See also: SingleBesselPlan,HankelPlan

GaPSE.FFTLog.evaluate_Hankel — Function

evaluate_Hankel(plan::HankelPlan, fx)::Union{Vector, Matrix}

Given an input plan::HankelPlan, evaluate the FFT fy of the fx y-axis data on the basis of the parameters stored in plan for an Hankel transform. The result is both stored in plan.fy and retuned as output.

See also: HankelPlan

GaPSE.FFTLog.evaluate_Hankel! — Function

evaluate_Hankel!(fy, plan::HankelPlan, fx)

Given an input plan::HankelPlan, evaluate the FFT fy of the fx y-axis data on the basis of the parameters stored in plan for an Hankel transform. The result is stored both in plan.fy and in the input fy.

See also: HankelPlan

The Power Spectrum with FFTLog

GaPSE.FFTLog_PS_multipole — Function

 FFTLog_PS_multipole(ss, xis;
      pr::Bool=true,
      L::Int=0, ν::Union{Float64,Nothing}=nothing,
      n_extrap_low::Int=500,
      n_extrap_high::Int=500, n_pad::Int=500,
      )

Computes the Power Spectrum through the FFTLog algorithm. More precisely, it computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and xis.

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

See also: PS_multipole

GaPSE.FFTLog_all_PS_multipole — Function

 FFTLog_all_PS_multipole(input::String,
      group::String=VALID_GROUPS[end];
      L::Int=0, pr::Bool=true,
      ν::Union{Float64,Nothing,Vector{Float64}}=nothing,
      n_extrap_low::Int=500,
      n_extrap_high::Int=500, n_pad::Int=500
 )

Computes the Power Spectrum through the FFTLog algorithm for a set of TPCFs. More precisely, it read the input file input, taking the first column as the x-axis ss vector and the following columns as the y-axis ones, and computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and each y-axis xis.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

pr::Bool=true : want to print the automatic messages to the screen?

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

See also: FFTLog_PS_multipole, PS_multipole

The Power Spectrum with TwoFAST

GaPSE.TwoFAST_PS_multipole — Function

 TwoFAST_PS_multipole(f_in;
      int_s_min::Float64 = 1e-1, int_s_max::Float64 = 1e3,
      L::Int = 0, N::Int = 1024, pr::Bool = true,
      k0::Union{Nothing,Float64} = nothing,
      right::Union{Float64,Nothing} = nothing
 ) ::Tuple{Vector{Float64}, Vector{Float64}}

Computes the Power Spectrum from the input spline f_in through the TwoFAST xicalc function of the TwoFAST Julia package. More precisely, it computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the inpunt spline.

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

N::Int = 1024 : number of points to be used in Fourier transform

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / int_s_max

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

See also: PS_multipole

 TwoFAST_PS_multipole(ss, fs;
      int_s_min::Float64 = 1e-1, int_s_max::Float64 = 1e3,
      epl::Bool = true, pr::Bool = true, L::Int = 0,
      N_left::Int = 12, N_right::Int = 12,
      p0_left = [-2.0, 1.0], p0_right = [-2.0, 1.0],
      k0::Union{Nothing,Float64} = nothing
 ) ::Tuple{Vector{Float64}, Vector{Float64}}

Takes the input data vector ss and fs and creates a spline from them, passing it as input tho the other TwoFAST_PS_multipole method. Depending on the options, it may create also a power law epansions on the edges.

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

See also: PS_multipole

GaPSE.TwoFAST_all_PS_multipole — Function

 TwoFAST_all_PS_multipole(input::String,
      group::String=VALID_GROUPS[end];
      L::Int = 0, pr::Bool = true, 
      kwargs...)

Computes the Power Spectrum through the TwoFAST xicalc function of the TwoFAST Julia package for a set of TPCFs. More precisely, it read the input file input, taking the first column as the x-axis ss vector and the following columns as the y-axis ones, and computes the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and each y-axis xis.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

See also: TwoFAST_PS_multipole, PS_multipole

The Power Spectrum multipole computation

GaPSE.PS_multipole — Function

 PS_multipole(ss, fs; 
      pr::Bool = true, L::Int = 0, 
      alg::Symbol = :fftlog, 
      cut_first_n::Int = 0, cut_last_n::Int = 0, 
      kwargs...
      ) ::Tuple{Vector{Float64}, Vector{Float64}}

 PS_multipole(input::String; 
      kwargs...)

Return the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation:

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

The second method reads the input file, takes the first column as ss and the second as fs and recalls the first method.

Currenlty, there are two algorithms you can choose in order to perform the computation; you can choose which one to use through the keyword value alg:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

IMPORTANT: no matter which algorithm you choose, you will need to give the input data in a LOGARITHMICALLY DISTRIBUTED scale. A linear distribution does not fit for the algorithms to apply.

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

Analytical derivation

The analytical expression previously showed can be easily obtained from the standard one:

\[\begin{split} P_L(k) = &\frac{2 L + 1}{A} (-i)^L \, \int_0^\infty \mathrm{d} s_1 \; s_1^2 \int_0^\infty \mathrm{d} s \; s^2 \int_{-1}^{+1} \mathrm{d} \mu \; j_L(ks) \, \xi(s_1, s, \mu) \, \phi(s_1) \, \phi(s_2) \, \mathcal{L}_L(\mu) F\left(\frac{s}{s_1}, \mu \right) \\ &\mathrm{with} \; \;s_2 = s_2(s_1, s, μ) = \sqrt{s_1^2 + s^2 + 2s_1s\mu} \; , \quad A(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max}) \frac{ V(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max}) }{4 \, \pi^2} \end{split}\]

with the definition

\[f_\mathrm{in}(s_1, s) = \int_{-1}^{+1} \mathrm{d} \mu \; \xi(s_1, s, \mu) \, \phi(s_2) \, \mathcal{L}_L(\mu) \, F\left(\frac{s}{s_1}, \mu \right)\]

and the application of the effective redshift approximation.

Returns

A Tuple{Vector{Float64}, Vector{Float64}} with:

the k values vector as first element;

the correspoding PS pk values vector as second one.

See also: V_survey, A, A_prime, EPLs, print_PS_multipole

GaPSE.print_PS_multipole — Function

 print_PS_multipole(ss, fs, out::String;
      L::Int=0, pr::Bool=true, alg::Symbol=:fftlog, kwargs...)
 print_PS_multipole(input::String, out::String;
      kwargs...)

Takes in input a filename input where is stored a TPCF multipole, calculate the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

where $f_\mathrm{in}$ is the function samples by ss and xis, and save it in the file out, together with the options used for the computation.

The second method reads the input file, takes the first column as ss and the second as fs and recalls the first method.

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

alg::Symbol = :fftlog : algorithm to be used for the computation. Currenlty, there are two algorithms you can coose in order to perform the computation:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

See also: V_survey, A, A_prime, EPLs, PS_multipole

GaPSE.all_PS_multipole — Function

 function all_PS_multipole(input::String,
      group::String=VALID_GROUPS[end];
      L::Int = 0, pr::Bool = true, 
      alg::Symbol=:fftlog, kwargs...
      ) ::Tuple{Vector{Float64}, Vector{Vector{Float64}}}

Given an input file where the first column is the x-axis data one and all the following columns are the corresponding y-data ones, this function computes all the Power Spectra of each y-data column and return a Tuple containing

as first element, the ks values, common to all the PS

as second element, a vector where in each position there is the Power Spectra corresponding to the associated inputy y-data.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

alg::Symbol = :fftlog : algorithm to be used for the computation. Currenlty, there are two algorithms you can coose in order to perform the computation:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

See also: EPLs, PS_multipole

GaPSE.print_all_PS_multipole — Function

 print_all_PS_multipole(input::String, out::String,
      group::String = VALID_GROUPS[end]; 
      L::Int = 0, pr::Bool = true, 
      alg::Symbol = :fftlog,
      kwargs...)

Given an inputfile where the first column is the x-axis data one and all the following columns are the corresponding y-data ones, this function computes all the Power Spectra of each y-data column and print in a file namedout`

as first column, the ks values, common to all the PS

as folowing columns, a vector where in each position there is the Power Spectra corresponding to the associated inputy y-data.

The group::String=VALID_GROUPS[end] argument allow you to specify the group of the input TPCF, if they were computed through GAPSE. The allowed values for this argument are: LD , GNC , GNCxLD , LDxGNC , generic ,

If you choose a group pay attention that the number of input TPCF must match the group number (16, 25, 20 and 20 respectively). The last group name (which is also the default value) is used in case the input xis do not belog to a specific group (and so no predefined number is expected).

Optional arguments

Depending on the algorithm you choose, the options would change. The options in common are:

pr::Bool=true : want to print the automatic messages to the screen?

L::Int=0 : which multipole order should I use for this computation? IT MUST MATCH THE MULTIPOLE ORDER OF THE INPUT TPCF!

cut_first_n::Int=0 and cut_last_n::Int=0 : you can cout the first and/or last n elements of the input data, if they are highly irregular.

alg::Symbol = :fftlog : algorithm to be used for the computation. Currenlty, there are two algorithms you can coose in order to perform the computation:

alg = :fftlog (default and recommended option) will employ the FFTLog algorithm.

alg = :twofast will employ the TwoFAST xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max. This alogrithm is not the ideal choise, because TwoFAST is conceived for the direction PS -> TPCF, while is not 100% trustworthy for the other way round.

The specific ones for alg = :fftlog are:

ν::Union{Float64,Nothing} = nothing : bias parameter, i.e. exponent used to "balance" the curve; if nothing, will be set automatically to 1.5

n_extrap_low::Int = 500 and n_extrap_high::Int = 500 : number of points to concatenate on the left/right of the input x-axis ss vector, logarithmically distributed with the same ratio of the left/right-edge elements of ss.

n_pad::Int = 500 : number of zeros to be concatenated both on the left and on the right of the input function. They stabilize a lot the algorithm.

The specific ones for alg = :twofast are:

epl::Bool=true : do you want to extend the edges of the input vectors using two fitted power-laws (obtained from EPLs)

N_left::Int = 12 and N_right::Int = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max.

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3: extremes of integration; if epl is set to false, their values will be automatically set to min(ss...) and max(ss...) respectively. Their values do matter only if epl=true.

p0_left=[-2.0, 1.0] and p0_right=[-2.0, 1.0]: vectors with the initial values for the left/right power-law fitting of EPLs; the power-law is in the form $y = f(x) = b * x^s$, so the first vector value is the initial value of $s$ (and of course the second is the one of $b$).

k0::Union{Nothing,Float64} = nothing : starting point for the xicalc function; if nothing, it will be set k0 = 1.0 / max(ss...)

right::Union{Float64,Nothing} = nothing : do you want to cut the output elements with ks .> right? if set to nothing, no cut will be done.

N::Int = 1024 : number of points to be used in Fourier transform

See also: EPLs, PS_multipole