The Power Spectrum multipole computation

GaPSE.PS_multipole — Function

 PS_multipole(input::String; N_left::Integer = 12, 
      N_right::Integer = 12, kwargs...
      ) ::Tuple{Vector{Float64}, Vector{Float64}}

 PS_multipole(f_in; int_s_min::Float64 = 1e-1, 
      int_s_max::Float64 = 1e3, L::Integer = 0, 
      N::Integer = 1024, pr::Bool = true
      ) ::Tuple{Vector{Float64}, Vector{Float64}}

Return the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation:

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

In the former method (recommended), you should pass the name of the file where the TPCF multipole in exam is saved in, while in the latter you have to give its function/spline f_in. Internally, the first method creates a spline of the considered ξ and call the second with that spline as input; the kwargs... refers to the latter method ones infact.

The analytical expression previously showed can be easily obtained from the standard one:

\[\begin{split} P_L(k) = &\frac{2 L + 1}{A} (-i)^L \, \int_0^\infty \mathrm{d} s_1 \; s_1^2 \int_0^\infty \mathrm{d} s \; s^2 \int_{-1}^{+1} \mathrm{d} \mu \; j_L(ks) \, \xi(s_1, s, \mu) \, \phi(s_1) \, \phi(s_2) \, \mathcal{L}_L(\mu) F\left(\frac{s}{s_1}, \mu \right) \\ &\mathrm{with} \; \;s_2 = s_2(s_1, s, μ) = \sqrt{s_1^2 + s^2 + 2s_1s\mu} \; , \quad A(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max}) \frac{ V(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max}) }{4 \, \pi^2} \end{split}\]

with the definition

\[f_\mathrm{in}(s_1, s) = \int_{-1}^{+1} \mathrm{d} \mu \; \xi(s_1, s, \mu) \, \phi(s_2) \, \mathcal{L}_L(\mu) \, F\left(\frac{s}{s_1}, \mu \right)\]

and the application of the effective redshift approximation.

The computation is made through the xicalc function of the TwoFAST Julia package. Note that in the computation the integration range $0\leq s \leq \infty$ is reduced to int_s_min ≤ s ≤ int_s_max`

Optional arguments

int_s_min::Float64 = 1e-1 and int_s_max::Float64 = 1e3 : extremes of integration. Do not worry if the input TPCF is not defined in all the integration range: it will be fitted with two power laws at its extremes, thanks to the function EPLs

N_left::Integer = 12 and N_right::Integer = 12 : number of points from left right edges to be used for the power law fitting in EPLs. They matters only if in the given input file ξ is not defined until the extremes of integration int_s_min and int_s_max

L::Integer = 0: order of the Legendre polynomial to be used; note that the multipole order L must be the same of the input TPCF in exam! Otherwise, the results would have no sense at all!

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

N::Integer = 1024 : number of points to be returned by xicalc

Returns

A Tuple{Vector{Float64}, Vector{Float64}} with:

the k values vector as first element;

the correspoding PS pk values vector as second one.

See also: V_survey, A, A_prime, EPLs, print_PS_multipole

GaPSE.print_PS_multipole — Function

 print_PS_multipole(input::String, out::String;
      L::Integer = 0, N::Integer = 1024,
      pr::Bool = true, kwargs...)

Takes in input a filename input where is stored a TPCF multipole, calculate the L-order PS multipole through the following Fast Fourier Transform and the effective redshift approximation

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = \frac{1}{4\,\pi}\]

and save it in the file out, together with the options used for the computation.

Optional arguments

L::Integer = 0: order of the Legendre polynomial to be used; note that the multipole order L must be the same of the input TPCF in exam! Otherwise, the results would have no sense at all!

pr::Bool = true : do you want the progress bar showed on screen, in order to check the time needed for the computation? (true recommended)

N::Integer = 1024 : number of points to be returned by xicalc

kwargs... : other keyword arguments that will be passed to PS_multipole

See also: V_survey, A, A_prime, EPLs, PS_multipole