The Power Spectrum multipole computation

GaPSE.PS_multipole — Function

 PS_multipole(in::String, int_s_min = 0.0, int_s_max = 1000.0; 
      L::Integer = 0, N::Integer = 128, pr::Bool = true)
 PS_multipole(
      f_in::Union{Function,Dierckx.Spline1D}, int_s_min = 0.0, int_s_max = 1000.0; 
      L::Integer = 0, N::Integer = 128, pr::Bool = true
      ) :: Tuple{Vector{Float64}, Vector{Float64}}

Return the L-order multipole from the input function f_in, through the following Fast Fourier Transform and the effective redshift approximation:

\[P_L(k) = \frac{2 L + 1}{A^{'}} (-i)^L \, \phi(s_\mathrm{eff}) \int_0^\infty \mathrm{d} s \; s^2 \, j_L(ks) \, f_\mathrm{in}(s) \; , \quad \; A^{'} = 4 \, \pi^2\]

This expression can be easily obtained from the standard one:

\[\begin{split} P_L(k) = &\frac{2 L + 1}{A} (-i)^L \, \int_0^\infty \mathrm{d} s_1 \; s_1^2 \int_0^\infty \mathrm{d} s \; s^2 \int_{-1}^{+1} \mathrm{d} \mu \; j_L(ks) \, \xi(s_1, s, \mu) \, \phi(s_1) \, \phi(s_2) \, \mathcal{L}_L(\mu) F\left(\frac{s}{s_1}, \mu \right) \\ &\mathrm{with} \; \;s_2 = s_2(s_1, s, μ) = \sqrt{s_1^2 + s^2 + 2s_1s\mu} \; , \quad A(s_\mathrm{max}, s_\mathrm{min}, \theta_\mathrm{max}) = 2 \, \pi \, V \end{split}\]

with the definition

\[f_\mathrm{in}(s_1, s) = \int_{-1}^{+1} \mathrm{d} \mu \; \xi(s_1, s, \mu) \, \phi(s_2) \, \mathcal{L}_L(\mu) \, F\left(\frac{s}{s_1}, \mu \right)\]

and the application of the effective redshift approximation.

See also: V_survey, A, A_prime